noirc_evaluator::acir
Module acir_context

The acir_context module acts as a high-level wrapper around generated_acir, maintaining a small amount of type information from SSA in order to allow efficient ACIR-gen.

The AcirContext struct defines how to translate SSA instructions into equivalent ACIR constructs. This layer of ACIR-gen is aware of a small amount of SSA type information in order to ensure that the generated ACIR is correct and for optimizations.

AcirContext also tracks Expressions which have been simplified into a Witness, or constant witnesses allowing these to be reused later in the program.

Re-exports

pub use generated_acir::GeneratedAcir;

Modules

black_box 🔒

brillig_call 🔒

generated_acir 🔒

The generated_acir module is responsible for the lowest level tracking of ACIR-gen.

Structs

AcirContext 🔒

Context object which holds the relationship between Variables(AcirVar) and types such as Expression and Witness which are placed into ACIR.

Enums

AcirVarData 🔒

Enum representing the possible values that a Variable can be given.

Functions

power_of_two 🔒

Returns an F representing the value 2**power