noirc_evaluator::acir::acir_context
Module generated_acir

The generated_acir module is responsible for the lowest level tracking of ACIR-gen.

This module defines the GeneratedAcir which is constructed as part of the ACIR-gen pass to accumulate all of the ACIR. This tracks properties such as which witnesses form part of the circuit interface and defines common snippets of ACIR to perform simple operations. Notably it has no concept of SSA or any types for the variables it works with, it instead only works on the [Expression]s and [Witness]es which will be reflected in the final circuit.

Modules

brillig_directive 🔒

Structs

GeneratedAcir

The output of the Acir-gen pass, which should only be produced for entry point Acir functions

Functions

black_box_expected_output_size 🔒

This function will return the number of outputs that a blackbox function expects. Returning None if there is no expectation.

black_box_func_expected_input_size 🔒

This function will return the number of inputs that a blackbox function expects. Returning None if there is no expectation.

intrinsics_check_inputs 🔒

Checks that the number of inputs being used to call the blackbox function is correct according to the function definition.

intrinsics_check_outputs 🔒

Checks that the number of outputs being used to call the blackbox function is correct according to the function definition.

Type Aliases

BrilligOpcodeToLocationsMap 🔒

BrilligProcedureRangeMap 🔒

OpcodeToLocationsMap 🔒

Correspondence between an opcode index (in opcodes) and the source code call stack which generated it