noirc_evaluator::acir::acir_context::generated_acir
Struct GeneratedAcir

pub struct GeneratedAcir<F: AcirField> {

    current_witness_index: Option<u32>,
    pub opcodes: Vec<Opcode<F>>,
    pub call_stacks: CallStackHelper,
    pub location_map: BTreeMap<OpcodeLocation, CallStackId>,
    pub return_witnesses: Vec<Witness>,
    pub input_witnesses: Vec<Witness>,
    pub brillig_locations: BTreeMap<BrilligFunctionId, BTreeMap<BrilligOpcodeLocation, CallStackId>>,
    pub(crate) call_stack_id: CallStackId,
    pub assertion_payloads: BTreeMap<OpcodeLocation, AssertionPayload<F>>,
    pub error_types: BTreeMap<ErrorSelector, ErrorType>,
    pub warnings: Vec<SsaReport>,
    pub name: String,
    pub brillig_stdlib_func_locations: BTreeMap<OpcodeLocation, BrilligStdlibFunc>,
    pub brillig_procedure_locs: BTreeMap<BrilligFunctionId, BTreeMap<ProcedureDebugId, (usize, usize)>>,

}

The output of the Acir-gen pass, which should only be produced for entry point Acir functions

Fields
current_witness_index: Option<u32>

The next witness index that may be declared. If witness index is None then we have not yet created a witness and thus next witness index that be declared is zero. This field is private should only ever be accessed through its getter and setter.

Equivalent to acvm::acir::circuit::Circuit’s field of the same name.

opcodes: Vec<Opcode<F>>

The opcodes of which the compiled ACIR will comprise.

call_stacks: CallStackHelperlocation_map: BTreeMap<OpcodeLocation, CallStackId>return_witnesses: Vec<Witness>

All witness indices that comprise the final return value of the program

input_witnesses: Vec<Witness>

All witness indices which are inputs to the main function

brillig_locations: BTreeMap<BrilligFunctionId, BTreeMap<BrilligOpcodeLocation, CallStackId>>

Brillig function id -> Opcodes locations map This map is used to prevent redundant locations being stored for the same Brillig entry point.

call_stack_id: CallStackId

Source code location of the current instruction being processed

assertion_payloads: BTreeMap<OpcodeLocation, AssertionPayload<F>>

Correspondence between an opcode index and the error message associated with it.

error_types: BTreeMap<ErrorSelector, ErrorType>

Correspondence between error selectors and types associated with them.

warnings: Vec<SsaReport>name: String

Name for the corresponding entry point represented by this Acir-gen output. Only used for debugging and benchmarking purposes

brillig_stdlib_func_locations: BTreeMap<OpcodeLocation, BrilligStdlibFunc>

Maps the opcode index to a Brillig std library function call. As to avoid passing the ACIR gen shared context into each individual ACIR we can instead keep this map and resolve the Brillig calls at the end of code generation.

brillig_procedure_locs: BTreeMap<BrilligFunctionId, BTreeMap<ProcedureDebugId, (usize, usize)>>

Brillig function id -> Brillig procedure locations map This maps allows a profiler to determine which Brillig opcodes originated from a reusable procedure.

Implementations

impl<F: AcirField> GeneratedAcir<F>

pub fn current_witness_index(&self) -> Witness

Returns the current witness index.

pub(crate) fn push_opcode(&mut self, opcode: AcirOpcode<F>)

Adds a new opcode into ACIR.

pub(crate) fn initialize_memory( &mut self, block_id: BlockId, init: Vec<Witness>, block_type: BlockType, )

Initializes memory block with given values.

pub(crate) fn get_call_stack(&self) -> CallStack

pub(crate) fn opcodes(&self) -> &[AcirOpcode<F>]

pub fn take_opcodes(&mut self) -> Vec<AcirOpcode<F>>

pub(crate) fn next_witness_index(&mut self) -> Witness

Updates the witness index counter and returns the next witness index.

pub(crate) fn get_or_create_witness(&mut self, expr: &Expression<F>) -> Witness

Converts Expression expr into a Witness.

If expr can be represented as a Witness then this function will return it, else a new opcode will be added to create a Witness that is equal to expr.

fn create_witness_for_expression( &mut self, expression: &Expression<F>, ) -> Witness

Creates a new Witness which is constrained to be equal to the passed Expression.

The reason we do this is because constraints in ACIR have a degree limit This means you cannot multiply an infinite amount of Expressions together. Once the Expression goes over degree-2, then it needs to be reduced to a Witness which has degree-1 in order to be able to continue the multiplication chain.

impl<F: AcirField> GeneratedAcir<F>

pub(crate) fn call_black_box( &mut self, func_name: BlackBoxFunc, function_inputs: Vec<Vec<FunctionInput<F>>>, constant_inputs: Vec<F>, num_bits: Option<u32>, outputs: Vec<Witness>, ) -> Result<(), InternalError>

Calls a black box function and returns the output of said blackbox function.

pub(crate) fn radix_le_decompose( &mut self, input_expr: &Expression<F>, radix: u128, limb_count: u32, bit_size: u32, ) -> Result<Vec<Witness>, RuntimeError>

Takes an input expression and returns witnesses that are constrained to be limbs decomposed from the input for the given radix and limb count.

Only radix that are a power of two are supported

pub(crate) fn brillig_to_radix( &mut self, expr: &Expression<F>, radix: u128, limb_count: u32, ) -> Vec<Witness>

Adds brillig opcode for to_radix

This code will decompose expr in a radix-base and return Witnesses which may (or not, because it does not apply constraints) be limbs resulting from the decomposition.

Safety: It is the callers responsibility to ensure that the resulting Witnesses are properly constrained.

pub(crate) fn brillig_inverse(&mut self, expr: Expression<F>) -> Witness

Adds an inversion brillig opcode.

This code will invert expr without applying constraints and return a Witness which may or may not be the result of inverting expr.

Safety: It is the callers responsibility to ensure that the resulting Witness is constrained to be the inverse.

pub(crate) fn assert_is_zero(&mut self, expr: Expression<F>)

Asserts expr to be zero.

If expr is not zero, then the constraint system will fail upon verification.

pub(crate) fn is_equal( &mut self, lhs: &Expression<F>, rhs: &Expression<F>, ) -> Witness

Returns a Witness that is constrained to be:

1 if lhs == rhs

0 otherwise

fn is_zero(&mut self, t_expr: Expression<F>) -> Witness

Returns a Witness that is constrained to be:

1 if t == 0

0 otherwise

Proof

First, let’s create a new variable y which will be the Witness that we will ultimately return indicating whether t == 0. Note: During this process we need to apply constraints that ensure that it is a boolean. But right now with no constraints applied to it, it is essentially a free variable.

Next we apply the following constraint y * t == 0. This implies that either y or t or both is 0.

If t == 0, then by definition t == 0.

If y == 0, this does not mean anything at this point in time, due to it having no constraints.

Naively, we could apply the following constraint: y == 1 - t. This along with the previous y * t == 0 constraint means that y or t needs to be zero, but they both cannot be zero.

This equation however falls short when t != 0 because then t may not be 1. If t is non-zero, then y is also non-zero due to y == 1 - t and the equation y * t == 0 fails.

To fix, we introduce another free variable called z and apply the following constraint instead: y == 1 - t * z.

When t == 0, y is 1. When t != 0, the prover can set z = 1/t which will make y = 1 - t * 1/t = 0.

We now arrive at the conclusion that when t == 0, y is 1 and when t != 0, then y is 0.

Bringing it all together, We introduce two variables y and z, With the following equations:

y == 1 - tz (z is a value that is chosen to be the inverse of t by the prover)

y * t == 0

Lets convince ourselves that the prover cannot prove an untrue statement.

Assume that t == 0, can the prover return y == 0?

When t == 0, there is no way to make y be zero since y = 1 - 0 * z = 1.

Assume that t != 0, can the prover return y == 1?

By setting z to be 0, we can make y equal to 1. This is easily observed: y = 1 - t * 0 Now since y is one, this means that t needs to be zero, or else y * t == 0 will fail.

pub(crate) fn range_constraint( &mut self, witness: Witness, num_bits: u32, ) -> Result<(), RuntimeError>

Adds a constraint which ensure thats witness is an integer within the range [0, 2^{num_bits} - 1]

pub(crate) fn brillig_call( &mut self, predicate: Option<Expression<F>>, generated_brillig: &GeneratedBrillig<F>, inputs: Vec<BrilligInputs<F>>, outputs: Vec<BrilligOutputs>, brillig_function_index: BrilligFunctionId, stdlib_func: Option<BrilligStdlibFunc>, )

pub(crate) fn resolve_brillig_stdlib_call( &mut self, opcode_location: OpcodeLocation, brillig_function_index: BrilligFunctionId, )

pub(crate) fn last_acir_opcode_location(&self) -> OpcodeLocation

pub(crate) fn record_error_type( &mut self, selector: ErrorSelector, typ: ErrorType, )

pub(crate) fn generate_assertion_message_payload( &mut self, message: String, ) -> AssertionPayload<F>