noirc_evaluator::acir::acir_context
Struct AcirContext

pub(crate) struct AcirContext<F: AcirField> {
    brillig_stdlib: BrilligStdLib<F>,
    vars: FxHashMap<AcirVar, AcirVarData<F>>,
    constant_witnesses: FxHashMap<F, Witness>,
    pub(super) acir_ir: GeneratedAcir<F>,
    expression_width: ExpressionWidth,
    pub(super) warnings: Vec<SsaReport>,
}

Context object which holds the relationship between Variables(AcirVar) and types such as Expression and Witness which are placed into ACIR.

Fields
brillig_stdlib: BrilligStdLib<F>vars: FxHashMap<AcirVar, AcirVarData<F>>constant_witnesses: FxHashMap<F, Witness>acir_ir: GeneratedAcir<F>

An in-memory representation of ACIR.

This struct will progressively be populated based on the methods called. For example, If one was to add two Variables together, then the acir_ir will be populated to assert this addition.

expression_width: ExpressionWidthwarnings: Vec<SsaReport>
Implementations

impl<F: AcirField> AcirContext<F>

pub(crate) fn black_box_function( &mut self, name: BlackBoxFunc, inputs: Vec<AcirValue>, num_bits: Option<u32>, output_count: usize, predicate: Option<AcirVar>, ) -> Result<Vec<AcirVar>, RuntimeError>

Calls a Blackbox function on the given inputs and returns a given set of outputs to represent the result of the blackbox function.

pub(super) fn prepare_inputs_for_black_box_func( &mut self, inputs: Vec<AcirValue>, name: BlackBoxFunc, ) -> Result<Vec<Vec<FunctionInput<F>>>, RuntimeError>

pub(super) fn prepare_inputs_for_black_box_func_call( &mut self, inputs: Vec<AcirValue>, allow_constant_inputs: bool, ) -> Result<Vec<Vec<FunctionInput<F>>>, RuntimeError>

Black box function calls expect their inputs to be in a specific data structure (FunctionInput).

This function will convert AcirVar into FunctionInput for a blackbox function call.

fn value_to_function_input( &mut self, value: AcirValue, ) -> Result<FunctionInput<F>, RuntimeError>

Converts an AcirValue into a FunctionInput for use in black box function calls.

If the value can be evaluated to a constant, it returns FunctionInput::Constant

Otherwise, it creates or retrieves a witness variable and returns FunctionInput::Witness

impl<F: AcirField> AcirContext<F>

pub(crate) fn stdlib_brillig_call( &mut self, predicate: AcirVar, brillig_stdlib_func: BrilligStdlibFunc, inputs: Vec<AcirValue>, outputs: Vec<AcirType>, ) -> Result<Vec<AcirValue>, RuntimeError>

Generates a brillig call to a handwritten section of brillig bytecode.

pub(crate) fn brillig_call( &mut self, predicate: AcirVar, generated_brillig: &GeneratedBrillig<F>, inputs: Vec<AcirValue>, outputs: Vec<AcirType>, unsafe_return_values: bool, brillig_function_index: BrilligFunctionId, brillig_stdlib_func: Option<BrilligStdlibFunc>, ) -> Result<Vec<AcirValue>, RuntimeError>

fn brillig_array_input( &mut self, var_expressions: &mut Vec<Expression<F>>, input: AcirValue, ) -> Result<(), InternalError>

fn zeroed_array_output( &mut self, element_types: &[AcirType], size: usize, ) -> AcirValue

Recursively create zeroed-out acir values for returned arrays. This is necessary because a brillig returned array can have nested arrays as elements.

fn brillig_array_output( &mut self, element_types: &[AcirType], size: usize, ) -> (AcirValue, Vec<Witness>)

Recursively create acir values for returned arrays. This is necessary because a brillig returned array can have nested arrays as elements. A singular array of witnesses is collected for a top level array, by deflattening the assigned witnesses at each level.

impl<F: AcirField> AcirContext<F>

pub(super) fn new(brillig_stdlib: BrilligStdLib<F>) -> Self

pub(crate) fn set_expression_width(&mut self, expression_width: ExpressionWidth)

pub(crate) fn current_witness_index(&self) -> Witness

pub(crate) fn extract_witnesses(&self, inputs: &[AcirValue]) -> Vec<Witness>

pub(crate) fn add_constant(&mut self, constant: impl Into<F>) -> AcirVar

Adds a constant to the context and assigns a Variable to represent it

pub(crate) fn constant(&self, var: AcirVar) -> &F

Returns the constant represented by the given variable.

Panics: if the variable does not represent a constant.

pub(crate) fn add_variable(&mut self) -> AcirVar

Adds a Variable to the context, whose exact value is resolved at runtime.

fn mark_variables_equivalent( &mut self, lhs: AcirVar, rhs: AcirVar, ) -> Result<(), InternalError>

pub(crate) fn get_call_stack(&self) -> CallStack

pub(crate) fn set_call_stack(&mut self, call_stack: CallStack)

pub(crate) fn set_call_stack_helper(&mut self, call_stack: CallStackHelper)

pub(crate) fn get_or_create_witness_var( &mut self, var: AcirVar, ) -> Result<AcirVar, InternalError>

pub(crate) fn var_to_witness( &mut self, var: AcirVar, ) -> Result<Witness, InternalError>

Converts an AcirVar to a Witness

pub(crate) fn var_to_expression( &self, var: AcirVar, ) -> Result<Expression<F>, InternalError>

Converts an AcirVar to an Expression

pub(crate) fn is_constant_one(&self, var: &AcirVar) -> bool

True if the given AcirVar refers to a constant one value

pub(crate) fn is_constant(&self, var: &AcirVar) -> bool

True if the given AcirVar refers to a constant value

pub(crate) fn neg_var(&mut self, var: AcirVar) -> AcirVar

Adds a new Variable to context whose value will be constrained to be the negation of var.

Note: Variables are immutable.

pub(crate) fn inv_var( &mut self, var: AcirVar, predicate: AcirVar, ) -> Result<AcirVar, RuntimeError>

Adds a new Variable to context whose value will be constrained to be the inverse of var.

fn expect_one_var(results: Vec<AcirValue>) -> AcirVar

pub(crate) fn eq_var( &mut self, lhs: AcirVar, rhs: AcirVar, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVar that is 1 if lhs equals rhs and 0 otherwise.

pub(crate) fn xor_var( &mut self, lhs: AcirVar, rhs: AcirVar, typ: NumericType, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVar that is the XOR result of lhs and rhs.

pub(crate) fn and_var( &mut self, lhs: AcirVar, rhs: AcirVar, typ: NumericType, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVar that is the AND result of lhs and rhs.

pub(crate) fn or_var( &mut self, lhs: AcirVar, rhs: AcirVar, typ: NumericType, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVar that is the OR result of lhs and rhs.

pub(crate) fn assert_eq_var( &mut self, lhs: AcirVar, rhs: AcirVar, assert_message: Option<AssertionPayload<F>>, ) -> Result<(), RuntimeError>

Constrains the lhs and rhs to be equal.

fn get_assertion_payload_message( &self, assert_message: Option<&AssertionPayload<F>>, ) -> Option<String>

Returns Some(String) if the assertion message is present and it refers to a static string.

pub(crate) fn assert_neq_var( &mut self, lhs: AcirVar, rhs: AcirVar, predicate: AcirVar, assert_message: Option<AssertionPayload<F>>, ) -> Result<(), RuntimeError>

Constrains the lhs and rhs to be non-equal.

This is done by asserting the existence of an inverse for the value lhs - rhs. The constraint (lhs - rhs) * inverse == 1 will only be satisfiable if lhs and rhs are non-equal.

pub(crate) fn assert_zero_var( &mut self, var: AcirVar, msg: String, ) -> Result<(), RuntimeError>

Assert that an AcirVar equals zero, or fail with a message.

pub(crate) fn assert_always_fail( &mut self, msg: String, ) -> Result<(), RuntimeError>

Add an always-fail assertion with a message.

pub(crate) fn values_to_expressions_or_memory( &self, values: &[AcirValue], ) -> Result<Vec<ExpressionOrMemory<F>>, RuntimeError>

pub(crate) fn div_var( &mut self, lhs: AcirVar, rhs: AcirVar, typ: NumericType, predicate: AcirVar, ) -> Result<AcirVar, RuntimeError>

Adds a new Variable to context whose value will be constrained to be the division of lhs and rhs

pub(crate) fn mul_var( &mut self, lhs: AcirVar, rhs: AcirVar, ) -> Result<AcirVar, RuntimeError>

Adds a new Variable to context whose value will be constrained to be the multiplication of lhs and rhs

pub(crate) fn sub_var( &mut self, lhs: AcirVar, rhs: AcirVar, ) -> Result<AcirVar, RuntimeError>

Adds a new Variable to context whose value will be constrained to be the subtraction of lhs and rhs

pub(crate) fn add_var( &mut self, lhs: AcirVar, rhs: AcirVar, ) -> Result<AcirVar, RuntimeError>

Adds a new Variable to context whose value will be constrained to be the addition of lhs and rhs

fn add_mul_var( &mut self, lhs: AcirVar, k: F, rhs: AcirVar, ) -> Result<AcirVar, RuntimeError>

Adds a new Variable to context whose value will be constrained to be the expression lhs + k * rhs

pub(crate) fn not_var( &mut self, x: AcirVar, typ: NumericType, ) -> Result<AcirVar, RuntimeError>

Adds a new variable that is constrained to be the logical NOT of x.

fn euclidean_division_var( &mut self, lhs: AcirVar, rhs: AcirVar, bit_size: u32, predicate: AcirVar, ) -> Result<(AcirVar, AcirVar), RuntimeError>

Returns the quotient and remainder such that lhs = rhs * quotient + remainder

pub(super) fn bound_constraint_with_offset( &mut self, lhs: AcirVar, rhs: AcirVar, offset: AcirVar, bits: u32, predicate: AcirVar, ) -> Result<(), RuntimeError>

Generate constraints that are satisfied iff:

lhs < rhs , when offset is 1, or

lhs <= rhs, when offset is 0

bits is the bit size of lhs and rhs (or an upper bound of the bit size) predicate is assumed to be 0 or 1, and when it is 0, the generated constrains will never fail.

lhs<=rhs is done by constraining rhs-lhs to a bit size of bits:

if lhs<=rhs, 0 <= rhs-lhs <= b < 2^bits

if lhs>rhs, rhs-lhs = p+rhs-lhs > p-2^bits >= 2^bits (if log(p) >= bits + 1)

n.b: we do NOT check here that lhs and rhs are indeed bits size

pub(crate) fn modulo_var( &mut self, lhs: AcirVar, rhs: AcirVar, bit_size: u32, predicate: AcirVar, ) -> Result<AcirVar, RuntimeError>

Returns a variable which is constrained to be lhs mod rhs

pub(crate) fn range_constrain_var( &mut self, variable: AcirVar, bit_size: u32, message: Option<String>, predicate: AcirVar, ) -> Result<AcirVar, RuntimeError>

Constrains the AcirVar variable to be of type NumericType.

If predicate is false, the constrain will never fail.

pub(crate) fn truncate_var( &mut self, lhs: AcirVar, rhs: u32, max_bit_size: u32, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVar which will be constrained to be lhs mod 2^{rhs} In order to do this, we ‘simply’ perform euclidean division of lhs by 2^{rhs} The remainder of the division is then lhs mod 2^{rhs}

pub(crate) fn more_than_eq_var( &mut self, lhs: AcirVar, rhs: AcirVar, max_bits: u32, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVar which will be 1 if lhs >= rhs and 0 otherwise.

pub(crate) fn less_than_var( &mut self, lhs: AcirVar, rhs: AcirVar, bit_size: u32, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVar which will be 1 if lhs < rhs and 0 otherwise.

pub(crate) fn radix_decompose( &mut self, endian: Endian, input_var: AcirVar, radix_var: AcirVar, limb_count: u32, result_element_type: NumericType, ) -> Result<AcirValue, RuntimeError>

Returns a vector of AcirVars constrained to be the decomposition of the given input over given radix.

The AcirVars for the radix_var and limb_count_var must be a constant

pub(crate) fn bit_decompose( &mut self, endian: Endian, input_var: AcirVar, limb_count: u32, result_element_type: NumericType, ) -> Result<AcirValue, RuntimeError>

Returns AcirVars constrained to be the bit decomposition of the provided input

pub(crate) fn flatten( &mut self, value: AcirValue, ) -> Result<Vec<(AcirVar, NumericType)>, InternalError>

Recursive helper to flatten a single AcirValue into the result vector. This helper differs from flatten() on the AcirValue type, as this method has access to the AcirContext which lets us flatten an AcirValue::DynamicArray by reading its variables from memory.

pub(crate) fn finish( self, return_values: Vec<Witness>, warnings: Vec<SsaReport>, ) -> GeneratedAcir<F>

Terminates the context and takes the resulting GeneratedAcir

fn add_data(&mut self, data: AcirVarData<F>) -> AcirVar

Adds Data into the context and assigns it a Variable.

Variable can be seen as an index into the context. We use a two-way map so that it is efficient to lookup either the key or the value.

pub(crate) fn read_from_memory( &mut self, block_id: BlockId, index: &AcirVar, ) -> Result<AcirVar, InternalError>

Returns a Variable that is constrained to be the result of reading from the memory block_id at the given index.

pub(crate) fn write_to_memory( &mut self, block_id: BlockId, index: &AcirVar, value: &AcirVar, ) -> Result<(), InternalError>

Constrains the Variable value to be the new value located at index in the memory block_id.

pub(crate) fn initialize_return_data( &mut self, block_id: BlockId, init: Vec<Witness>, )

Insert the MemoryInit for the Return Data array, using the provided witnesses

pub(crate) fn initialize_array( &mut self, block_id: BlockId, len: usize, optional_value: Option<AcirValue>, databus: BlockType, ) -> Result<(), InternalError>

Initializes an array in memory with the given values optional_values. If optional_values is empty, then the array is initialized with zeros.

fn initialize_array_inner( &mut self, witnesses: &mut Vec<Witness>, input: AcirValue, ) -> Result<(), InternalError>

pub(crate) fn call_acir_function( &mut self, id: AcirFunctionId, inputs: Vec<AcirValue>, output_count: usize, predicate: AcirVar, ) -> Result<Vec<AcirVar>, RuntimeError>

pub(crate) fn generate_assertion_message_payload( &mut self, message: String, ) -> AssertionPayload<F>