noirc_evaluator::acir
Struct Context

struct Context<'a> {
    ssa_values: FxHashMap<Id<Value>, AcirValue>,
    current_side_effects_enabled_var: AcirVar,
    acir_context: AcirContext<FieldElement, Bn254BlackBoxSolver>,
    initialized_arrays: HashSet<BlockId>,
    memory_blocks: FxHashMap<Id<Value>, BlockId>,
    internal_memory_blocks: FxHashMap<Id<Value>, BlockId>,
    internal_mem_block_lengths: FxHashMap<BlockId, usize>,
    max_block_id: u32,
    data_bus: DataBus,
    shared_context: &'a mut SharedContext<FieldElement>,
    brillig: &'a Brillig,
    brillig_options: &'a BrilligOptions,
}

Context struct for the acir generation pass. May be similar to the Evaluator struct in the current SSA IR.

Fields
ssa_values: FxHashMap<Id<Value>, AcirValue>

Maps SSA values to AcirVar.

This is needed so that we only create a single AcirVar per SSA value. Before creating an AcirVar for an SSA value, we check this map. If an AcirVar already exists for this Value, we return the AcirVar.

current_side_effects_enabled_var: AcirVar

The AcirVar that describes the condition belonging to the most recently invoked SideEffectsEnabled instruction.

acir_context: AcirContext<FieldElement, Bn254BlackBoxSolver>

Manages and builds the AcirVars to which the converted SSA values refer.

initialized_arrays: HashSet<BlockId>

Track initialized acir dynamic arrays

An acir array must start with a MemoryInit ACIR opcodes and then have MemoryOp opcodes This set is used to ensure that a MemoryOp opcode is only pushed to the circuit if there is already a MemoryInit opcode.

memory_blocks: FxHashMap<Id<Value>, BlockId>

Maps SSA values to BlockId A BlockId is an ACIR structure which identifies a memory block Each acir memory block corresponds to a different SSA array.

internal_memory_blocks: FxHashMap<Id<Value>, BlockId>

Maps SSA values to a BlockId used internally A BlockId is an ACIR structure which identifies a memory block Each memory blocks corresponds to a different SSA value which utilizes this internal memory for ACIR generation.

internal_mem_block_lengths: FxHashMap<BlockId, usize>

Maps an internal memory block to its length

This is necessary to keep track of an internal memory block’s size. We do not need a separate map to keep track of memory_blocks as the length is set when we construct a AcirValue::DynamicArray and is tracked as part of the AcirValue in the ssa_values map. The length of an internal memory block is determined before an array operation takes place thus we track it separate here in this map.

max_block_id: u32

Number of the next BlockId, it is used to construct a new BlockId

data_bus: DataBusshared_context: &'a mut SharedContext<FieldElement>

Contains state that is generated and also used across ACIR functions

brillig: &'a Brilligbrillig_options: &'a BrilligOptions

Options affecting Brillig code generation.

Implementations

impl Context<'_>

pub(super) fn block_id(&mut self, value: &ValueId) -> BlockId

Get the BlockId corresponding to the ValueId If there is no matching BlockId, we create a new one.

fn internal_block_id(&mut self, value: &ValueId) -> BlockId

Get the next BlockId for internal memory used during ACIR generation. This is useful for referencing information that can only be computed dynamically, such as the type structure of non-homogenous arrays.

pub(super) fn initialize_databus( &mut self, witnesses: &Vec<Witness>, dfg: &DataFlowGraph, ) -> Result<(), RuntimeError>

pub(super) fn handle_array_operation( &mut self, instruction: Id<Instruction>, dfg: &DataFlowGraph, ) -> Result<(), RuntimeError>

Handles an ArrayGet or ArraySet instruction. To set an index of the array (and create a new array in doing so), pass Some(value) for store_value. To just retrieve an index of the array, pass None for store_value.

fn handle_constant_index_wrapper( &mut self, instruction: Id<Instruction>, dfg: &DataFlowGraph, array: ValueId, index: ValueId, store_value: Option<ValueId>, ) -> Result<bool, RuntimeError>

fn handle_constant_index( &mut self, instruction: Id<Instruction>, dfg: &DataFlowGraph, array: Vector<AcirValue>, index: FieldElement, store_value: Option<AcirValue>, ) -> Result<bool, RuntimeError>

Handle constant index: if there is no predicate and we have the array values, we can perform the operation directly on the array

fn convert_array_operation_inputs( &mut self, array_id: ValueId, dfg: &DataFlowGraph, index: ValueId, store_value: Option<ValueId>, offset: usize, ) -> Result<(AcirVar, Option<AcirValue>), RuntimeError>

We need to properly setup the inputs for array operations in ACIR. From the original SSA values we compute the following AcirVars:

new_index is the index of the array. ACIR memory operations work with a flat memory, so we fully flattened the specified index in case we have a nested array. The index for SSA array operations only represents the flattened index of the current array. Thus internal array element type sizes need to be computed to accurately transform the index.

predicate_index is offset, or the index if the predicate is true

new_value is the optional value when the operation is an array_set When there is a predicate, it is predicate*value + (1-predicate)*dummy, where dummy is the value of the array at the requested index. It is a dummy value because in the case of a false predicate, the value stored at the requested index will be itself.

fn convert_array_set_store_value( &mut self, store_value: &AcirValue, dummy_value: &AcirValue, ) -> Result<AcirValue, RuntimeError>

fn get_from_call_data( &mut self, offset: &mut AcirVar, call_data_block: BlockId, typ: &Type, ) -> Result<AcirValue, RuntimeError>

Returns the acir value at the provided databus offset

fn array_get( &mut self, instruction: Id<Instruction>, array: ValueId, var_index: AcirVar, dfg: &DataFlowGraph, index_side_effect: bool, ) -> Result<AcirValue, RuntimeError>

Generates a read opcode for the array index_side_effect == false means that we ensured var_index will have a type matching the value in the array

pub(super) fn array_get_value( &mut self, ssa_type: &Type, block_id: BlockId, var_index: &mut AcirVar, ) -> Result<AcirValue, RuntimeError>

pub(super) fn array_zero_value( &mut self, ssa_type: &Type, ) -> Result<AcirValue, RuntimeError>

Construct a value with all zero values, which we can use to provide a default value when we cannot use array_get_value because the array length itself is zero, yet we also don’t want a memory operation to fail, because the operation will never actually run, because we know that the side effect variable is false.

fn array_set( &mut self, instruction: Id<Instruction>, var_index: AcirVar, store_value: AcirValue, dfg: &DataFlowGraph, mutate_array: bool, ) -> Result<(), RuntimeError>

If mutate_array is:

true: Mutate the array directly

false: Copy the array and generates a write opcode on the new array. This is generally very inefficient and should be avoided if possible. Currently this is controlled by SSA’s array set optimization pass.

pub(super) fn array_set_value( &mut self, value: &AcirValue, block_id: BlockId, var_index: &mut AcirVar, ) -> Result<(), RuntimeError>

pub(super) fn init_element_type_sizes_array( &mut self, array_typ: &Type, array_id: ValueId, supplied_acir_value: Option<&AcirValue>, dfg: &DataFlowGraph, ) -> Result<BlockId, RuntimeError>

pub(super) fn read_array( &mut self, array: AcirValue, ) -> Result<Vector<AcirValue>, RuntimeError>

pub(super) fn copy_array( &mut self, source: ValueId, destination: BlockId, dfg: &DataFlowGraph, ) -> Result<(), RuntimeError>

fn read_dynamic_array( &mut self, source: BlockId, array_len: usize, ) -> Result<Vector<AcirValue>, RuntimeError>

fn copy_dynamic_array( &mut self, source: BlockId, destination: BlockId, array_len: usize, ) -> Result<(), RuntimeError>

pub(super) fn get_flattened_index( &mut self, array_typ: &Type, array_id: ValueId, var_index: AcirVar, dfg: &DataFlowGraph, ) -> Result<AcirVar, RuntimeError>

pub(super) fn flattened_size( &mut self, array: ValueId, dfg: &DataFlowGraph, ) -> usize

pub(super) fn ensure_array_is_initialized( &mut self, array: ValueId, dfg: &DataFlowGraph, ) -> Result<BlockId, RuntimeError>

pub(super) fn initialize_array( &mut self, array: BlockId, len: usize, value: Option<AcirValue>, ) -> Result<(), InternalError>

Initializes an array with the given values and caches the fact that we have initialized this array.

impl<'a> Context<'a>

fn new( shared_context: &'a mut SharedContext<FieldElement>, expression_width: ExpressionWidth, brillig: &'a Brillig, brillig_stdlib: BrilligStdLib<FieldElement>, brillig_options: &'a BrilligOptions, ) -> Context<'a>

fn convert_ssa_function( self, ssa: &Ssa, function: &Function, ) -> Result<Option<GeneratedAcir<FieldElement>>, RuntimeError>

fn convert_acir_main( self, main_func: &Function, ssa: &Ssa, ) -> Result<GeneratedAcir<FieldElement>, RuntimeError>

fn convert_brillig_main( self, main_func: &Function, ) -> Result<GeneratedAcir<FieldElement>, RuntimeError>

fn convert_ssa_block_params( &mut self, params: &[ValueId], dfg: &DataFlowGraph, ) -> Result<Vec<Witness>, RuntimeError>

Adds and binds AcirVars for each numeric block parameter or block parameter array element.

fn convert_ssa_block_param( &mut self, param_type: &Type, ) -> Result<AcirValue, RuntimeError>

fn create_value_from_type( &mut self, param_type: &Type, make_var: &mut impl FnMut(&mut Self, NumericType) -> Result<AcirVar, RuntimeError>, ) -> Result<AcirValue, RuntimeError>

fn add_numeric_input_var( &mut self, numeric_type: &NumericType, ) -> Result<AcirVar, RuntimeError>

Creates an AcirVar corresponding to a parameter witness to appears in the abi. A range constraint is added if the numeric type requires it.

This function is used not only for adding numeric block parameters, but also for adding any array elements that belong to reference type block parameters.

fn convert_ssa_instruction( &mut self, instruction_id: Id<Instruction>, dfg: &DataFlowGraph, ssa: &Ssa, ) -> Result<Vec<SsaReport>, RuntimeError>

Converts an SSA instruction into its ACIR representation

fn convert_ssa_call( &mut self, instruction: &Instruction, dfg: &DataFlowGraph, ssa: &Ssa, result_ids: &[ValueId], ) -> Result<Vec<SsaReport>, RuntimeError>

fn handle_ssa_call_outputs( &mut self, result_ids: &[ValueId], output_values: Vec<AcirValue>, dfg: &DataFlowGraph, ) -> Result<(), RuntimeError>

fn gen_brillig_parameters( &self, values: &[ValueId], dfg: &DataFlowGraph, ) -> Vec<BrilligParameter>

fn define_result( &mut self, dfg: &DataFlowGraph, instruction: Id<Instruction>, result: AcirValue, )

Remember the result of an instruction returning a single value

fn define_result_var( &mut self, dfg: &DataFlowGraph, instruction: Id<Instruction>, result: AcirVar, )

Remember the result of instruction returning a single numeric value

fn get_num_return_witnesses( &self, terminator: &TerminatorInstruction, dfg: &DataFlowGraph, ) -> usize

Converts an SSA terminator’s return values into their ACIR representations

fn convert_ssa_return( &mut self, terminator: &TerminatorInstruction, dfg: &DataFlowGraph, ) -> Result<(Vec<AcirVar>, Vec<SsaReport>), RuntimeError>

Converts an SSA terminator’s return values into their ACIR representations

fn convert_value(&mut self, value_id: ValueId, dfg: &DataFlowGraph) -> AcirValue

Gets the cached AcirVar that was converted from the corresponding ValueId. If it does not already exist in the cache, a conversion is attempted and cached for simple values that require no further context such as numeric types - values requiring more context should have already been cached elsewhere.

Conversion is assumed to have already been performed for instruction results and block parameters. This is because block parameters are converted before anything else, and because instructions results are converted when the corresponding instruction is encountered. (An instruction result cannot be referenced before the instruction occurs.)

It is not safe to call this function on value ids that represent addresses. Instructions involving such values are evaluated via a separate path and stored in ssa_value_to_array_address instead.

fn convert_numeric_value( &mut self, value_id: ValueId, dfg: &DataFlowGraph, ) -> Result<AcirVar, InternalError>

fn convert_ssa_binary( &mut self, binary: &Binary, dfg: &DataFlowGraph, predicate: AcirVar, ) -> Result<AcirVar, RuntimeError>

Processes a binary operation and converts the result into an AcirVar

fn check_unsigned_overflow( &mut self, result: AcirVar, bit_size: u32, binary: &Binary, predicate: AcirVar, ) -> Result<AcirVar, RuntimeError>

Adds a range check against the bit size of the result of addition, subtraction or multiplication

fn type_of_binary_operation(&self, binary: &Binary, dfg: &DataFlowGraph) -> Type

Operands in a binary operation are checked to have the same type.

In Noir, binary operands should have the same type due to the language semantics.

There are some edge cases to consider:

Constants are not explicitly type casted, so we need to check for this and return the type of the other operand, if we have a constant.

0 is not seen as Field 0 but instead as Unit 0

TODO: The latter seems like a bug, if we cannot differentiate between a function returning TODO nothing and a 0.

TODO: This constant coercion should ideally be done in the type checker.

fn convert_ssa_truncate( &mut self, value_id: ValueId, bit_size: u32, max_bit_size: u32, dfg: &DataFlowGraph, ) -> Result<AcirVar, RuntimeError>

Returns an AcirVarthat is constrained to be result of the truncation.

fn convert_ssa_intrinsic_call( &mut self, intrinsic: Intrinsic, arguments: &[ValueId], dfg: &DataFlowGraph, result_ids: &[ValueId], ) -> Result<Vec<AcirValue>, RuntimeError>

Returns a vector of AcirVars constrained to be result of the function call.

The function being called is required to be intrinsic.

fn convert_vars_to_values( &self, vars: Vec<AcirVar>, dfg: &DataFlowGraph, result_ids: &[ValueId], ) -> Vec<AcirValue>

Convert a Vec<AcirVar> into a Vec<AcirValue> using the given result ids. If the type of a result id is an array, several acir vars are collected into a single AcirValue::Array of the same length. If the type of a result id is a slice, the slice length must precede it and we can convert to an AcirValue::Array when the length is known (constant).

fn convert_var_type_to_values( result_type: &Type, vars: &mut impl Iterator<Item = AcirVar>, ) -> AcirValue

Recursive helper for convert_vars_to_values. If the given result_type is an array of length N, this will create an AcirValue::Array with the first N elements of the given iterator. Otherwise, the result is a single AcirValue::Var wrapping the first element of the iterator.