noirc_evaluator::acir
Module types

Structs

AcirDynamicArray 🔒

Represents an array whose size is not known during compile time, and which might be indexed with a dynamic index.

AcirVar 🔒

A Reference to an AcirVarData

Enums

AcirType 🔒

High level Type descriptor for Variables.

AcirValue 🔒

All possible variants of values in ACIR.

Functions

collect_flat_numeric_types 🔒

Helper function for flat_numeric_types that recursively collects all numeric types into flat_types.

collect_fully_flattened_numeric_types 🔒

Helper function for flat_element_types that fully flattens arrays using the length. This is different from collect_flat_numeric_types which returns just the first element.

flat_element_types 🔒

Returns the fully flattened numeric types for one element of a vector/array, recursively flattening nested arrays. For example, for Vector([(u32, u32, [Field; 3])]), this returns [u32, u32, Field, Field, Field].

flat_numeric_types 🔒

Assumes typ is an array or vector type with nested numeric types, arrays or vectors (recursively) and returns a flat list of all the contained numeric types. Panics if self is not an array or vector type or if a function or reference type is found along the way.