noirc_evaluator::brillig::brillig_check
Struct AdvisoryCollector

struct AdvisoryCollector<'a> {
    addr_range: &'a Range<usize>,
    reads_in_descendants: &'a HashSet<MemoryAddress>,
    reads: HashMap<MemoryAddress, usize>,
    writes: HashMap<MemoryAddress, usize>,
    advisories: HashMap<usize, Vec<OpcodeAdvisory>>,
    in_call_region: u8,
    in_return_region: bool,
    is_fallible: bool,
}

Collect advisories about the opcodes in a single block.

Assumes that we are visiting the opcodes going backwards.

Fields
addr_range: &'a Range<usize>reads_in_descendants: &'a HashSet<MemoryAddress>

Addresses read in any of the descendants of the block.

reads: HashMap<MemoryAddress, usize>

Last location the address was read from in this block.

writes: HashMap<MemoryAddress, usize>

Last location the address was written to in this block.

advisories: HashMap<usize, Vec<OpcodeAdvisory>>

Advisories collected in this block.

in_call_region: u8

Indicate that we are in a region where call parameters are being passed.

2 means we are in a call region

1 means we are 1 step before the start of the call region

0 means we are not in a call region

in_return_region: bool

Indicate that we are in the region where we are passing out return parameters.

is_fallible: bool

Indicate whether the current opcode can fail, which means it needs to be kept even if its value is unused.

Implementations

impl<'a> AdvisoryCollector<'a>

fn new( addr_range: &'a Range<usize>, reads_in_descendants: &'a HashSet<MemoryAddress>, ) -> Self

fn into_advisories(self) -> HashMap<usize, Vec<OpcodeAdvisory>>

fn addr_in_range(&self, addr: &MemoryAddress) -> bool

fn add_advisory(&mut self, location: usize, advisory: OpcodeAdvisory)