noirc_evaluator::brillig::brillig_check
Trait OpcodeAddressVisitor

trait OpcodeAddressVisitor {
    // Required methods
    fn should_visit_opcode<F: AcirField>(&mut self, opcode: &Opcode<F>) -> bool;
    fn read(&mut self, addr: &MemoryAddress, location: usize);
    fn write(&mut self, addr: &MemoryAddress, location: usize);

    // Provided methods
    fn visit_opcode<F: AcirField>(
        &mut self,
        opcode: &Opcode<F>,
        location: usize,
    ) { ... }
    fn read_heap_array(&mut self, array: &HeapArray, location: usize) { ... }
    fn read_heap_vector(&mut self, vector: &HeapVector, location: usize) { ... }
    fn read_value_or_array(&mut self, value: &ValueOrArray, location: usize) { ... }
    fn write_heap_array(&mut self, array: &HeapArray, location: usize) { ... }
    fn write_heap_vector(&mut self, vector: &HeapVector, location: usize) { ... }
    fn write_value_or_array(&mut self, value: &ValueOrArray, location: usize) { ... }
    fn visit_black_box_op(&mut self, op: &BlackBoxOp, location: usize) { ... }
}

Required Methods

fn should_visit_opcode<F: AcirField>(&mut self, opcode: &Opcode<F>) -> bool

Decide if an opcode should be visited.

Can mutate self to set flags that allows it to ignore upcoming opcodes as well.

fn read(&mut self, addr: &MemoryAddress, location: usize)

Called with all the addresses read by the opcode.

fn write(&mut self, addr: &MemoryAddress, location: usize)

Called with all addresses written by the opcode.

Provided Methods

fn visit_opcode<F: AcirField>(&mut self, opcode: &Opcode<F>, location: usize)

Call with each opcode to record reads and writes.

The expected order of visit depends on the purpose of the visitor implementation.

fn read_heap_array(&mut self, array: &HeapArray, location: usize)

fn read_heap_vector(&mut self, vector: &HeapVector, location: usize)

fn read_value_or_array(&mut self, value: &ValueOrArray, location: usize)

fn write_heap_array(&mut self, array: &HeapArray, location: usize)

fn write_heap_vector(&mut self, vector: &HeapVector, location: usize)

fn write_value_or_array(&mut self, value: &ValueOrArray, location: usize)

fn visit_black_box_op(&mut self, op: &BlackBoxOp, location: usize)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.