noirc_evaluator::brillig

Module brillig_gen

The code generation logic for converting crate::ssa objects into their respective Brillig artifacts.

Modules

brillig_block 🔒

Module containing Brillig-gen logic specific to an SSA function’s basic blocks.

brillig_block_variables 🔒

This module handles allocation, tracking, and lifetime management of variables within a Brillig compiled SSA basic block.

brillig_call 🔒

brillig_fn 🔒

Module containing Brillig-gen logic specific to SSA Function’s.

brillig_globals 🔒

Codegen for converting SSA globals to Brillig bytecode.

brillig_instructions 🔒

constant_allocation 🔒

This module analyzes the usage of constants in a given function and decides an allocation point for them. The allocation point will be the common dominator of all the places where the constant is used. By allocating in the common dominator, we can cache the constants for all subsequent uses.

variable_liveness 🔒

This module analyzes the liveness of variables (non-constant values) throughout a function. It uses the approach detailed in the section 4.2 of this paper https://inria.hal.science/inria-00558509v2/document

Functions

gen_brillig_for 🔒

Generates a complete Brillig entry point artifact for a given SSA-level Function, linking all dependencies.