noirc_evaluator::brillig::brillig_gen::brillig_block
Struct BrilligBlock

pub(crate) struct BrilligBlock<'block, Registers: RegisterAllocator> {
    pub(crate) function_context: &'block mut FunctionContext,
    pub(crate) block_id: BasicBlockId,
    pub(crate) brillig_context: &'block mut BrilligContext<FieldElement, Registers>,
    pub(crate) variables: BlockVariables,
    pub(crate) last_uses: FxHashMap<Id<Instruction>, FxHashSet<ValueId>>,
    pub(crate) globals: &'block FxHashMap<ValueId, BrilligVariable>,
    pub(crate) hoisted_global_constants: &'block FxHashMap<(FieldElement, NumericType), BrilligVariable>,
    pub(crate) building_globals: bool,
}

Context structure for compiling a function block into Brillig bytecode.

Fields
function_context: &'block mut FunctionContext

Per-function context shared across all of a function’s blocks

block_id: BasicBlockId

The basic block that is being converted

brillig_context: &'block mut BrilligContext<FieldElement, Registers>

Context for creating brillig opcodes

variables: BlockVariables

Tracks the available variable during the codegen of the block

last_uses: FxHashMap<Id<Instruction>, FxHashSet<ValueId>>

For each instruction, the set of values that are not used anymore after it.

globals: &'block FxHashMap<ValueId, BrilligVariable>

Mapping of SSA ValueIds to their already instantiated values in the Brillig IR.

hoisted_global_constants: &'block FxHashMap<(FieldElement, NumericType), BrilligVariable>

Pre-instantiated constants values shared across functions which have hoisted to the global memory space.

building_globals: bool

Status variable for whether we are generating Brillig bytecode for a function or globals. This is primarily used for gating local variable specific logic. For example, liveness analysis for globals is unnecessary (and adds complexity), and instead globals live throughout the entirety of the program.

Implementations

impl<'block, Registers: RegisterAllocator> BrilligBlock<'block, Registers>

pub(crate) fn compile_block( function_context: &'block mut FunctionContext, brillig_context: &'block mut BrilligContext<FieldElement, Registers>, block_id: BasicBlockId, dfg: &DataFlowGraph, call_stacks: &mut CallStackHelper, globals: &'block HashMap<ValueId, BrilligVariable>, hoisted_global_constants: &'block FxHashMap<(FieldElement, NumericType), BrilligVariable>, )

Converts an SSA basic block into a sequence of Brillig opcodes.

This method contains the necessary initial variable and register setup for compiling an SSA block by accessing the pre-computed liveness context.

pub(crate) fn compile_globals( &mut self, globals: &DataFlowGraph, used_globals: &HashSet<ValueId>, call_stacks: &mut CallStackHelper, hoisted_global_constants: &BTreeSet<(FieldElement, NumericType)>, ) -> HashMap<(FieldElement, NumericType), Allocated<BrilligVariable, Registers>>

Converts SSA globals into Brillig global values.

Global values can be:

Numeric constants

Instructions that compute global values

Pre-hoisted constants (shared across functions and stored in global memory)

This method expects SSA globals to already be converted to a DataFlowGraph as to share codegen logic with standard SSA function blocks.

This method also emits any necessary debugging initialization logic (e.g., allocating a counter used to track array copies).

Returns

A map of hoisted (constant, type) pairs to their allocated Brillig variables, which are used to resolve references to these constants throughout Brillig lowering.

Panics

Globals graph contains values other than a constant or instruction

fn convert_block( &mut self, dfg: &DataFlowGraph, call_stacks: &mut CallStackHelper, )

Internal method for BrilligBlock::compile_block that actually kicks off the Brillig compilation process.

At this point any Brillig context should be contained in BrilligBlock, and this function should only need to accept external SSA and debugging structures.

fn create_block_label_for_current_function( &self, block_id: BasicBlockId, ) -> Label

Creates a unique global label for a block.

This uses the current function’s function ID and the block ID, making the assumption that the block ID passed in belongs to this function.

fn create_block_label( function_id: Id<Function>, block_id: BasicBlockId, ) -> Label

Creates a unique label for a block using the function Id and the block ID.

We implicitly assume that the function ID and the block ID is enough for us to create a unique label across functions and blocks.

This is so that during linking there are no duplicates or labels being overwritten.

fn convert_ssa_terminator( &mut self, terminator_instruction: &TerminatorInstruction, dfg: &DataFlowGraph, )

Converts an SSA terminator instruction into the necessary opcodes:

allocates the hoisted constants which are used by dominated blocks

for jumps:

copies the arguments to the registers allocated in Self::convert_block_params

adds jump opcodes to the labels of the destination blocks

for return it allocates registers for the return values and copies from variables.

fn convert_block_params(&mut self, dfg: &DataFlowGraph)

Allocates the block parameters that the given block is defining.

We don’t allocate the block parameters of the block itself here, we allocate the parameters the block is defining for the descendant blocks it immediately dominates. Since predecessors to a block have to know where the parameters of the block are allocated to pass data to it in Self::convert_ssa_terminator, the block parameters need to be defined/allocated before the given block. VariableLiveness decides when the block parameters are defined.

For the entry block, the defined block params will be the params of the function + any extra params of blocks it’s the immediate dominator of.

fn convert_ssa_instruction( &mut self, instruction_id: Id<Instruction>, dfg: &DataFlowGraph, call_stacks: &mut CallStackHelper, )

Converts an SSA instruction into a sequence of Brillig opcodes.

If this is the last time a variable is used in this block, its memory slot gets deallocated, unless it’s a global.

fn convert_cast( &mut self, destination: SingleAddrVariable, source: SingleAddrVariable, )

Converts an SSA cast to a sequence of Brillig opcodes. Casting is only necessary when shrinking the bit size of a numeric value.

fn initialize_constants( &mut self, dfg: &DataFlowGraph, location: InstructionLocation, )

Initializes constants allocated to a InstructionLocation by ConstantAllocation.

It is expected that this method is called before converting an SSA instruction to Brillig and the constants to be initialized have been precomputed and stored in FunctionContext::constant_allocation.

pub(crate) fn convert_ssa_value( &mut self, value_id: ValueId, dfg: &DataFlowGraph, ) -> BrilligVariable

Converts an SSA ValueId into a BrilligVariable. Initializes if necessary, or returns an existing allocation.

This method also first checks whether the SSA value is a hoisted global constant. If it has already been initialized in the global space, we return the already existing variable. If an SSA value is a Value::Global, we check whether the value exists in the BrilligBlock::globals map, otherwise the method panics.

pub(crate) fn convert_ssa_single_addr_value( &mut self, value_id: ValueId, dfg: &DataFlowGraph, ) -> SingleAddrVariable

Converts an SSA ValueId into a single MemoryAddress. Initializes if necessary.

Panics if the value was converted to an array or vector.

fn get_hoisted_global( &self, dfg: &DataFlowGraph, value_id: ValueId, ) -> Option<BrilligVariable>

If the supplied value is a numeric constant check whether it is exists within the precomputed hoisted globals map. If the variable exists as a hoisted global return that value, otherwise return None.

fn codegen_not( &mut self, instruction_id: Id<Instruction>, value: ValueId, dfg: &DataFlowGraph, )

Define the result variable, convert the value, then generate Brillig opcodes to negate the variable.

fn codegen_truncate( &mut self, instruction_id: Id<Instruction>, value: ValueId, bit_size: u32, dfg: &DataFlowGraph, )

Define the result variable, convert the value, then generate Brillig opcodes to truncate the variable.

fn codegen_cast( &mut self, instruction_id: Id<Instruction>, value: ValueId, dfg: &DataFlowGraph, )

Define the result variable, convert the value, then generate Brillig opcodes to truncate the variable.

fn codegen_range_check( &mut self, value: ValueId, max_bit_size: u32, assert_message: Option<&String>, dfg: &DataFlowGraph, )

Convert the value, and if its bit size is larger than the maximum bit size, then generate opcodes to constrain the value to be no greater than the maximum value as a Field.

fn codegen_if_else( &mut self, instruction_id: Id<Instruction>, then_condition: ValueId, then_value: ValueId, else_value: ValueId, dfg: &DataFlowGraph, )

Convert the condition and the values, define a variable for the result, then generate opcodes for a conditional move.

Panics if the then_value and else_value are incompatible.

impl<Registers: RegisterAllocator> BrilligBlock<'_, Registers>

fn write_variables( &mut self, write_pointer: MemoryAddress, variables: &[BrilligVariable], )

Take a list of BrilligVariable and copy the memory they point at to the write_pointer, increasing the address by 1 between each variable.

pub(crate) fn vector_push_back_operation( &mut self, target_vector: BrilligVector, source_len: SingleAddrVariable, source_vector: BrilligVector, variables_to_insert: &[BrilligVariable], )

Prepare a vector for pushing a number of new (flattened) items to the back, then write those variables to the returned write pointer.

pub(crate) fn vector_push_front_operation( &mut self, target_vector: BrilligVector, source_len: SingleAddrVariable, source_vector: BrilligVector, variables_to_insert: &[BrilligVariable], )

Prepare a vector for pushing a number of new (flattened) items to the front, then write those variables to the returned write pointer.

fn read_variables( &mut self, read_pointer: MemoryAddress, variables: &[BrilligVariable], )

Read the memory starting at the pointer into successive variables.

pub(crate) fn vector_pop_front_operation( &mut self, target_vector: BrilligVector, source_len: SingleAddrVariable, source_vector: BrilligVector, removed_items: &[BrilligVariable], )

Read the popped number of variables from the front of the source vector, then create the target vector from the source by skipping the number of popped items.

pub(crate) fn vector_pop_back_operation( &mut self, target_vector: BrilligVector, source_len: SingleAddrVariable, source_vector: BrilligVector, removed_items: &[BrilligVariable], )

Create new vector with a number of (flattened) items popped from the back, then read the popped items into the variables representing the removed items.

pub(crate) fn vector_insert_operation( &mut self, target_vector: BrilligVector, source_vector: BrilligVector, index: SingleAddrVariable, items: &[BrilligVariable], )

Prepare a vector for inserting a number of (flattened) items at a specific index by making a hole for them, then write the variables to the returned write pointer.

pub(crate) fn vector_remove_operation( &mut self, target_vector: BrilligVector, source_vector: BrilligVector, index: SingleAddrVariable, removed_items: &[BrilligVariable], )

Read a number of (flattened) items at a specific index of a vector into the variables representing the removed items, then create a new vector with the same number of items removed and subsequent items shifted to the left.

impl<Registers: RegisterAllocator> BrilligBlock<'_, Registers>

fn convert_ssa_foreign_call( &mut self, func_name: &str, arguments: &[ValueId], instruction_id: Id<Instruction>, dfg: &DataFlowGraph, )

Converts a foreign function call into Brillig bytecode.

Foreign functions are external host functions that interact with the Brillig VM. This method handles the conversion of inputs/outputs and manages memory allocation.

fn output_variables_to_destinations( &mut self, output_variables: &[BrilligVariable], ) -> Vec<Allocated<ValueOrArray, Registers>>

Convert output BrilligVariables to [ValueOrArray] destinations on the heap.

fn convert_ssa_field_less_than( &mut self, arguments: &[ValueId], instruction_id: Id<Instruction>, dfg: &DataFlowGraph, )

Converts a field less than comparison intrinsic to Brillig bytecode.

Compares two field elements and returns a boolean result.

fn convert_ssa_black_box_call( &mut self, bb_func: &BlackBoxFunc, arguments: &[ValueId], instruction_id: Id<Instruction>, dfg: &DataFlowGraph, )

Converts a black box function call into Brillig bytecode.

Black box functions are native cryptographic operations or other optimized primitives. Some black box functions (ECDSA, MultiScalarMul, EmbeddedCurveAdd) have a predicate argument in SSA that is removed for Brillig as CFG flattening is not needed.

fn convert_ssa_as_vector( &mut self, arguments: &[ValueId], instruction_id: Id<Instruction>, dfg: &DataFlowGraph, )

Converts an array to a vector by copying the array contents into a vector.

This intrinsic converts a fixed-size array into a dynamically-sized vector (vector).

pub(crate) fn codegen_call( &mut self, instruction_id: Id<Instruction>, func: ValueId, arguments: &[ValueId], dfg: &DataFlowGraph, )

impl<Registers: RegisterAllocator> BrilligBlock<'_, Registers>

fn allocate_external_call_results( &mut self, results: &[ValueId], dfg: &DataFlowGraph, ) -> Vec<BrilligVariable>

Allocates a variable to hold the result of an external function call (e.g., foreign or black box). For more information about foreign function calls in Brillig take a look at the [foreign call opcode][acvm::acir::brillig::Opcode::ForeignCall].

This is typically used during Brillig codegen for calls to Value::ForeignFunction, where external host functions return values back into the program.

Numeric types and fixed-sized array results are directly allocated.

As vectors are determined at runtime they are allocated differently:

Allocates memory for a BrilligVariable::BrilligVector, which holds a pointer and dynamic size.

Initializes the pointer using the free memory pointer.

The actual size will be updated after the foreign function call returns.

Returns

A BrilligVariable representing the allocated memory structure to store the foreign call’s result.

Panics

If there is a vector among the output variables and it’s followed by another vector: when we allocate memory for a vector, we don’t know its length, so it just points at the current free memory pointer without increasing it; a second vector gets allocated at the same memory slot.

fn allocate_foreign_call_result_array( &mut self, typ: &Type, array: BrilligArray, )

Recursively allocates memory on the heap for a nested array returned from a foreign function call.

Panics

If the provided typ is not an array.

If any vector types are encountered within the nested structure, since vectors require runtime size information and cannot be allocated statically here.

fn convert_ssa_function_call( &mut self, func_id: Id<Function>, arguments: &[ValueId], dfg: &DataFlowGraph, result_ids: &[ValueId], )

Internal method to codegen an crate::ssa::ir::instruction::Instruction::Call to a Value::Function

fn update_vector_length( &mut self, target_len: SingleAddrVariable, source_len: SingleAddrVariable, binary_op: BrilligBinaryOp, )

Increase or decrease the vector length by 1.

Vectors have a tuple structure (vector length, vector contents) to enable logic that uses dynamic vector lengths (such as with merging vectors in the flattening pass). This method codegens an update to the vector length.

The binary operation performed on the vector length is always an addition or subtraction of 1. This is because the vector length holds the user length (length as displayed by a .len() call), and not a flattened length used internally to represent arrays of tuples. The length inside of RegisterOrMemory::HeapVector represents the entire flattened number of fields in the vector.

Note that when we subtract a value, we expect that there is a constraint in SSA to check that the length isn’t already 0. We could add a constraint opcode here, but if it’s in SSA, there is a chance it can be optimized out.

fn convert_ssa_identity_call( &mut self, arguments: &[ValueId], dfg: &DataFlowGraph, result_ids: &[ValueId], )

Copy the input arguments to the results.

fn convert_ssa_vector_intrinsic_call( &mut self, dfg: &DataFlowGraph, intrinsic: &Value, instruction_id: Id<Instruction>, arguments: &[ValueId], )

Convert the SSA vector operations to brillig vector operations

impl<Registers: RegisterAllocator> BrilligBlock<'_, Registers>

fn convert_ssa_binary( &mut self, binary: &Binary, dfg: &DataFlowGraph, result_variable: SingleAddrVariable, )

Converts the Binary instruction into a sequence of Brillig opcodes, writing the results to the result_variable.

fn convert_signed_modulo( &mut self, left: SingleAddrVariable, right: SingleAddrVariable, result: SingleAddrVariable, )

fn convert_signed_shr( &mut self, left: SingleAddrVariable, right: SingleAddrVariable, result: SingleAddrVariable, )

fn convert_signed_less_than( &mut self, left: SingleAddrVariable, right: SingleAddrVariable, result: SingleAddrVariable, )

fn add_overflow_check( &mut self, left: SingleAddrVariable, right: SingleAddrVariable, result: SingleAddrVariable, binary: &Binary, dfg: &DataFlowGraph, is_signed: bool, )

Overflow checks for the following unsigned binary operations

Checked Add/Sub/Mul

pub(crate) fn codegen_binary( &mut self, instruction_id: Id<Instruction>, binary: &Binary, dfg: &DataFlowGraph, )

Define a variable for the result and convert a binary operation to Brillig opcodes.

pub(crate) fn codegen_constrain( &mut self, lhs: ValueId, rhs: ValueId, assert_message: &Option<ConstrainError>, dfg: &DataFlowGraph, )

Generate Brillig opcodes to constrain two values to be equal to each other.

impl<Registers: RegisterAllocator> BrilligBlock<'_, Registers>

fn initialize_constant_array( &mut self, data: &Vector<ValueId>, typ: &Type, dfg: &DataFlowGraph, pointer: MemoryAddress, )

Initializes a constant array in Brillig memory.

This method is responsible for writing a constant array’s contents into memory, starting from the given pointer. It chooses between compile-time or runtime initialization depending on the data pattern and size.

If the array is large (>10 items), its elements are all numeric, and all items are identical, a runtime loop is generated to perform the initialization more efficiently.

Otherwise, the method falls back to a straightforward compile-time initialization, where each array element is emitted explicitly.

This optimization helps reduce Brillig bytecode size and runtime cost when initializing large, uniform arrays.

Example

For an array like [5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5], a runtime loop will be used For an array like [1, 2, 3, 4], each element will be set explicitly

fn initialize_constant_array_runtime( &mut self, item_types: Arc<Vec<Type>>, item_to_repeat: Vec<ValueId>, item_count: usize, pointer: MemoryAddress, dfg: &DataFlowGraph, )

Codegens Brillig instructions to initialize a large, constant array using a runtime loop.

This method assumes the array consists of identical items repeated multiple times. It generates a Brillig loop that writes the repeated item into memory efficiently, reducing bytecode size and instruction count compared to unrolling each element.

For complex types (e.g., tuples), multiple memory writes happen per loop iteration. For primitive type (e.g., u32, Field), a single memory write happens per loop iteration.

fn initialize_constant_array_comptime( &mut self, data: &Vector<ValueId>, dfg: &DataFlowGraph, pointer: MemoryAddress, )

Codegens Brillig instructions to initialize a constant array at compile time.

This method generates one store instruction per array element, writing each value from the SSA into consecutive memory addresses starting at pointer.

Unlike initialize_constant_array_runtime, this does not use loops and emits one instruction per write, which can increase bytecode size but provides fine-grained control.

fn convert_ssa_array_get( &mut self, array_variable: BrilligVariable, index_variable: SingleAddrVariable, destination_variable: BrilligVariable, has_offset: bool, )

Load from an array variable at a specific index into a specified destination.

If has_offset is set, then the index_variable is expected to have accounted for the array/vector specific offset between the array_variable and the start of the items; otherwise opcodes are emitted to calculate an adjusted base pointer.

Panics

The array variable is not a BrilligVariable::BrilligArray or BrilligVariable::BrilligVector when has_offset is false

fn convert_ssa_array_set( &mut self, source_variable: BrilligVariable, destination_variable: BrilligVariable, index_register: SingleAddrVariable, value_variable: BrilligVariable, mutable: bool, has_offset: bool, )

Array set operation in SSA returns a new array or vector that is a copy of the parameter array or vector with a specific value changed.

Whether an actual copy other the array occurs or we write into the same source array is determined by the call into the array copy procedure. If the reference count of an array pointer is one, we write directly to the array. Look at the procedure compilation for the exact procedure’s codegen.

If has_offset is set, then the index_variable is expected to have accounted for the array/vector specific offset between the source_variable and the start of the items; otherwise opcodes are emitted to calculate an adjusted base pointer.

fn codegen_assert_rc_neq_zero(&mut self, rc_register: MemoryAddress)

Debug utility method to determine whether an array’s reference count (RC) is zero. If RC’s have drifted down to zero it means the RC increment/decrement instructions have been written incorrectly.

Should only be called if BrilligContext::enable_debug_assertions returns true.

pub(crate) fn codegen_allocate( &mut self, instruction_id: Id<Instruction>, dfg: &DataFlowGraph, )

Define the result variable on the stack, then allocate 1 memory slot on the heap point the reference variable at it.

pub(crate) fn codegen_store( &mut self, address: ValueId, value: ValueId, dfg: &DataFlowGraph, )

Convert the address and value to Brillig variables, then generate opcodes to store value at address.

pub(crate) fn codegen_load( &mut self, instruction_id: Id<Instruction>, address: ValueId, dfg: &DataFlowGraph, )

Define the result variable, then generate opcodes to load the converted address into it.

pub(crate) fn codegen_array_get( &mut self, instruction_id: Id<Instruction>, array: ValueId, index: ValueId, dfg: &DataFlowGraph, )

Define a variable for the result, convert the array and indexes, then generate opcodes to load an array item.

pub(crate) fn codegen_array_set( &mut self, instruction_id: Id<Instruction>, array: ValueId, index: ValueId, value: ValueId, mutable: bool, dfg: &DataFlowGraph, )

Define a variable for the result, convert the array, index and value, then generate opcodes to store an array item.

pub(crate) fn codegen_make_array( &mut self, instruction_id: Id<Instruction>, array: &Vector<ValueId>, typ: &Type, dfg: &DataFlowGraph, )

Define the variable for the array or vector, allocate memory for the ref-counter and items, then load all items from array into the memory.

pub(crate) fn codegen_increment_rc( &mut self, value: ValueId, dfg: &DataFlowGraph, )

pub(crate) fn codegen_decrement_rc( &mut self, value: ValueId, dfg: &DataFlowGraph, )