noirc_evaluator::brillig::brillig_gen::brillig_fn
Struct FunctionContext

pub(crate) struct FunctionContext {
    function_id: Option<Id<Function>>,
    pub(crate) ssa_value_allocations: FxHashMap<ValueId, BrilligVariable>,
    blocks: Vec<BasicBlockId>,
    pub(crate) liveness: VariableLiveness,
    pub(crate) constant_allocation: ConstantAllocation,
    pub(crate) is_entry_point: bool,
}

Information required to compile an SSA Function into Brillig bytecode.

This structure is instantiated once per function and used throughout basic block code generation. It can also represent a non-function context (e.g., global instantiation) to reuse block codegen logic by leaving its function_id field unset.

Fields
function_id: Option<Id<Function>>

A FunctionContext is necessary for using a Brillig block’s code gen, but sometimes such as with globals, we are not within a function and do not have a FunctionId.

ssa_value_allocations: FxHashMap<ValueId, BrilligVariable>

Map from SSA values its allocation. Since values can be only defined once in SSA form, we insert them here on when we allocate them at their definition.

Multiple variables could be assigned the same slot, because this structure accumulates historical allocations, not just the currently active ones. This is needed so that when we start processing a block, we can always look up the allocation of the variables which are live at the beginning of it, even if they were deemed dead by another block we already visited.

Note that we don’t use Allocated<BrilligVariable> here, because we create a fresh allocator for each block we process, and something that is allocated in e.g. block 1 might be deallocated in block 2, so it has to be done manually.

blocks: Vec<BasicBlockId>

The block ids of the function in Reverse Post Order.

liveness: VariableLiveness

Liveness information for each variable in the function.

constant_allocation: ConstantAllocation

Information on where to allocate constants

is_entry_point: bool

True if this function is a brillig entry point

Implementations

impl FunctionContext

pub(crate) fn new(function: &Function, is_entry_point: bool) -> Self

Creates a new function context. It will allocate parameters for all blocks and compute the liveness of every variable.

pub(crate) fn function_id(&self) -> Id<Function>

Get the ID of the function this context was created for.

Panics if we call it when in the context created to hold data structures for global codegen only.

pub(crate) fn return_values(func: &Function) -> Vec<BrilligParameter>

Collects the return values of a given function

pub(crate) fn ssa_type_to_parameter(typ: &Type) -> BrilligParameter

Converts an SSA Type into a corresponding BrilligParameter.

This conversion defines the calling convention for Brillig functions, ensuring that SSA values are correctly mapped to memory layouts understood by the VM.

Panics

Panics if called with a slice type, as a slice’s memory layout cannot be inferred without runtime data.

pub(crate) fn post_order(&self) -> impl ExactSizeIterator<Item = BasicBlockId>

Iterate blocks in Post Order.

pub(crate) fn reverse_post_order( &self, ) -> impl ExactSizeIterator<Item = BasicBlockId>

Iterate blocks in Reverse Post Order.