noirc_evaluator::brillig::brillig_gen
Module variable_liveness

This module analyzes the liveness of variables (non-constant values) throughout a function. It uses the approach detailed in the section 4.2 of this paper https://inria.hal.science/inria-00558509v2/document

Structs

BackEdge 🔒

A back edge is an edge from a node to one of its ancestors. It denotes a loop in the CFG.

VariableLiveness 🔒

A struct representing the liveness of variables throughout a function.

Functions

is_variable 🔒

Check if the Value behind the ValueId requires register allocation (like a function parameter), rather than a global value like a user-defined function, intrinsic, or foreign function.

variables_used_in_block 🔒

Collect all ValueIds used in an BasicBlock which refer to Variables.

variables_used_in_instruction 🔒

Collect all ValueIds used in an Instruction which refer to variables (not functions).

Type Aliases

LastUses 🔒

The set variables which are dead after a given instruction (in a given block).

LoopMap 🔒

Maps a loop (identified by its header and back-edge) to the blocks in the loop body.

Variables 🔒

A set of ValueIds referring to SSA variables (not functions).