noirc_evaluator::brillig
Module brillig_ir

This module is an abstraction layer over Brillig. To allow for separation of concerns, it knows nothing about SSA types, and can therefore be tested independently. brillig_gen is therefore the module which combines both ssa types and types in this module. A similar paradigm can be seen with the acir_ir module.

The brillig IR provides instructions and codegens. The instructions are low level operations that are printed via debug_show. They should emit few opcodes. Codegens on the other hand orchestrate the low level instructions to emit the desired high level operation.

Modules

artifact

brillig_variable 🔒

codegen_binary 🔒

codegen_calls 🔒

codegen_control_flow 🔒

codegen_intrinsic 🔒

codegen_memory 🔒

codegen_stack 🔒

debug_show 🔒

This module contains functions for producing a higher level view disassembler of Brillig.

entry_point 🔒

instructions 🔒

procedures 🔒

registers 🔒

Memory layout and register allocation for Brillig programs.

Structs

BrilligContext 🔒

Brillig context object that is used while constructing the Brillig bytecode.

ReservedRegisters 🔒

Registers reserved in runtime for special purposes.

Constants

BRILLIG_MEMORY_ADDRESSING_BIT_SIZE 🔒

The Brillig VM does not apply a limit to the memory address space, As a convention, we take use 32 bits. This means that we assume that memory has 2^32 memory slots.

Functions

assert_u32 🔒

Converts a usize value to u32, panicking if the conversion fails.

assert_usize 🔒

Converts a u32 value to usize, panicking if the conversion fails.