noirc_evaluator::brillig::brillig_ir
Module artifact

Structs

BrilligArtifact

Artifacts resulting from the compilation of a function into brillig byte code. It includes the bytecode of the function and all the metadata that allows linking with other functions.

GeneratedBrillig 🔒

The result of compiling and linking brillig artifacts. This is ready to run bytecode with attached metadata.

Label 🔒

An identifier for a location in the code.

Enums

BrilligParameter 🔒

Represents a parameter or a return value of an entry point function.

LabelType 🔒

Type Aliases

JumpInstructionPosition 🔒

Pointer to a unresolved Jump instruction in the bytecode.

OpcodeLocation 🔒

A pointer to a location in the opcode.

UnresolvedJumpLocation 🔒

When constructing the bytecode, there may be instructions which require one to jump to a specific region of code (function)