noirc_evaluator::brillig::brillig_ir
Module procedures

Modules

array_copy 🔒

array_reverse 🔒

check_max_stack_depth 🔒

error_with_string 🔒

mem_copy 🔒

prepare_vector_insert 🔒

prepare_vector_push 🔒

vector_copy 🔒

vector_pop_back 🔒

vector_pop_front 🔒

vector_remove 🔒

Enums

ProcedureId

Procedures are a set of complex operations that are common in the noir language. Extracting them to reusable procedures allows us to reduce the size of the generated Brillig. Procedures receive their arguments on scratch space to avoid stack dumping&restoring.

Functions

compile_procedure 🔒

Compile a procedure as a stand-alone Brillig artifact, generating byte code for a specific operation, reading and returning arguments through the ScratchSpace.