noirc_evaluator::brillig::brillig_ir::procedures::prepare_vector_push
Function reallocate_vector_for_insertion

pub(crate) fn reallocate_vector_for_insertion<F: AcirField + DebugToString, Registers: RegisterAllocator>(
    brillig_context: &mut BrilligContext<F, Registers>,
    source_vector: BrilligVector,
    source_rc: SingleAddrVariable,
    source_capacity: SingleAddrVariable,
    target_vector: BrilligVector,
    target_size: SingleAddrVariable,
)

Reallocates the target vector for insertion, skipping reallocation if the source vector can be reused:

if the capacity accommodates the target size:

if the RC is 1, we can increase the size of the source vector as reuse it as the destination

if the RC is not 1, we allocate a new destination vector with the source capacity and target size

if the capacity is too small for the target size, we allocate the destination vector with a capacity that is double the target size.

Does not copy the items, only reallocates the vector.

Whether a copy is necessary can be gleaned by comparing the target address to the source address.