noirc_evaluator::brillig::brillig_ir::registers
Struct Allocated

pub(crate) struct Allocated<T, R: RegisterAllocator> {
    inner: Option<AllocatedInner<T, R>>,
}

Wrapper for a memory address which automatically deallocates itself when it goes out of scope.

Fields
inner: Option<AllocatedInner<T, R>>

The inner structure is optional, so that we can map it to a different value and let the original wrapper be dropped without any effect.

Implementations

impl<R: RegisterAllocator> Allocated<MemoryAddress, R>

Construct a new allocated address.

fn new_addr(addr: MemoryAddress, registers: Rc<RefCell<R>>) -> Self

Wrap single allocated memory address.

impl<A, R: RegisterAllocator> Allocated<A, R>

pub(crate) fn pure(value: A) -> Self

Wrap a value in Allocated without deallocating any address at the end.

pub(crate) fn new( value: A, addresses: SmallVec<[MemoryAddress; 2]>, registers: Rc<RefCell<R>>, ) -> Self

Create an Allocated value that deallocates its associated addresses at the end.

fn from_inner(inner: AllocatedInner<A, R>) -> Self

fn into_inner(self) -> AllocatedInner<A, R>

Take the inner, leaving self ready to be dropped without any effect.

fn inner(&self) -> &AllocatedInner<A, R>

fn inner_mut(&mut self) -> &mut AllocatedInner<A, R>

pub(crate) fn map<B>(self, f: impl FnOnce(A) -> B) -> Allocated<B, R>

Map the value to something else.

The deallocation of addresses is deferred to when the new value goes out of scope.

pub(crate) fn map2<B>( self, other: Self, f: impl FnOnce(A, A) -> B, ) -> Allocated<B, R>

Map two values into one.

The addresses of the two values are merged, with the implicit assumption that the underlying registers are the same.

fn merge_addresses( a: SmallVec<[MemoryAddress; 2]>, b: SmallVec<[MemoryAddress; 2]>, ) -> SmallVec<[MemoryAddress; 2]>

pub(crate) fn deallocate(self)

Manually deallocate a register when itโs no longer needed.

pub(crate) fn detach(self) -> A

Detach the allocated value from the registers, promising to take care of ownership in a different way.

This is used when we allocate an address in a register, then create a new register where we consider the value pre-allocated, and finally want to deallocate the address in the new register, not the old one, so that we can reuse the memory for something else.