noirc_evaluator::brillig::brillig_ir::registers
Trait RegisterAllocator

pub(crate) trait RegisterAllocator {
    // Required methods
    fn start(&self) -> usize;
    fn end(&self) -> usize;
    fn allocate_register(&mut self) -> MemoryAddress;
    fn deallocate_register(&mut self, register: MemoryAddress);
    fn ensure_register_is_allocated(&mut self, register: MemoryAddress);
    fn from_preallocated_registers(
        preallocated_registers: Vec<MemoryAddress>,
        layout: LayoutConfig,
    ) -> Self;
    fn empty_registers_start(&self) -> MemoryAddress;
    fn layout(&self) -> LayoutConfig;
}

Required Methods

fn start(&self) -> usize

First valid memory address

fn end(&self) -> usize

Last valid memory address

fn allocate_register(&mut self) -> MemoryAddress

Allocates a new register.

fn deallocate_register(&mut self, register: MemoryAddress)

Push a register to the deallocation list, ready for reuse.

fn ensure_register_is_allocated(&mut self, register: MemoryAddress)

Ensures a register is allocated, allocating it if necessary.

fn from_preallocated_registers( preallocated_registers: Vec<MemoryAddress>, layout: LayoutConfig, ) -> Self

Creates a new register context from a set of registers allocated previously.

fn empty_registers_start(&self) -> MemoryAddress

Finds the first register which is followed only by free registers.

Always returns a MemoryAddress::Relative address.

fn layout(&self) -> LayoutConfig

Return the memory layout used by this allocator.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.