noirc_evaluator::brillig::brillig_ir
Struct BrilligContext

pub(crate) struct BrilligContext<F, Registers> {
    obj: BrilligArtifact<F>,
    registers: Registers,
    context_label: Label,
    current_section: usize,
    next_section: usize,
    debug_show: DebugShow,
    can_call_procedures: bool,
    enable_debug_assertions: bool,
    count_arrays_copied: bool,
    globals_memory_size: Option<usize>,
}

Brillig context object that is used while constructing the Brillig bytecode.

Fields
obj: BrilligArtifact<F>registers: Registers

Tracks register allocations

context_label: Label

Context label, must be unique with respect to the function being linked.

current_section: usize

Section label, used to separate sections of code

next_section: usize

Stores the next available section

debug_show: DebugShow

IR printer

can_call_procedures: bool

Whether this context can call procedures or not. This is used to prevent a procedure from calling another procedure.

enable_debug_assertions: bool

Insert extra assertions that we expect to be true, at the cost of larger bytecode size.

count_arrays_copied: bool

Count the number of arrays that are copied, and output this to stdout

globals_memory_size: Option<usize>
Implementations

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_array_copy_procedure( &mut self, source_array: BrilligArray, destination_array: BrilligArray, )

Conditionally copies a source array to a destination array. If the reference count of the source array is 1, then we can directly copy the pointer of the source array to the destination array.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn emit_println_of_array_copy_counter(&mut self)

emit: println(f"Total arrays copied: {array_copy_counter}")

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_array_reverse_procedure( &mut self, pointer: MemoryAddress, size: MemoryAddress, )

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_check_max_stack_depth_procedure(&mut self)

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_mem_copy_procedure( &mut self, source_pointer: MemoryAddress, destination_pointer: MemoryAddress, num_elements_variable: MemoryAddress, )

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_prepare_vector_insert_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, index: SingleAddrVariable, write_pointer: MemoryAddress, item_count: usize, )

It prepares a vector for a insert operation, leaving a hole at the index position which is returned as the write_pointer.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_prepare_vector_push_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, write_pointer: MemoryAddress, item_push_count: usize, back: bool, )

Prepares a vector for a push operation, allocating a larger vector and copying the source vector into the destination vector. It returns the write pointer to where to put the new items.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_revert_with_string_procedure( &mut self, revert_string: String, )

Reverts with the error selector of the given string This procedure is useful to deduplicate generating code for the same selector.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_copy_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, )

Conditionally copies a source array to a destination array. If the reference count of the source array is 1, then we can directly copy the pointer of the source array to the destination array.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_pop_back_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, read_pointer: MemoryAddress, item_pop_count: usize, )

Pops items from the back of a vector, returning the new vector and the pointer to the popped items in read_pointer.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_pop_front_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, item_pop_count: usize, )

Pops items from the front of a vector, returning the new vector

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_remove_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, index: SingleAddrVariable, item_count: usize, )

Removes items from the vector, returning the new vector.

impl<F, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn allocate_register(&mut self) -> MemoryAddress

Allocates an unused register.

pub(crate) fn set_allocated_registers( &mut self, allocated_registers: Vec<MemoryAddress>, )

pub(crate) fn deallocate_register(&mut self, register_index: MemoryAddress)

Push a register to the deallocation list, ready for reuse.

pub(crate) fn deallocate_single_addr(&mut self, var: SingleAddrVariable)

Deallocates the address where the single address variable is stored

pub(crate) fn deallocate_heap_array(&mut self, arr: HeapArray)

pub(crate) fn deallocate_heap_vector(&mut self, vec: HeapVector)

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_usize_op_in_place( &mut self, destination: MemoryAddress, op: BrilligBinaryOp, constant: usize, )

Utility method to perform a binary instruction with a constant value in place

pub(crate) fn codegen_usize_op( &mut self, operand: MemoryAddress, destination: MemoryAddress, op: BrilligBinaryOp, constant: usize, )

Utility method to perform a binary instruction with a constant value

pub(crate) fn codegen_increment_array_copy_counter(&mut self)

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_call( &mut self, func_id: Id<Function>, arguments: &[BrilligVariable], returns: &[BrilligVariable], )

pub(crate) fn codegen_return(&mut self, return_registers: &[MemoryAddress])

Codegens a return from the current function.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_generic_iteration<T>( &mut self, make_iterator: impl FnOnce(&mut BrilligContext<F, Registers>) -> T, update_iterator: impl FnOnce(&mut BrilligContext<F, Registers>, &T), make_finish_condition: impl FnOnce(&mut BrilligContext<F, Registers>, &T) -> SingleAddrVariable, on_iteration: impl FnOnce(&mut BrilligContext<F, Registers>, &T), clean_iterator: impl FnOnce(&mut BrilligContext<F, Registers>, T), )

pub(crate) fn codegen_for_loop( &mut self, loop_start: Option<MemoryAddress>, loop_bound: MemoryAddress, step: Option<MemoryAddress>, on_iteration: impl FnOnce(&mut BrilligContext<F, Registers>, SingleAddrVariable), )

This codegen will issue a loop for (let iterator_register = loop_start; i < loop_bound; i += step) The body of the loop should be issued by the caller in the on_iteration closure.

pub(crate) fn codegen_loop( &mut self, iteration_count: MemoryAddress, on_iteration: impl FnOnce(&mut BrilligContext<F, Registers>, SingleAddrVariable), )

This codegen will issue a loop that will iterate from 0 to iteration_count The body of the loop should be issued by the caller in the on_iteration closure.

pub(crate) fn codegen_branch( &mut self, condition: MemoryAddress, f: impl FnMut(&mut BrilligContext<F, Registers>, bool), )

This codegen will issue an if-then branch that will check if the condition is true and if so, perform the instructions given in f(self, true) and otherwise perform the instructions given in f(self, false). A boolean is passed instead of two separate functions to allow the given function to mutably alias its environment.

pub(crate) fn codegen_if( &mut self, condition: MemoryAddress, f: impl FnOnce(&mut BrilligContext<F, Registers>), )

This codegen issues a branch that jumps over the code generated by the given function if the condition is false

pub(crate) fn codegen_if_not( &mut self, condition: MemoryAddress, f: impl FnOnce(&mut BrilligContext<F, Registers>), )

This codegen issues a branch that jumps over the code generated by the given function if the condition is truthy

pub(crate) fn codegen_constrain_with_revert_data( &mut self, condition: SingleAddrVariable, revert_data_items: Vec<BrilligVariable>, revert_data_types: Vec<BrilligParameter>, error_selector: ErrorSelector, )

Emits brillig bytecode to jump to a trap condition if condition is false. The trap will include the given message as revert data.

pub(crate) fn codegen_constrain( &mut self, condition: SingleAddrVariable, assert_message: Option<String>, )

Emits brillig bytecode to jump to a trap condition if condition is false.

pub(super) fn revert_with_string(&mut self, revert_string: String)

pub(super) fn flattened_size(param: &BrilligParameter) -> usize

Computes the size of a parameter if it was flattened

pub(super) fn flattened_tuple_size(tuple: &[BrilligParameter]) -> usize

Computes the size of a parameter if it was flattened

pub(crate) fn has_nested_arrays(tuple: &[BrilligParameter]) -> bool

Computes the size of a parameter if it was flattened

pub(super) fn flatten_array( &mut self, item_type: &[BrilligParameter], item_count: usize, flattened_array_pointer: MemoryAddress, deflattened_items_pointer: MemoryAddress, )

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_truncate( &mut self, destination_of_truncated_value: SingleAddrVariable, value_to_truncate: SingleAddrVariable, bit_size: u32, )

Codegens a truncation of a value to the given bit size

pub(crate) fn codegen_to_radix( &mut self, source_field: SingleAddrVariable, target_array: BrilligArray, radix: SingleAddrVariable, little_endian: bool, output_bits: bool, )

Issues a to_radix instruction. This instruction will write the modulus of the source register And the radix register limb_count times to the target vector.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_allocate_immediate_mem( &mut self, pointer_register: MemoryAddress, size: usize, )

Allocates an array of size size and stores the pointer to the array in pointer_register

pub(crate) fn codegen_allocate_mem( &mut self, pointer_register: MemoryAddress, size_register: MemoryAddress, )

Allocates an array of size contained in size_register and stores the pointer to the array in pointer_register

pub(crate) fn codegen_load_with_offset( &mut self, base_ptr: MemoryAddress, index: SingleAddrVariable, result: MemoryAddress, )

Gets the value stored at base_ptr + index and stores it in result

pub(crate) fn codegen_store_with_offset( &mut self, base_ptr: MemoryAddress, index: SingleAddrVariable, value: MemoryAddress, )

Stores value at base_ptr + index

pub(crate) fn codegen_mem_copy( &mut self, source_pointer: MemoryAddress, destination_pointer: MemoryAddress, num_elements_variable: SingleAddrVariable, )

Copies the values of memory pointed by source with length stored in num_elements_register After the address pointed by destination

pub(crate) fn codegen_mem_copy_from_the_end( &mut self, source_start: MemoryAddress, target_start: MemoryAddress, num_elements_variable: SingleAddrVariable, )

Copies num_elements_variable from the source pointer to the target pointer, starting from the end

pub(crate) fn codegen_array_reverse( &mut self, items_pointer: MemoryAddress, size: MemoryAddress, )

This instruction will reverse the order of the size elements pointed by pointer.

pub(crate) fn codegen_brillig_array_to_heap_array( &mut self, array: BrilligArray, ) -> HeapArray

Converts a BrilligArray (pointer to [RC, ...items]) to a HeapArray (pointer to [items])

pub(crate) fn codegen_brillig_vector_to_heap_vector( &mut self, vector: BrilligVector, ) -> HeapVector

pub(crate) fn variable_to_value_or_array( &mut self, variable: BrilligVariable, ) -> ValueOrArray

pub(crate) fn codegen_make_vector_length( &mut self, vector: BrilligVector, ) -> SingleAddrVariable

Returns a variable holding the length of a given vector

pub(crate) fn codegen_update_vector_length( &mut self, vector: BrilligVector, new_length: SingleAddrVariable, )

pub(crate) fn codegen_make_vector_capacity( &mut self, vector: BrilligVector, ) -> SingleAddrVariable

Returns a variable holding the capacity of a given vector

pub(crate) fn codegen_vector_items_pointer( &mut self, vector: BrilligVector, result: MemoryAddress, )

Writes a pointer to the items of a given vector

pub(crate) fn codegen_make_vector_items_pointer( &mut self, vector: BrilligVector, ) -> MemoryAddress

Returns a pointer to the items of a given vector

pub(crate) fn codegen_read_vector_metadata( &mut self, vector: BrilligVector, rc: SingleAddrVariable, size: SingleAddrVariable, capacity: SingleAddrVariable, items_pointer: SingleAddrVariable, )

Reads the metadata of a vector and stores it in the given variables

pub(crate) fn codegen_make_array_items_pointer( &mut self, array: BrilligArray, ) -> MemoryAddress

Returns a pointer to the items of a given array

pub(crate) fn codegen_make_array_or_vector_items_pointer( &mut self, variable: BrilligVariable, ) -> MemoryAddress

pub(crate) fn codegen_initialize_array(&mut self, array: BrilligArray)

Initializes an array, allocating memory to store its representation and initializing the reference counter.

fn initialize_rc(&mut self, pointer: MemoryAddress, rc_value: usize)

Initialize the reference counter for an array or vector. This should only be used internally in the array and vector initialization methods

pub(crate) fn codegen_initialize_vector( &mut self, vector: BrilligVector, size: SingleAddrVariable, capacity: Option<SingleAddrVariable>, )

Initializes a vector, allocating memory to store its representation and initializing the reference counter, size and capacity

pub(super) fn codegen_initialize_vector_metadata( &mut self, vector: BrilligVector, size: SingleAddrVariable, capacity: Option<SingleAddrVariable>, )

Writes vector metadata (reference count, size, and capacity) into the allocated memory

pub(crate) fn initialize_externally_returned_vector( &mut self, vector: BrilligVector, resulting_heap_vector: HeapVector, )

We don’t know the length of a vector returned externally before the call so we pass the free memory pointer and then use this function to allocate after the fact when we know the length.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_mov_registers_to_registers( &mut self, sources: Vec<MemoryAddress>, destinations: Vec<MemoryAddress>, )

This function moves values from a set of registers to another set of registers. The only requirement is that every destination needs to be written at most once.

fn perform_movements( &mut self, movements: &HashMap<MemoryAddress, HashSet<MemoryAddress>>, current_source: MemoryAddress, )

impl<F: AcirField + DebugToString> BrilligContext<F, Stack>

pub(crate) fn new_entry_point_artifact( arguments: Vec<BrilligParameter>, return_parameters: Vec<BrilligParameter>, target_function: Id<Function>, globals_init: bool, globals_memory_size: usize, options: &BrilligOptions, ) -> BrilligArtifact<F>

Creates an entry point artifact that will jump to the function label provided.

fn calldata_start_offset(&self) -> usize

fn return_data_start_offset(&self, calldata_size: usize) -> usize

fn codegen_entry_point( &mut self, arguments: &[BrilligParameter], return_parameters: &[BrilligParameter], )

Adds the instructions needed to handle entry point parameters The runtime will leave the parameters in calldata. Arrays will be passed flattened.

fn allocate_function_arguments( &mut self, arguments: &[BrilligParameter], ) -> Vec<BrilligVariable>

fn copy_and_cast_calldata(&mut self, arguments: &[BrilligParameter])

fn deflatten_array( &mut self, item_type: &[BrilligParameter], item_count: usize, flattened_array_pointer: MemoryAddress, is_vector: bool, ) -> MemoryAddress

Deflatten an array by recursively allocating nested arrays and copying the plain values. Returns the pointer to the deflattened items.

fn codegen_exit_point( &mut self, arguments: &[BrilligParameter], return_parameters: &[BrilligParameter], )

Adds the instructions needed to handle return parameters The runtime expects the results in a contiguous memory region. Arrays are expected to be returned with all the nested arrays flattened. However, the function called returns variables (that have extra data) and the returned arrays are deflattened.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

Low level instructions of the brillig IR, used by the brillig ir codegens and brillig_gen Printed using debug_slow

pub(crate) fn binary_instruction( &mut self, lhs: SingleAddrVariable, rhs: SingleAddrVariable, result: SingleAddrVariable, operation: BrilligBinaryOp, )

Processes a binary instruction according operation.

This method will compute lhs <operation> rhs and store the result in the result register.

pub(crate) fn not_instruction( &mut self, input: SingleAddrVariable, result: SingleAddrVariable, )

Processes a not instruction.

Not is computed using a subtraction operation as there is no native not instruction in Brillig.

pub(crate) fn memory_op_instruction( &mut self, lhs: MemoryAddress, rhs: MemoryAddress, destination: MemoryAddress, op: BrilligBinaryOp, )

Utility method to perform a binary instruction with a memory address

pub(crate) fn conditional_move_instruction( &mut self, condition: SingleAddrVariable, then_address: SingleAddrVariable, else_address: SingleAddrVariable, destination: SingleAddrVariable, )

Insert a conditional move instruction

fn binary( &mut self, lhs: SingleAddrVariable, rhs: SingleAddrVariable, result: SingleAddrVariable, operation: BrilligBinaryOp, )

fn modulo( &mut self, result: SingleAddrVariable, left: SingleAddrVariable, right: SingleAddrVariable, )

Computes left % right by emitting the necessary Brillig opcodes.

This is done by using the following formula:

a % b = a - (b * (a / b))

Brillig does not have an explicit modulo operation, so we must emit multiple opcodes and process it differently to other binary instructions.

fn binary_result_bit_size( operation: BrilligBinaryOp, arguments_bit_size: u32, ) -> u32

pub(crate) fn foreign_call_instruction( &mut self, func_name: String, inputs: &[ValueOrArray], input_value_types: &[HeapValueType], outputs: &[ValueOrArray], output_value_types: &[HeapValueType], )

Processes a foreign call instruction.

Note: the function being called is external and will not be linked during brillig generation.

pub(crate) fn add_external_call_instruction(&mut self, func_id: Id<Function>)

Adds a unresolved external Call instruction to the bytecode. This calls into another function compiled into this brillig artifact.

pub(super) fn add_procedure_call_instruction( &mut self, procedure_id: ProcedureId, )

pub(super) fn add_globals_init_instruction(&mut self, func_id: Id<Function>)

pub(crate) fn jump_instruction(&mut self, target_label: Label)

Adds a unresolved Jump instruction to the bytecode.

pub(crate) fn jump_if_instruction( &mut self, condition: MemoryAddress, target_label: Label, )

Adds a unresolved JumpIf instruction to the bytecode.

fn add_unresolved_jump( &mut self, jmp_instruction: BrilligOpcode<F>, destination: Label, )

Adds a unresolved Jump to the bytecode.

pub(crate) fn enter_context(&mut self, label: Label)

Adds a label to the next opcode

pub(super) fn enter_section(&mut self, section: usize)

Enter the given section within a basic block

pub(super) fn reserve_next_section_label(&mut self) -> (usize, Label)

Create, reserve, and return a new section label.

fn compute_section_label(&self, section: usize) -> Label

Internal function used to compute the section labels

fn current_section_label(&self) -> Label

Returns the current section label

pub(crate) fn return_instruction(&mut self)

Emits a return instruction

pub(crate) fn stop_instruction(&mut self, return_data: HeapVector)

Emits a stop instruction with return data

pub(crate) fn black_box_op_instruction(&mut self, op: BlackBoxOp)

Issues a blackbox operation.

pub(crate) fn load_free_memory_pointer_instruction( &mut self, pointer_register: MemoryAddress, )

pub(crate) fn increase_free_memory_pointer_instruction( &mut self, size_register: MemoryAddress, )

pub(crate) fn store_instruction( &mut self, destination_pointer: MemoryAddress, source: MemoryAddress, )

Emits a store instruction

pub(crate) fn load_instruction( &mut self, destination: MemoryAddress, source_pointer: MemoryAddress, )

Emits a load instruction

pub(crate) fn mov_instruction( &mut self, destination: MemoryAddress, source: MemoryAddress, )

Emits a mov instruction.

Copies the value at source into destination

pub(crate) fn cast_instruction( &mut self, destination: SingleAddrVariable, source: SingleAddrVariable, )

Cast truncates the value to the given bit size and converts the type of the value in memory to that bit size.

pub(crate) fn cast( &mut self, destination: SingleAddrVariable, source: SingleAddrVariable, )

pub(crate) fn const_instruction( &mut self, result: SingleAddrVariable, constant: F, )

Stores the value of constant in the result register

pub(crate) fn indirect_const_instruction( &mut self, result_pointer: MemoryAddress, bit_size: u32, constant: F, )

Stores the value of constant in the result_pointer

fn constant( &mut self, result: MemoryAddress, bit_size: u32, constant: F, indirect: bool, )

pub(crate) fn usize_const_instruction( &mut self, result: MemoryAddress, constant: F, )

pub(crate) fn make_constant_instruction( &mut self, constant: F, bit_size: u32, ) -> SingleAddrVariable

Returns a register which holds the value of a constant

pub(crate) fn make_usize_constant_instruction( &mut self, constant: F, ) -> SingleAddrVariable

Returns a register which holds the value of an usize constant

pub(super) fn calldata_copy_instruction( &mut self, destination: MemoryAddress, calldata_size: usize, offset: usize, )

pub(super) fn trap_instruction(&mut self, revert_data: HeapVector)

impl<F, R> BrilligContext<F, R>

pub(crate) fn enable_debug_assertions(&self) -> bool

Enable the insertion of bytecode with extra assertions during testing.

pub(crate) fn array_copy_counter_address(&self) -> MemoryAddress

Returns the address of the implicit debug variable containing the count of implicitly copied arrays as a result of RC’s copy on write semantics.

pub(crate) fn count_array_copies(&self) -> bool

pub(crate) fn set_globals_memory_size(&mut self, new_size: Option<usize>)

Set the globals memory size if it is not already set. If it is already set, this will assert that the two values must be equal.

impl<F: AcirField + DebugToString> BrilligContext<F, Stack>

Regular brillig context to codegen user defined functions

pub(crate) fn new(options: &BrilligOptions) -> BrilligContext<F, Stack>

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn absolute_value( &mut self, num: SingleAddrVariable, absolute_value: SingleAddrVariable, result_is_negative: SingleAddrVariable, )

Splits a two’s complement signed integer in the sign bit and the absolute value. For example, -6 i8 (11111010) is split to 00000110 (6, absolute value) and 1 (is_negative).

pub(crate) fn convert_signed_division( &mut self, left: SingleAddrVariable, right: SingleAddrVariable, result: SingleAddrVariable, )

impl<F: AcirField + DebugToString> BrilligContext<F, ScratchSpace>

Special brillig context to codegen compiler intrinsic shared procedures

pub(crate) fn new_for_procedure( procedure_id: ProcedureId, options: &BrilligOptions, ) -> BrilligContext<F, ScratchSpace>

impl<F: AcirField + DebugToString> BrilligContext<F, GlobalSpace>

Special brillig context to codegen global values initialization

pub(crate) fn new_for_global_init( options: &BrilligOptions, entry_point: Id<Function>, ) -> BrilligContext<F, GlobalSpace>

pub(crate) fn global_space_size(&self) -> usize

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

fn push_opcode(&mut self, opcode: BrilligOpcode<F>)

Adds a brillig instruction to the brillig byte code

pub(crate) fn artifact(self) -> BrilligArtifact<F>

Returns the artifact

pub(crate) fn set_call_stack(&mut self, call_stack: CallStackId)

Sets a current call stack that the next pushed opcodes will be associated with.