noirc_evaluator
Module ssa

SSA stands for Single Static Assignment The IR presented in this module will already be in SSA form and will be used to apply conventional optimizations like Common Subexpression elimination and constant folding.

This module heavily borrows from Cranelift

Modules

artifact 🔒

builder 🔒

checks 🔒

function_builder

interpreter

ir

opt

This folder contains each optimization pass for the SSA IR.

parser

ssa_gen

validation 🔒

Validator that checks whether a function is well formed.

visit_once_deque 🔒

visit_once_priority_queue 🔒

Structs

ArtifactsAndWarnings

SsaBuilder

SsaCircuitArtifact

SsaEvaluatorOptions

SsaPass

An SSA pass reified as a construct we can put into a list, which facilitates equivalence testing between different stages of the processing pipeline.

SsaProgramArtifact

Enums

SsaLogging

Functions

combine_artifacts

convert_generated_acir_into_circuit

create_program

Compiles the [Program] into [ACIR][acvm::acir::circuit::Program].

create_program_with_minimal_passes

Compiles the [Program] into [ACIR][acvm::acir::circuit::Program] using the minimum amount of SSA passes.

create_program_with_passes

Compiles the [Program] into [ACIR][acvm::acir::circuit::Program] by running it through SSA passes.

minimal_passes

For testing purposes we want a list of the minimum number of SSA passes that should return the same result as the full pipeline.

optimize_into_acir

Optimize the given program by converting it into SSA form and performing optimizations there. When finished, convert the final SSA into an ACIR program and return it. An ACIR program is made up of both ACIR functions and Brillig functions for unconstrained execution.

optimize_ssa_builder_into_acir

Optimize the given SsaBuilder by converting it into SSA form and performing optimizations there. When finished, convert the final SSA into an ACIR program and return it. An ACIR program is made up of both ACIR functions and Brillig functions for unconstrained execution.

primary_passes

The default SSA optimization pipeline.

resolve_function_signature 🔒

split_public_and_private_inputs 🔒