noirc_evaluator::ssa::checks

Module check_for_missing_brillig_constraints

This module defines the Ssa::check_for_missing_brillig_constraints method.

It verifies that the output of Brillig calls is connected to the inputs of the calls by assertions; in other words, that the circuit has constraints that the output is correct, given the inputs.

To do so, it tracks the ancestry of every expression, and checks that any variable which is an output of a Brillig call has a descendant which appears in an assertion, where the other side has an ancestor that is an input of the call.

Essentially, to consider a particular Brillig call constrained, we are looking for a constraint where the ancestors of the constraint arguments intersect both of:

the descendants of the results of the call (outputs)

the ancestors of the arguments of the call (inputs)

For example take the following graph of variables feeding into calls:

  v1     v2      v3
   \   /  \    /
    \ /    \  /
     v4     v5 = call(v2, v3)
     |\     |
     | \    |
     |  \   |
     |   \  |
     |    \ |
     |      v6 = call(v5, v4)
     |     /
     |    /
     |   /
     |  /
constrain(v4, v6)

Both calls are considered constrained:

The output of the 2nd call (v6) is constrained directly against its input (v4)

The output of the 1st call (v5) has a descendant (v6) which is constrained against a value (v4) that has an ancestor (v2) which is also an ancestor of an argument of of the call itself.

The goal isn’t to verify that the constraint is correct, just that some (indirect) connection between inputs and outputs is made.

Structs

Context 🔒

TaintedDescendants 🔒

Outputs of a Brillig call and their descendants.

ValueSet 🔒

A more compact representation of a HashSet<ValueId> to limit memory use.

Constants

DEFAULT_MAX_ANCESTOR_DISTANCE

Limit how far back the BFS traverses to try to find a relation between the ancestors of constrained values and Brillig inputs/outputs.

DEFAULT_MAX_ARRAY_OUTPUT_LENGTH

The maximum length of arrays that we attempt to constrain item-by-item.

Functions

any_ancestor 🔒

Returns true if start itself, or any value reachable from start by following parents (and equivalences from intermediate nodes), satisfies a predicate.

as_equivalence 🔒

Whether the instruction is a constrain v1 == v2 with non-constant variables.

bfs_ancestors 🔒

Compute the set of all values reachable (inclusive) from any of the starts by following parents and equivalences edges backwards.

bfs_traverse_ancestors 🔒

Compute the set of all values reachable (inclusive) from any of the starts by following parents and equivalences edges backwards.

has_call_to_brillig 🔒

Whether there is at least one instruction making a call to a Brillig function with non-empty results.

instruction_arguments 🔒

Collect non-constant arguments of an instruction.

instruction_results 🔒

Collect non-constant results of an instruction.

is_call_to_brillig 🔒

Whether the instruction is a call to a Brillig function with a non-empty results.

is_constraint 🔒

Whether an instruction puts constraints on its inputs.

is_side_effect 🔒

Whether the instruction enables side effects with a non-constant variable.

parent_arguments 🔒

Like instruction_arguments, but tailored for the parent-graph edges that drive the ancestry-based check. For an Instruction::ArrayGet with a non-constant index the result is some single element of array, not a function of the whole array; treating array as a parent would propagate every element’s ancestry to the result, letting a constraint on the result trivially “explain” outputs that are actually free for the prover when index happens to point at a different slot.