noirc_evaluator::ssa::checks::check_for_underconstrained_values
Struct Context

struct Context {
    block_queue: VisitOnceDeque,
    value_sets: Vec<BTreeSet<ValueId>>,
    brillig_return_to_argument: HashMap<ValueId, Vec<ValueId>>,
    brillig_return_to_instruction_id: HashMap<ValueId, Id<Instruction>>,
}

Fields
block_queue: VisitOnceDequevalue_sets: Vec<BTreeSet<ValueId>>brillig_return_to_argument: HashMap<ValueId, Vec<ValueId>>brillig_return_to_instruction_id: HashMap<ValueId, Id<Instruction>>
Implementations

impl Context

fn compute_sets_of_connected_value_ids( &mut self, function: &Function, all_functions: &BTreeMap<Id<Function>, Function>, )

Compute sets of variable ValueIds that are connected with constraints

Additionally, store information about Brillig calls in the context

fn find_sets_connected_to_function_inputs_or_outputs( &mut self, function: &Function, ) -> BTreeSet<usize>

Find sets that contain input or output value of the function

Goes through each set of connected ValueIds and see if function arguments or return values are in the set

fn find_disconnecting_brillig_calls_with_results_in_set( &self, current_set: &BTreeSet<ValueId>, all_brillig_generated_values: &BTreeSet<ValueId>, function: &Function, ) -> Vec<SsaReport>

Find which Brillig calls separate this set from others and return bug warnings about them

fn connect_value_ids_in_block( &mut self, function: &Function, block: BasicBlockId, all_functions: &BTreeMap<Id<Function>, Function>, )

Go through each instruction in the block and add a set of ValueIds connected through that instruction

Additionally, this function adds mappings of Brillig return values to call arguments and instruction ids from calls to Brillig functions in the block

fn merge_sets(current: &[BTreeSet<ValueId>]) -> Vec<BTreeSet<ValueId>>

Merge all small sets into larger ones based on whether the sets intersect or not

If two small sets have a common ValueId, we merge them into one

fn merge_sets_par(sets: &[BTreeSet<ValueId>]) -> Vec<BTreeSet<ValueId>>

Parallel version of merge_sets The sets are merged by chunks, and then the chunks are merged together