noirc_evaluator::ssa::function_builder
Struct FunctionBuilder

pub struct FunctionBuilder {
    pub current_function: Function,
    current_block: BasicBlockId,
    finished_functions: Vec<Function>,
    call_stack: CallStackId,
    error_types: BTreeMap<ErrorSelector, Type>,
    pub simplify: bool,
    globals: Arc<GlobalsGraph>,
    purities: Arc<FxHashMap<Id<Function>, Purity>>,
}

The per-function context for each ssa function being generated.

This is split from the global SsaBuilder context to allow each function to be potentially built concurrently.

Contrary to the name, this struct has the capacity to build as many functions as needed, although it is limited to one function at a time.

Fields
current_function: Functioncurrent_block: BasicBlockIdfinished_functions: Vec<Function>call_stack: CallStackIderror_types: BTreeMap<ErrorSelector, Type>simplify: bool

Whether instructions are simplified as soon as they are inserted into this builder. This is true by default unless changed to false after constructing a builder.

globals: Arc<GlobalsGraph>purities: Arc<FxHashMap<Id<Function>, Purity>>
Implementations

impl FunctionBuilder

fn add_to_data_bus(&mut self, value: ValueId, databus: &mut DataBusBuilder)

Insert a value into a data bus builder

pub(crate) fn initialize_data_bus( &mut self, values: &[ValueId], databus: DataBusBuilder, call_data_id: Option<u32>, ) -> DataBusBuilder

Create a data bus builder from a list of values

pub(crate) fn call_data_bus( &mut self, flattened_databus_visibilities: Vec<DatabusVisibility>, ) -> Vec<DataBusBuilder>

Generate the data bus for call-data, based on the parameters of the entry block and a vector telling which ones are call-data

pub(crate) fn set_data_bus(&mut self, data_bus: DataBus)

Forcefully sets the databus of the current function.

fn deflatten_databus_visibilities( &self, ssa_params: &[ValueId], flattened_params_databus_visibility: Vec<DatabusVisibility>, ) -> Vec<DatabusVisibility>

This function takes the flattened databus visibilities and generates the databus visibility for each ssa parameter asserting that an ssa parameter is not assigned two different databus visibilities

impl FunctionBuilder

pub fn new(function_name: String, function_id: Id<Function>) -> Self

Creates a new FunctionBuilder to build the function with the given FunctionId.

This creates the new function internally so there is no need to call .new_function() right after constructing a new FunctionBuilder.

pub fn from_existing(function: &Function, function_id: Id<Function>) -> Self

Create a function builder with a new function created with the same name, globals, and function purities taken from an existing function.

pub fn set_runtime(&mut self, runtime: RuntimeType)

Set the runtime of the initial function that is created internally after constructing the FunctionBuilder. A function’s default runtime type is RuntimeType::Acir(InlineType::Inline). This should only be used immediately following construction of a FunctionBuilder and will panic if there are any already finished functions.

pub fn set_globals(&mut self, globals: Arc<GlobalsGraph>)

fn apply_globals(&mut self)

pub fn set_purities(&mut self, purities: Arc<FxHashMap<Id<Function>, Purity>>)

fn new_function_with_type( &mut self, name: String, function_id: Id<Function>, runtime_type: RuntimeType, )

Finish the current function and create a new function.

A FunctionBuilder can always only work on one function at a time, so care should be taken not to finish a function that is still in progress by calling new_function before the current function is finished.

pub fn new_function( &mut self, name: String, function_id: Id<Function>, inline_type: InlineType, )

Finish the current function and create a new ACIR function.

pub fn new_brillig_function( &mut self, name: String, function_id: Id<Function>, inline_type: InlineType, )

Finish the current function and create a new unconstrained function.

pub fn finish(self) -> Ssa

Consume the FunctionBuilder returning all the functions it has generated.

pub fn add_parameter(&mut self, typ: Type) -> ValueId

Add a parameter to the current function with the given parameter type. Returns the newly-added parameter.

pub fn numeric_constant( &mut self, value: impl Into<FieldElement>, typ: NumericType, ) -> ValueId

Insert a numeric constant into the current function

pub fn field_constant(&mut self, value: impl Into<FieldElement>) -> ValueId

Insert a numeric constant into the current function of type Field

pub fn length_constant(&mut self, value: impl Into<FieldElement>) -> ValueId

Insert a numeric constant into the current function of type Type::length_type()

pub fn type_of_value(&self, value: ValueId) -> Type

Returns the type of the given value.

pub fn insert_block(&mut self) -> BasicBlockId

Insert a new block into the current function and return it. Note that this block is unreachable until another block is set to jump to it.

pub fn add_block_parameter(&mut self, block: BasicBlockId, typ: Type) -> ValueId

Adds a parameter with the given type to the given block. Returns the newly-added parameter.

pub fn block_parameters(&self, block: BasicBlockId) -> &[ValueId]

Returns the parameters of the given block in the current function.

pub fn insert_instruction( &mut self, instruction: Instruction, ctrl_typevars: Option<Vec<Type>>, ) -> InsertInstructionResult<'_>

Inserts a new instruction at the end of the current block and returns its results

pub fn switch_to_block(&mut self, block: BasicBlockId)

Switch to inserting instructions in the given block. Expects the given block to be within the same function. If you want to insert instructions into a new function, call new_function instead.

pub(crate) fn current_block(&mut self) -> BasicBlockId

Returns the block currently being inserted into

pub fn get_current_block_index(&self) -> BasicBlockId

pub fn insert_allocate(&mut self, element_type: Type) -> ValueId

Insert an allocate instruction at the end of the current block, allocating the given amount of field elements. Returns the result of the allocate instruction, which is always a Reference to the allocated data.

pub fn set_location(&mut self, location: Location) -> &mut FunctionBuilder

pub fn set_call_stack( &mut self, call_stack: CallStackId, ) -> &mut FunctionBuilder

pub fn get_call_stack(&self) -> CallStack

pub fn insert_load(&mut self, address: ValueId, type_to_load: Type) -> ValueId

Insert a Load instruction at the end of the current block, loading from the given address which should point to a previous Allocate instruction. Note that this is limited to loading a single value. Loading multiple values (such as a tuple) will require multiple loads. Returns the element that was loaded.

pub fn insert_store(&mut self, address: ValueId, value: ValueId)

Insert a Store instruction at the end of the current block, storing the given element at the given address. Expects that the address points somewhere within a previous Allocate instruction.

pub fn insert_binary( &mut self, lhs: ValueId, operator: BinaryOp, rhs: ValueId, ) -> ValueId

Insert a binary instruction at the end of the current block. Returns the result of the binary instruction.

pub fn insert_not(&mut self, rhs: ValueId) -> ValueId

Insert a not instruction at the end of the current block. Returns the result of the instruction.

pub fn insert_cast(&mut self, value: ValueId, typ: NumericType) -> ValueId

Insert a cast instruction at the end of the current block. Returns the result of the cast instruction.

pub fn insert_truncate( &mut self, value: ValueId, bit_size: u32, max_bit_size: u32, ) -> ValueId

Insert a truncate instruction at the end of the current block. Returns the result of the truncate instruction.

pub fn insert_constrain( &mut self, lhs: ValueId, rhs: ValueId, assert_message: Option<ConstrainError>, )

Insert a constrain instruction at the end of the current block.

pub fn insert_range_check( &mut self, value: ValueId, max_bit_size: u32, assert_message: Option<String>, )

Insert a Instruction::RangeCheck instruction at the end of the current block.

pub fn insert_call( &mut self, func: ValueId, arguments: Vec<ValueId>, result_types: Vec<Type>, ) -> Cow<'_, [ValueId]>

Insert a call instruction at the end of the current block and return the results of the call.

pub fn insert_array_get( &mut self, array: ValueId, index: ValueId, element_type: Type, ) -> ValueId

Insert an instruction to extract an element from an array

pub fn insert_array_set( &mut self, array: ValueId, index: ValueId, value: ValueId, mutable: bool, ) -> ValueId

Insert an instruction to create a new array with the given index replaced with a new value

pub fn insert_inc_rc(&mut self, value: ValueId)

Insert an instruction to increment an array’s reference count. This only has an effect in unconstrained code where arrays are reference counted and copy on write.

pub(crate) fn insert_dec_rc(&mut self, value: ValueId)

Insert an instruction to decrement an array’s reference count. This only has an effect in unconstrained code where arrays are reference counted and copy on write.

pub fn insert_enable_side_effects_if(&mut self, condition: ValueId)

Insert an enable_side_effects_if instruction. These are normally only automatically inserted during the flattening pass when branching is removed.

pub fn insert_make_array( &mut self, elements: Vector<ValueId>, typ: Type, ) -> ValueId

Insert a make_array instruction to create a new array or slice. Returns the new array value. Expects typ to be an array or slice type.

fn terminate_block_with(&mut self, terminator: TerminatorInstruction)

Terminates the current block with the given terminator instruction if the current block does not already have a terminator instruction.

pub fn terminate_with_jmp( &mut self, destination: BasicBlockId, arguments: Vec<ValueId>, )

Terminate the current block with a jmp instruction to jmp to the given block with the given arguments.

pub fn terminate_with_jmpif( &mut self, condition: ValueId, then_destination: BasicBlockId, else_destination: BasicBlockId, )

Terminate the current block with a jmpif instruction to jmp with the given arguments block with the given arguments.

pub fn terminate_with_return(&mut self, return_values: Vec<ValueId>)

Terminate the current block with a return instruction

pub fn terminate_with_unreachable(&mut self)

Terminate the current block with an unreachable instruction

pub fn import_function(&mut self, function: Id<Function>) -> ValueId

Returns a ValueId pointing to the given function or imports the function into the current function if it was not already, and returns that ID.

pub fn import_foreign_function(&mut self, function: &str) -> ValueId

Returns a ValueId pointing to the given oracle/foreign function or imports the oracle into the current function if it was not already, and returns that ID.

pub fn import_intrinsic(&mut self, name: &str) -> Option<ValueId>

Retrieve a value reference to the given intrinsic operation. Returns None if there is no intrinsic matching the given name.

pub fn import_intrinsic_id(&mut self, intrinsic: Intrinsic) -> ValueId

Retrieve a value reference to the given intrinsic operation.

pub fn get_intrinsic_from_value(&mut self, value: ValueId) -> Option<Intrinsic>

pub(crate) fn increment_array_reference_count( &mut self, value: ValueId, ) -> Option<ValueId>

Insert instructions to increment the reference count of any array(s) stored within the given value. If the given value is not an array and does not contain any arrays, this does nothing.

Returns the ID of the array that was affected, if any.

pub(crate) fn decrement_array_reference_count( &mut self, value: ValueId, ) -> Option<ValueId>

Insert instructions to decrement the reference count of any array(s) stored within the given value. If the given value is not an array and does not contain any arrays, this does nothing.

Returns the ID of the array that was affected, if any.

fn update_array_reference_count( &mut self, value: ValueId, increment: bool, ) -> Option<ValueId>

Increment or decrement the given value’s reference count if it is an array. If it is not an array, this does nothing. Note that inc_rc and dec_rc instructions are ignored outside of unconstrained code.

If there is an original value it indicates that we’re now decrementing it, but that should only happen if it hasn’t been changed by an array_set in the meantime, which in itself would make sure its RC is decremented.

Returns the ID of the array that was affected, if any.

pub fn record_error_type(&mut self, selector: ErrorSelector, typ: HirType)