noirc_evaluator::ssa::interpreter
Macro apply_fit_binop

macro_rules! apply_fit_binop {
    ($cons:expr, $lhs:expr, $rhs:expr, $f:expr, $g:expr, $lhs_num:expr, $rhs_num:expr) => { ... };
}

Applies a fallible integer binary operation on Fitted values, promoting values to Field in case there is an overflow, thus turning the operation infallible.

If the result is an overflow, it promotes the values to Field and performs the operation there. If the operation is applied on Unfit values, and the result fits in the original numeric type, it is converted back to a Fit value.

For example we would normally have an infallible wrapped_add, but we want to match ACIR by not wrapping around but extending into larger bit sizes.

Parameters

$cons: Constructor for a NumericValue

$lhs, $rhs: The Fitted values in the left-hand side and right-hand side operands.

$f: The function to apply on the integer values if both are Fit; returns None on overflow.

$g: The function to apply on Field values.

$lhs_num, $rhs_num: The original NumericValues.