noirc_evaluator::ssa::interpreter
Struct Interpreter

pub(crate) struct Interpreter<'ssa, W> {
    call_stack: Vec<CallContext>,
    functions: &'ssa BTreeMap<Id<Function>, Function>,
    options: InterpreterOptions,
    output: W,
    step_counter: usize,
}

Fields
call_stack: Vec<CallContext>

Contains each function called with main (or the first called function if the interpreter was manually invoked on a different function) at the front of the Vec.

functions: &'ssa BTreeMap<Id<Function>, Function>options: InterpreterOptions

The options the interpreter was created with.

output: W

Print output.

step_counter: usize

Number of instructions and terminators executed.

Implementations

impl<W: Write> Interpreter<'_, W>

pub(super) fn call_intrinsic( &mut self, intrinsic: Intrinsic, args: &[ValueId], results: &[ValueId], ) -> IResults

fn convert_error(err: BlackBoxResolutionError) -> InterpreterError

fn to_radix( &self, endian: Endian, element_type: NumericType, field_id: ValueId, field: FieldElement, radix: u32, result: ValueId, ) -> IResults

fn vector_push_back(&self, args: &[ValueId]) -> IResults

(length, vector, elem…) -> (length, vector)

fn vector_push_front(&self, args: &[ValueId]) -> IResults

(length, vector, elem…) -> (length, vector)

fn vector_pop_back(&self, args: &[ValueId]) -> IResults

(length, vector) -> (length, vector, elem…)

fn vector_pop_front(&self, args: &[ValueId]) -> IResults

(length, vector) -> (elem…, length, vector)

fn vector_insert(&self, args: &[ValueId]) -> IResults

(length, vector, index:u32, elem…) -> (length, vector)

fn vector_remove(&self, args: &[ValueId]) -> IResults

(length, vector, index:u32) -> (length, vector, elem…)

pub(super) fn call_print(&mut self, args: Vec<Value>) -> IResults

Print is not an intrinsic but it is treated like one.

impl<'ssa, W: Write> Interpreter<'ssa, W>

fn new(ssa: &'ssa Ssa, options: InterpreterOptions, output: W) -> Self

pub(crate) fn new_from_functions( functions: &'ssa BTreeMap<Id<Function>, Function>, options: InterpreterOptions, output: W, ) -> Self

pub(crate) fn functions(&self) -> &BTreeMap<Id<Function>, Function>

fn inc_step_counter(&mut self) -> Result<(), InterpreterError>

Increment the step counter, or return InterpreterError::OutOfBudget.

If there is no step limit, then it doesn’t increment the counter.

fn call_context(&self) -> &CallContext

fn call_context_mut(&mut self) -> &mut CallContext

fn global_scope(&self) -> &HashMap<ValueId, Value>

fn try_current_function(&self) -> Option<&'ssa Function>

fn current_function(&self) -> &'ssa Function

fn dfg(&self) -> &'ssa DataFlowGraph

fn in_unconstrained_context(&self) -> bool

fn define(&mut self, id: ValueId, value: Value) -> Result<(), InterpreterError>

Define or redefine a value. Redefinitions are expected in the case of loops.

pub(crate) fn interpret_globals(&mut self) -> Result<(), InterpreterError>

Interpret the global instructions.

Once this is complete, the interpreter can be reused for multiple function calls within the same SSA.

pub(crate) fn interpret_function( &mut self, function_id: Id<Function>, arguments: Vec<Value>, ) -> IResults

Interpret an entry point, assuming the globals have already been interpreted.

This resets any previous call stack and step counter.

fn call_function( &mut self, function_id: Id<Function>, arguments: Vec<Value>, ) -> IResults

Interpret a function call.

Unlike interpret_function this does not reset the state; it is meant to be used for internal calls.

fn lookup(&self, id: ValueId) -> Result<Value, InterpreterError>

fn lookup_helper<T>( &self, value_id: ValueId, instruction: &'static str, expected_type: &'static str, convert: impl FnOnce(&Value) -> Option<T>, ) -> Result<T, InterpreterError>

fn lookup_bool( &self, value_id: ValueId, instruction: &'static str, ) -> Result<bool, InterpreterError>

fn lookup_u32( &self, value_id: ValueId, instruction: &'static str, ) -> Result<u32, InterpreterError>

fn lookup_field( &self, value_id: ValueId, instruction: &'static str, ) -> Result<FieldElement, InterpreterError>

fn lookup_numeric( &self, value_id: ValueId, instruction: &'static str, ) -> Result<NumericValue, InterpreterError>

fn lookup_array_or_vector( &self, value_id: ValueId, instruction: &'static str, ) -> Result<ArrayValue, InterpreterError>

fn lookup_array_index( &self, value_id: ValueId, instruction: &'static str, length: u32, ) -> Result<u32, InterpreterError>

Look up an array index.

If the value exists but it’s Unfit, returns IndexOutOfBounds.

fn lookup_bytes( &self, value_id: ValueId, instruction: &'static str, ) -> Result<Vec<u8>, InterpreterError>

fn lookup_vec_u32( &self, value_id: ValueId, instruction: &'static str, ) -> Result<Vec<u32>, InterpreterError>

fn lookup_vec_u64( &self, value_id: ValueId, instruction: &'static str, ) -> Result<Vec<u64>, InterpreterError>

fn lookup_vec_field( &self, value_id: ValueId, instruction: &'static str, ) -> Result<Vec<FieldElement>, InterpreterError>

fn lookup_string( &self, value_id: ValueId, instruction: &'static str, ) -> Result<String, InterpreterError>

fn lookup_reference( &self, value_id: ValueId, instruction: &'static str, ) -> Result<ReferenceValue, InterpreterError>

fn lookup_all(&self, ids: &[ValueId]) -> Result<Vec<Value>, InterpreterError>

fn side_effects_enabled(&self, instruction: &Instruction) -> bool

fn interpret_instruction( &mut self, instruction: &Instruction, results: &[ValueId], ) -> Result<(), InterpreterError>

fn interpret_not( &mut self, id: ValueId, result: ValueId, ) -> Result<(), InterpreterError>

fn interpret_truncate( &mut self, value_id: ValueId, bit_size: u32, max_bit_size: u32, result: ValueId, ) -> Result<(), InterpreterError>

fn interpret_range_check( &mut self, value_id: ValueId, max_bit_size: u32, error_message: Option<&String>, side_effects_enabled: bool, ) -> Result<(), InterpreterError>

fn interpret_call( &mut self, function_id: ValueId, argument_ids: &[ValueId], results: &[ValueId], side_effects_enabled: bool, ) -> Result<(), InterpreterError>

fn try_get_function_name(&self, function: ValueId) -> String

Try to get a function’s name or approximate it if it is not known

fn reset_array_state(value: &mut Value) -> Result<(), InterpreterError>

Reset the value’s Shared states in each array within. This is used to mimic each invocation of the brillig vm receiving fresh values. No matter the history of this value (e.g. even if they were previously returned from another brillig function) the reference count should always be 1 and it shouldn’t alias any other arrays.

fn interpret_allocate( &mut self, result: ValueId, ) -> Result<(), InterpreterError>

fn interpret_load( &mut self, address: ValueId, result: ValueId, ) -> Result<(), InterpreterError>

fn interpret_store( &mut self, address: ValueId, value: ValueId, ) -> Result<(), InterpreterError>

fn interpret_array_get( &mut self, array: ValueId, index: ValueId, result: ValueId, side_effects_enabled: bool, ) -> Result<(), InterpreterError>

fn interpret_array_set( &mut self, array: ValueId, index: ValueId, value: ValueId, mutable: bool, result: ValueId, side_effects_enabled: bool, ) -> Result<(), InterpreterError>

fn decrement_rc(_array: &ArrayValue)

Decrement the ref-count of an array by 1.

fn interpret_inc_rc(&self, value_id: ValueId) -> Result<(), InterpreterError>

fn interpret_dec_rc(&self, value_id: ValueId) -> Result<(), InterpreterError>

fn interpret_if_else( &mut self, then_condition_id: ValueId, then_value: ValueId, else_condition_id: ValueId, else_value: ValueId, result: ValueId, ) -> Result<(), InterpreterError>

fn interpret_make_array( &mut self, elements: &Vector<ValueId>, result: ValueId, result_type: &Type, ) -> Result<(), InterpreterError>

fn interpret_binary( &self, binary: &Binary, side_effects_enabled: bool, ) -> Result<Value, InterpreterError>