noirc_evaluator::ssa

Module ir

Modules

basic_block

call_graph 🔒

Call graph analysis

cfg 🔒

dfg 🔒

dom 🔒

The dominator tree of a function, represented as a hash map of each reachable block id to its immediate dominator.

function

function_inserter 🔒

instruction

integer 🔒

map

post_order 🔒

The post-order for a given function represented as a vector of basic block ids.

printer 🔒

This file is for pretty-printing the SSA IR in a human-readable form for debugging.

target_cost 🔒

Brillig target cost estimation for SSA IR types.

types

value

Functions

is_printable_byte