noirc_evaluator::ssa::ir::call_graph
Struct CallGraph

pub(crate) struct CallGraph {
    graph: DiGraph<Id<Function>, usize>,
    ids_to_indices: FxHashMap<Id<Function>, NodeIndex>,
    indices_to_ids: FxHashMap<NodeIndex, Id<Function>>,
}

Represents a function call graph built from the SSA Internally, this is a directed graph where each node is a FunctionId and each edge represents a call from one function to another. Each edge has an associated weight representing the number of times a descendant node was called by an ancestor node.

Fields
graph: DiGraph<Id<Function>, usize>ids_to_indices: FxHashMap<Id<Function>, NodeIndex>indices_to_ids: FxHashMap<NodeIndex, Id<Function>>
Implementations

impl CallGraph

pub(crate) fn from_ssa(ssa: &Ssa) -> Self

Construct a CallGraph from the Ssa The edges in this graph are unweighted. Thus, it is a pure dependency map.

pub(crate) fn from_ssa_weighted(ssa: &Ssa) -> Self

Construct a CallGraph from the Ssa with its edges weighted. An edges weight refers to the numbers of times a descendant node was called by its ancestor.

fn from_deps_weighted( dependencies: BTreeMap<Id<Function>, BTreeMap<Id<Function>, usize>>, ) -> Self

pub(crate) fn from_deps( dependencies: HashMap<Id<Function>, BTreeSet<Id<Function>>>, ) -> Self

Construct a CallGraph from an explicit dependency mapping of (caller -> callees) The edges in this graph all have a weight of one. Thus, it is a pure dependency graph.

pub(crate) fn get_recursive_functions(&self) -> HashSet<Id<Function>>

Returns the set of all recursive functions.

A function is considered recursive if:

It is self-recursive (calls itself), or

It is part of a mutual recursion cycle with other functions.

pub(crate) fn sccs(&self) -> (Vec<Vec<Id<Function>>>, HashSet<Id<Function>>)

Returns both the call graph’s strongly connected components (SCCs) as well as a utility set of all recursive functions.

pub(crate) fn graph(&self) -> &DiGraph<Id<Function>, usize>

pub(crate) fn ids_to_indices(&self) -> &HashMap<Id<Function>, PetGraphIndex>

pub(crate) fn indices_to_ids(&self) -> &HashMap<PetGraphIndex, Id<Function>>

pub(crate) fn build_acyclic_subgraph( &self, recursive_functions: &HashSet<Id<Function>>, ) -> CallGraph

pub(crate) fn callees( &self, ) -> BTreeMap<Id<Function>, BTreeMap<Id<Function>, usize>>

Fetch for each function the set of functions called by it, and how many times it does so.

pub(crate) fn callers( &self, ) -> BTreeMap<Id<Function>, BTreeMap<Id<Function>, usize>>

Fetch for each function the set of functions that call it, and how many times they do so.

pub(crate) fn times_called(&self) -> HashMap<Id<Function>, usize>

Compute the times each function is called from any other function.

pub(crate) fn reachable_from( &self, roots: impl IntoIterator<Item = Id<Function>>, ) -> HashSet<Id<Function>>

Returns all functions reachable from the provided root(s).

This function uses DFS internally to find all nodes reachable from the provided root(s).