noirc_evaluator::ssa::ir::dfg
Module simplify

Modules

binary 🔒

call 🔒

cast 🔒

constrain 🔒

value_merger 🔒

Enums

SimplifyResult 🔒

Contains the result to Instruction::simplify, specifying how the instruction should be simplified.

Functions

optimize_length_one_array_read 🔒

Given an array access on a length 1 array such as:

simplify 🔒

Try to simplify this instruction. If the instruction can be simplified to a known value, that value is returned. Otherwise None is returned.

try_optimize_array_get_from_previous_set 🔒

Given a chain of operations like: v1 = array_set [10, 11, 12], index 1, value: 5 v2 = array_set v1, index 2, value: 6 v3 = array_set v2, index 2, value: 7 v4 = array_get v3, index 1

try_optimize_array_set_from_previous_get 🔒

If we have an array set whose value is from an array get on the same array at the same index, we can simplify that array set to the array we were looking to perform an array set upon.