noirc_evaluator::ssa::ir::dfg::simplify
Module call

Modules

blackbox 🔒

Functions

array_is_constant 🔒

constant_to_radix 🔒

decrement_slice_length 🔒

increment_slice_length 🔒

make_array 🔒

make_constant_array 🔒

simplify_black_box_func 🔒

Try to simplify this black box call. If the call can be simplified to a known value, that value is returned. Otherwise SimplifyResult::None is returned.

simplify_call 🔒

Try to simplify this call instruction. If the instruction can be simplified to a known value, that value is returned. Otherwise None is returned.

simplify_constant_to_radix 🔒

Returns a slice (represented by a tuple (len, slice)) of constants corresponding to the limbs of the radix decomposition.

simplify_derive_generators 🔒

Replaces a call to derive_pedersen_generators with the results of the computation.

simplify_slice_pop_back 🔒

simplify_slice_push_back 🔒

to_u8_vec 🔒

update_slice_length 🔒

Slices have a tuple structure (slice length, slice contents) to enable logic that uses dynamic slice lengths (such as with merging slices in the flattening pass). This method codegens an update to the slice length.