noirc_evaluator::ssa::ir::dfg::simplify::call
Function simplify_call

pub(super) fn simplify_call(
    func: ValueId,
    arguments: &[ValueId],
    dfg: &mut DataFlowGraph,
    block: BasicBlockId,
    ctrl_typevars: Option<Vec<Type>>,
    call_stack: CallStackId,
) -> SimplifyResult

Try to simplify this call instruction. If the instruction can be simplified to a known value, that value is returned. Otherwise None is returned.

The block parameter indicates the block any new instructions that are part of a call’s simplification will be inserted into. For example, all vector intrinsics require updates to the vector length, which requires inserting a binary instruction. This update instruction must be inserted into the same block that the call itself is being simplified into.