noirc_evaluator::ssa::ir::dfg::simplify
Function optimize_length_one_array_read

fn optimize_length_one_array_read(
    dfg: &mut DataFlowGraph,
    block: BasicBlockId,
    call_stack: CallStackId,
    array: ValueId,
    index: ValueId,
) -> SimplifyResult

Given an array access on a length 1 array such as:

v2 = make_array [v0] : [Field; 1]
v3 = array_get v2, index v1 -> Field

We want to replace the array read with the only valid value which can be read from the array while ensuring that if there is an attempt to read past the end of the array then the program fails.

We then inject an explicit assertion that the index variable has the value zero while replacing the value being used in the array_get instruction with a constant value of zero. This then results in the SSA:

v2 = make_array [v0] : [Field; 1]
constrain v1 == u32 0, "Index out of bounds"
v4 = array_get v2, index u32 0 -> Field

We then attempt to resolve the array read immediately.

Note that this does not work if the array has length 1, but contains a complex type such as tuple, which consists of multiple elements. If that is the case than the index will most likely not be a constant, but a base plus an offset, and if the array contains repeated elements of the same type for example, we wouldn’t be able to come up with a constant offset even if we knew the return type.