noirc_evaluator::ssa::ir::dfg
Struct DataFlowGraph

pub(crate) struct DataFlowGraph {

    runtime: RuntimeType,
    instructions: DenseMap<Instruction>,
    results: FxHashMap<Id<Instruction>, SmallVec<[ValueId; 1]>>,
    values: DenseMap<Value>,
    constants: FxHashMap<(FieldElement, NumericType), ValueId>,
    functions: FxHashMap<Id<Function>, ValueId>,
    intrinsics: FxHashMap<Intrinsic, ValueId>,
    foreign_functions: FxHashMap<String, ValueId>,
    blocks: DenseMap<BasicBlock>,
    locations: FxHashMap<Id<Instruction>, CallStackId>,
    pub(crate) call_stack_data: CallStackHelper,
    pub(crate) data_bus: DataBus,
    pub(crate) globals: Arc<GlobalsGraph>,
    pub(crate) function_purities: Arc<FxHashMap<Id<Function>, Purity>>,
    pub(crate) brillig_arrays_offset: bool,

}

The DataFlowGraph contains most of the actual data in a function including its blocks, instructions, and values. This struct is largely responsible for owning most data in a function and handing out Ids to this data that can be shared without worrying about ownership.

Fields
runtime: RuntimeType

Runtime of the function that owns this DataFlowGraph. This might change during the runtime_separation pass where ACIR functions are cloned as Brillig functions.

instructions: DenseMap<Instruction>

All of the instructions in a function

results: FxHashMap<Id<Instruction>, SmallVec<[ValueId; 1]>>

Stores the results for a particular instruction.

An instruction may return multiple values and for this, we will also use the cranelift strategy to fetch them via indices.

Currently, we need to define them in a better way Call instructions require the func signature, but other instructions may need some more reading on my part

values: DenseMap<Value>

Storage for all of the values defined in this function.

constants: FxHashMap<(FieldElement, NumericType), ValueId>

Each constant is unique, attempting to insert the same constant twice will return the same ValueId.

functions: FxHashMap<Id<Function>, ValueId>

Contains each function that has been imported into the current function. A unique ValueId for each function’s Value::Function is stored so any given FunctionId will always have the same ValueId within this function.

intrinsics: FxHashMap<Intrinsic, ValueId>

Contains each intrinsic that has been imported into the current function. This map is used to ensure that the ValueId for any given intrinsic is always represented by only 1 ValueId within this function.

foreign_functions: FxHashMap<String, ValueId>

Contains each foreign function that has been imported into the current function. This map is used to ensure that the ValueId for any given foreign function is always represented by only 1 ValueId within this function.

blocks: DenseMap<BasicBlock>

All blocks in a function

locations: FxHashMap<Id<Instruction>, CallStackId>

Source location of each instruction for debugging and issuing errors.

The CallStack here corresponds to the entire callstack of locations. Initially this only contains the actual location of the instruction. During inlining, a new location will be pushed to each instruction for the location of the function call of the function the instruction was originally located in. Once inlining is complete, the locations Vec here should contain the entire callstack for each instruction.

Instructions inserted by internal SSA passes that don’t correspond to user code may not have a corresponding location.

call_stack_data: CallStackHelperdata_bus: DataBusglobals: Arc<GlobalsGraph>function_purities: Arc<FxHashMap<Id<Function>, Purity>>brillig_arrays_offset: bool

Indicate whether the Brillig array index offset optimizations have been performed.

Implementations

impl DataFlowGraph

pub(crate) fn runtime(&self) -> RuntimeType

Runtime type of the function.

pub(crate) fn set_runtime(&mut self, runtime: RuntimeType)

Set runtime type of the function.

pub(crate) fn make_block(&mut self) -> BasicBlockId

Creates a new basic block with no parameters. After being created, the block is unreachable in the current function until another block is made to jump to it.

pub(crate) fn make_block_with_parameters_from_block( &mut self, block: BasicBlockId, ) -> BasicBlockId

Create a new block with the same parameter count and parameter types from the given block. This is a somewhat niche operation used in loop unrolling but is included here as doing it outside the DataFlowGraph would require cloning the parameters.

pub(crate) fn values_iter( &self, ) -> impl DoubleEndedIterator<Item = (ValueId, &Value)>

Iterate over every Value in this DFG in no particular order, including unused Values

pub(crate) fn block_parameters(&self, block: BasicBlockId) -> &[ValueId]

Returns the parameters of the given block

pub(crate) fn make_instruction( &mut self, instruction_data: Instruction, ctrl_typevars: Option<Vec<Type>>, ) -> Id<Instruction>

Inserts a new instruction into the DFG. This does not add the instruction to the block. Returns the id of the new instruction and its results.

Populates the instruction’s results with the given ctrl_typevars if the instruction is a Load, Call, or Intrinsic. Otherwise the instruction’s results will be known by the instruction itself and None can safely be passed for this parameter.

pub(crate) fn is_handled_by_runtime(&self, instruction: &Instruction) -> bool

Check if the function runtime would simply ignore this instruction.

fn insert_instruction_without_simplification( &mut self, instruction_data: Instruction, block: BasicBlockId, ctrl_typevars: Option<Vec<Type>>, call_stack: CallStackId, ) -> Id<Instruction>

pub(crate) fn insert_instruction_and_results_without_simplification( &mut self, instruction: Instruction, block: BasicBlockId, ctrl_typevars: Option<Vec<Type>>, call_stack: CallStackId, ) -> InsertInstructionResult<'_>

pub(crate) fn insert_instruction_and_results( &mut self, instruction: Instruction, block: BasicBlockId, ctrl_typevars: Option<Vec<Type>>, call_stack: CallStackId, ) -> InsertInstructionResult<'_>

Simplifies a new instruction and inserts it at the end of the given block and returns its results. If the instruction is not handled by the current runtime, InstructionRemoved is returned.

pub(crate) fn insert_instruction_and_results_if_simplified( &mut self, instruction: Instruction, block: BasicBlockId, ctrl_typevars: Option<Vec<Type>>, call_stack: CallStackId, existing_id: Option<Id<Instruction>>, ) -> InsertInstructionResult<'_>

Simplifies a potentially existing instruction and inserts it only if it changed.

pub(crate) fn set_instruction( &mut self, id: Id<Instruction>, instruction: Instruction, )

Replace an existing instruction with a new one.

pub(crate) fn replace_values_in_block( &mut self, block: BasicBlockId, mapping: &ValueMapping, )

Replaces values in the given block according to the given mapping.

pub(crate) fn replace_values_in_block_instructions( &mut self, block: BasicBlockId, mapping: &ValueMapping, )

Replaces values in the given block instructions according to the given mapping.

pub(crate) fn replace_values_in_block_terminator( &mut self, block: BasicBlockId, mapping: &ValueMapping, )

Replaces values in the given block terminator (if it has any) according to the given mapping.

pub(crate) fn set_type_of_value(&mut self, value_id: ValueId, target_type: Type)

Set the type of value_id to the target_type.

pub(crate) fn make_constant( &mut self, constant: FieldElement, typ: NumericType, ) -> ValueId

Creates a new constant value, or returns the Id to an existing one if one already exists.

pub(crate) fn make_global(&mut self, typ: Type) -> ValueId

pub(crate) fn import_function(&mut self, function: Id<Function>) -> ValueId

Gets or creates a ValueId for the given FunctionId.

pub(crate) fn import_foreign_function(&mut self, function: &str) -> ValueId

Gets or creates a ValueId for the given FunctionId.

pub(crate) fn import_intrinsic(&mut self, intrinsic: Intrinsic) -> ValueId

Gets or creates a ValueId for the given Intrinsic.

pub(crate) fn get_intrinsic(&self, intrinsic: Intrinsic) -> Option<&ValueId>

pub(crate) fn make_instruction_results( &mut self, instruction: Id<Instruction>, ctrl_typevars: Option<Vec<Type>>, )

Attaches results to the instruction, clearing any previous results.

This does not normally need to be called manually as it is called within make_instruction automatically.

Returns the results of the instruction

fn for_each_instruction_result_type( &mut self, instruction_id: Id<Instruction>, ctrl_typevars: Option<Vec<Type>>, f: impl FnMut(&mut Self, Type), )

Return the result types of this instruction.

In the case of Load, Call, and Intrinsic, the function’s result type may be unknown. In this case, the given ctrl_typevars are returned instead. ctrl_typevars is taken in as an Option since it is common to omit them when getting the type of an instruction that does not require them. Compared to passing an empty Vec, Option has the benefit of panicking if it is accidentally used for a Call instruction, rather than silently returning the empty Vec and continuing.

pub(crate) fn type_of_value(&self, value: ValueId) -> Type

Returns the type of a given value

pub(crate) fn get_value_max_num_bits(&self, value: ValueId) -> u32

Returns the maximum possible number of bits that value can potentially be.

Should value be a numeric constant then this function will return the exact number of bits required, otherwise it will return the minimum number of bits based on type information.

pub(crate) fn value_is_reference(&self, value: ValueId) -> bool

True if the type of this value is Type::Reference. Using this method over type_of_value avoids cloning the value’s type.

pub(crate) fn instruction_results( &self, instruction_id: Id<Instruction>, ) -> &[ValueId]

Returns all of result values which are attached to this instruction.

pub(crate) fn instruction_result<const N: usize>( &self, instruction_id: Id<Instruction>, ) -> [ValueId; N]

Returns N results, asserting that there are exactly N items.

pub(crate) fn remove_instruction(&mut self, instruction: Id<Instruction>)

Remove an instruction by replacing it with a Noop instruction. Doing this avoids shifting over each instruction after this one in its block’s instructions vector.

pub(crate) fn retain_instructions( &mut self, keep: impl Fn(Id<Instruction>) -> bool, )

Remove instructions for which the keep functions returns false.

pub(crate) fn add_block_parameter( &mut self, block_id: BasicBlockId, typ: Type, ) -> ValueId

Add a parameter to the given block

pub(crate) fn get_numeric_constant( &self, value: ValueId, ) -> Option<FieldElement>

Returns the field element represented by this value if it is a numeric constant. Returns None if the given value is not a numeric constant.

Use get_integer_constant if the underlying values need to be compared as signed integers.

pub(crate) fn get_integer_constant( &self, value: ValueId, ) -> Option<IntegerConstant>

Similar to get_numeric_constant but returns the value as a signed or unsigned integer. Returns None if the given value is not an integer constant.

pub(crate) fn get_numeric_constant_with_type( &self, value: ValueId, ) -> Option<(FieldElement, NumericType)>

Returns the field element and type represented by this value if it is a numeric constant. Returns None if the given value is not a numeric constant.

pub(crate) fn get_array_constant( &self, value: ValueId, ) -> Option<(Vector<ValueId>, Type)>

Returns the item values in with this ValueId if it refers to an array constant, along with the type of the array item. Otherwise, this returns None.

pub(crate) fn try_get_array_length(&self, value: ValueId) -> Option<u32>

If this value is an array, return the length of the array as indicated by its type. Otherwise, return None.

pub(crate) fn try_get_vector_capacity(&self, value: ValueId) -> Option<u32>

pub(crate) fn get_string(&self, value: ValueId) -> Option<String>

If this value points to an array of constant bytes, returns a string consisting of those bytes if they form a valid UTF-8 string.

pub(crate) fn is_safe_index(&self, index: ValueId, array: ValueId) -> bool

A constant index less than the array length is safe

pub(crate) fn set_block_terminator( &mut self, block: BasicBlockId, terminator: TerminatorInstruction, )

Sets the terminator instruction for the given basic block

pub(crate) fn inline_block( &mut self, source: BasicBlockId, destination: BasicBlockId, )

Moves the entirety of the given block’s contents into the destination block. The source block afterward will be left in a valid but emptied state. The destination block will also have its terminator overwritten with that of the source block.

pub(crate) fn get_instruction_call_stack( &self, instruction: Id<Instruction>, ) -> CallStack

pub(crate) fn get_instruction_call_stack_id( &self, instruction: Id<Instruction>, ) -> CallStackId

pub(crate) fn get_call_stack(&self, call_stack: CallStackId) -> CallStack

pub(crate) fn get_value_call_stack(&self, value: ValueId) -> CallStack

pub(crate) fn get_value_call_stack_id(&self, value: ValueId) -> CallStackId

pub(crate) fn is_constant(&self, argument: ValueId) -> bool

True if the given ValueId refers to a constant value. A MakeArray instruction is considered constant if all its elements are constants.

pub(crate) fn is_constant_true(&self, argument: ValueId) -> bool

True that the input is a non-zero Value::NumericConstant

pub(crate) fn is_global(&self, value: ValueId) -> bool

pub(crate) fn get_local_or_global_instruction( &self, value: ValueId, ) -> Option<&Instruction>

Uses value information to determine whether an instruction is from this function’s DFG or the global space’s DFG.

pub(crate) fn set_function_purities( &mut self, purities: Arc<FxHashMap<Id<Function>, Purity>>, )

pub(crate) fn purity_of(&self, function: Id<Function>) -> Option<Purity>

pub(crate) fn array_offset(&self, array: ValueId, index: ValueId) -> ArrayOffset

Determine the appropriate ArrayOffset to use for indexing an array or vector.

pub(crate) fn is_returned_in_databus( &self, instruction_id: Id<Instruction>, ) -> bool

Check if the results of an instruction are used in the databus to return a value..

This only applies to ACIR, as in Brillig the databus will always be empty.