noirc_evaluator::ssa::ir::function
Struct Function

pub struct Function {
    entry_block: BasicBlockId,
    name: String,
    id: Option<Id<Function>>,
    pub(crate) dfg: DataFlowGraph,
}

A function holds a list of instructions. These instructions are further grouped into Basic blocks

All functions outside of the current function are seen as external. To reference external functions its FunctionId can be used but this cannot be checked for correctness until inlining is performed.

Fields
entry_block: BasicBlockId

The first basic block in the function

name: String

Name of the function for debugging only

id: Option<Id<Function>>dfg: DataFlowGraph

The DataFlowGraph holds the majority of data pertaining to the function including its blocks, instructions, and values.

Implementations

impl Function

pub(crate) fn new(name: String, id: Id<Function>) -> Self

Creates a new function with an automatically inserted entry block.

Note that any parameters or attributes of the function must be manually added later.

pub(crate) fn new_for_globals() -> Self

Globals are generated using the same codegen process as functions. To avoid a recursive global context we should create a pseudo function to mock a globals context.

pub(crate) fn clone_with_id(id: Id<Function>, another: &Function) -> Self

Creates a new function as a clone of the one passed in with the passed in id.

pub(crate) fn clone_signature(id: Id<Function>, another: &Function) -> Self

Takes the signature (function name & runtime) from a function but does not copy the body.

pub fn name(&self) -> &str

The name of the function. Used exclusively for debugging purposes.

pub(crate) fn id(&self) -> Id<Function>

The id of the function.

pub(crate) fn runtime(&self) -> RuntimeType

Runtime type of the function.

pub(crate) fn set_runtime(&mut self, runtime: RuntimeType)

Set runtime type of the function.

pub(crate) fn set_globals(&mut self, globals: Arc<GlobalsGraph>)

pub(crate) fn is_no_predicates(&self) -> bool

pub(crate) fn entry_block(&self) -> BasicBlockId

Retrieves the entry block of a function.

A function’s entry block contains the instructions to be executed first when the function is called. The function’s parameters are also stored as the entry block’s parameters.

pub(crate) fn parameters(&self) -> &[ValueId]

Returns the parameters of this function. The parameters will always match that of this function’s entry block.

pub(crate) fn returns(&self) -> Option<&[ValueId]>

Returns the return types of this function. None might be returned if the function ends up with all of its block terminators being jmp, jmpif or unreachable.

pub(crate) fn reachable_blocks(&self) -> BTreeSet<BasicBlockId>

Collects all the reachable blocks of this function.

Note that self.dfg.basic_blocks_iter() iterates over all blocks, whether reachable or not. This function should be used if you want to iterate only reachable blocks.

pub(crate) fn signature(&self) -> Signature

pub(crate) fn find_last_block(&self) -> BasicBlockId

Finds the block of the function with the Return instruction

pub(crate) fn num_instructions(&self) -> usize

Total number of instructions in the reachable blocks of this function.

pub fn view(&self) -> FunctionView<'_>

impl Function

pub(crate) fn array_set_optimization(&mut self)

impl Function

pub(crate) fn as_vector_optimization(&mut self)

impl Function

fn brillig_array_get_and_set(&mut self)

impl Function

fn check_u128_mul_overflow(&mut self)

impl Function

fn checked_to_unchecked(&mut self)

impl Function

pub(crate) fn constant_fold( &mut self, use_constraint_info: bool, max_iter: usize, interpreter: &mut Option<Interpreter<'_, Empty>>, )

The structure of this pass is simple: Go through each block and re-insert all instructions.

impl Function

fn prune_dead_parameters( &mut self, unused_params: &HashMap<BasicBlockId, Vec<ValueId>>, main_id: Id<Function>, ) -> Option<Vec<bool>>

See prune_dead_parameters module for more information

fn update_predecessor_terminators( &mut self, cfg: &ControlFlowGraph, target_block: BasicBlockId, keep_list: &[bool], )

Update terminator arguments of predecessor blocks after pruning.

impl Function

fn dead_instruction_elimination( &mut self, insert_out_of_bounds_checks: bool, flattened: bool, ) -> HashMap<BasicBlockId, Vec<ValueId>>

Removes any unused instructions in the reachable blocks of the given function.

This method is designed to be run within the context of the full SSA, not in isolation. Running DIE on a single function may cause inconsistencies, such as leaving dangling unused parameters. The pruning of block parameters depends on the full SSA context. Therefore, this method must remain private, and DIE should run over the entire SSA, ensuring proper tracking of unused parameters across all blocks.

The blocks of the function are iterated in post order, such that any blocks containing instructions that reference results from an instruction in another block are evaluated first. If we did not iterate blocks in this order we could not safely say whether or not the results of its instructions are needed elsewhere.

Returns

After processing all functions, the union of these sets enables determining the unused globals.

A mapping of (block id -> unused parameters) for the given function. This can be used by follow-up passes to prune unused parameters from blocks.

impl Function

fn evaluate_static_assert_and_assert_constant( &mut self, ) -> Result<(), RuntimeError>

impl Function

fn expand_signed_checks(&mut self)

The structure of this pass is simple: Go through each block and re-insert all instructions, decomposing any checked signed arithmetic to have explicit overflow checks.

impl Function

fn expand_signed_math(&mut self)

The structure of this pass is simple: Go through each block and re-insert all instructions, decomposing any signed “less than”, “div” and “mod” operations to be done using unsigned types, but only if this is an ACIR function.

impl Function

pub(super) fn inlined( &self, ssa: &Ssa, should_inline_call: &impl Fn(&Function) -> bool, ) -> Result<Function, RuntimeError>

Create a new function which has the functions called by this one inlined into its body.

impl Function

pub(super) fn loop_invariant_code_motion(&mut self)

impl Function

fn make_constrain_not_equal(&mut self)

impl Function

pub(crate) fn mem2reg(&mut self)

impl Function

fn is_pure(&self) -> Purity

impl Function

pub(crate) fn remove_paired_rc(&mut self)

This function is very simplistic for now. It takes advantage of the fact that dec_rc instructions are currently issued only at the end of a function for parameters and will only check the first and last block for inc & dec rc instructions to be removed. The rest of the function is still checked for array_set instructions.

This restriction lets this function largely ignore merging intermediate results from other blocks and handling loops.

impl Function

fn remove_bit_shifts(&mut self)

If this is an ACIR function, go through every instruction, replacing bit shifts with more primitive arithmetic operations,

impl Function

fn remove_enable_side_effects(&mut self)

impl Function

pub(crate) fn remove_if_else(&mut self) -> RtResult<()>

impl Function

pub(crate) fn remove_truncate_after_range_check(&mut self)

impl Function

fn remove_unreachable_instructions(&mut self, is_main: bool)

impl Function

pub(crate) fn remove_unused_instructions(&mut self)

Remove instructions which aren’t used in any blocks.

impl Function

pub(crate) fn simple_optimization<F>(&mut self, f: F)
where F: FnMut(&mut SimpleOptimizationContext<'_, '_>),

Performs a simple optimization according to the given callback.

The function’s reverse post order are traversed in turn, and instructions in those blocks are then traversed in turn. For each one, f will be called with a context.

The current instruction will be inserted at the end of the callback given to mutate unless remove_current_instruction or insert_current_instruction are called.

insert_current_instruction is useful if you need to insert new instructions after the current one, so this can be done before the callback ends.

replace_value can be used to replace a value with another one. This substitution will be performed in all subsequent instructions.

pub(crate) fn simple_optimization_result<F>(&mut self, f: F) -> RtResult<()>
where F: FnMut(&mut SimpleOptimizationContext<'_, '_>) -> RtResult<()>,

Performs a simple optimization according to the given callback, returning early if an error occurred.

The function’s reverse post order are traversed in turn, and instructions in those blocks are then traversed in turn. For each one, f will be called with a context.

The current instruction will be inserted at the end of the callback given to mutate unless remove_current_instruction or insert_current_instruction are called.

insert_current_instruction is useful if you need to insert new instructions after the current one, so this can be done before the callback ends.

replace_value can be used to replace a value with another one. This substitution will be performed in all subsequent instructions.

impl Function

pub(crate) fn simplify_function(&mut self)

Simplify a function’s cfg by going through each block to check for any simple blocks that can be inlined into their predecessor.

impl Function

pub(super) fn unroll_loops_iteratively(&mut self) -> Result<bool, RuntimeError>

Try to unroll loops in the function.

Returns an Err if it cannot be done, for example because the loop bounds cannot be determined at compile time. This can happen during pre-processing, but it should still leave the function in a partially unrolled, but valid state.

If successful, returns a flag indicating whether any loops have been unrolled.

fn try_unroll_loops(&mut self) -> (bool, Vec<RuntimeError>)

Unroll all loops within the function. Any loops which fail to be unrolled (due to using non-constant indices) will be unmodified. Returns a flag indicating whether any blocks have been modified.

Loop unrolling in brillig can lead to a code explosion currently. This can also be true for ACIR, but we have no alternative to unrolling in ACIR. Brillig also generally prefers smaller code rather than faster code, so we only attempt to unroll small loops, which we decide on a case-by-case basis.

impl Function

pub(crate) fn verify_no_dynamic_indices_to_references(&self) -> RtResult<()>