noirc_evaluator::ssa::ir::function_inserter
Struct FunctionInserter

pub(crate) struct FunctionInserter<'f> {
    pub(crate) function: &'f mut Function,
    values: FxHashMap<ValueId, ValueId>,
}

The FunctionInserter can be used to help modify existing Functions and map old values to new values after re-inserting optimized versions of old instructions.

Fields
function: &'f mut Functionvalues: FxHashMap<ValueId, ValueId>
Implementations

impl<'f> FunctionInserter<'f>

pub(crate) fn new(function: &'f mut Function) -> FunctionInserter<'f>

pub(crate) fn resolve(&self, value: ValueId) -> ValueId

Resolves a ValueId to its new, updated value. If there is no updated value for this id, this returns the same ValueId that was passed in.

pub(crate) fn try_map_value(&mut self, key: ValueId, value: ValueId)

Insert a key, value pair if the key isn’t already present in the map

pub(crate) fn map_value(&mut self, key: ValueId, value: ValueId)

Insert a key, value pair in the map

fn validate_map_value(&self, key: ValueId, value: ValueId)

Sanity check that we are not creating back-and-forth cycles. Doesn’t catch longer cycles, but has detected mistakes with reusing instructions.

pub(crate) fn map_instruction( &mut self, id: Id<Instruction>, ) -> (Instruction, CallStackId)

Get an instruction and make sure all the values in it are freshly resolved.

pub(crate) fn map_instruction_in_place(&mut self, id: Id<Instruction>)

Get an instruction, map all its values, and replace it with the resolved instruction.

pub(crate) fn map_terminator_in_place(&mut self, block: BasicBlockId)

Maps a terminator in place, replacing any ValueId in the terminator with the resolved version of that value id from this FunctionInserter’s internal value mapping.

pub(crate) fn map_data_bus_in_place(&mut self)

Maps the data bus in place, replacing any ValueId in the data bus with the resolved version of that value id from this FunctionInserter’s internal value mapping.

pub(crate) fn push_instruction( &mut self, id: Id<Instruction>, block: BasicBlockId, allow_reinsert: bool, ) -> Option<Id<Instruction>>

Push an instruction by ID, after re-mapping the values in it, to the given block and return its new InstructionId. If the instruction was simplified out of the program, None is returned.

pub(crate) fn push_instruction_value( &mut self, instruction: Instruction, id: Id<Instruction>, block: BasicBlockId, call_stack: CallStackId, allow_reinsert: bool, ) -> InsertInstructionResult<'_>

Push a instruction that already exists with an ID, given as an instance with potential modification already applied to it.

If allow_insert is set, we consider reinserting the instruction as it is, if it hasn’t changed and cannot be simplified. Consider using this if we are re-inserting when we take instructions from a block and re-insert them into the same block.

fn insert_new_instruction_results( values: &mut HashMap<ValueId, ValueId>, old_results: &[ValueId], new_results: &InsertInstructionResult<'_>, )

Modify the values HashMap to remember the mapping between an instruction result’s previous ValueId (from the source_function) and its new ValueId in the destination function.

pub(crate) fn remember_block_params( &mut self, block: BasicBlockId, new_values: &[ValueId], )

Associates each block parameter with a value, unless the parameter already has a value, in which case it is kept as-is.

pub(crate) fn remember_block_params_from_block( &mut self, block: BasicBlockId, new_block: BasicBlockId, )

pub(crate) fn extract_mapping(&self, mapping: &mut HashMap<ValueId, ValueId>)

Merge the internal mapping into the given mapping The merge is guaranteed to be coherent because ambiguous cases are prevented

pub(crate) fn set_mapping(&mut self, mapping: HashMap<ValueId, ValueId>)