noirc_evaluator::ssa::ir::instruction
Enum Instruction

pub enum Instruction {

    Binary(Binary),
    Cast(ValueId, NumericType),
    Not(ValueId),
    Truncate {
        value: ValueId,
        bit_size: u32,
        max_bit_size: u32,
    },
    Constrain(ValueId, ValueId, Option<ConstrainError>),
    ConstrainNotEqual(ValueId, ValueId, Option<ConstrainError>),
    RangeCheck {
        value: ValueId,
        max_bit_size: u32,
        assert_message: Option<String>,
    },
    Call {
        func: ValueId,
        arguments: Vec<ValueId>,
    },
    Allocate,
    Load {
        address: ValueId,
    },
    Store {
        address: ValueId,
        value: ValueId,
    },
    EnableSideEffectsIf {
        condition: ValueId,
    },
    ArrayGet {
        array: ValueId,
        index: ValueId,
    },
    ArraySet {
        array: ValueId,
        index: ValueId,
        value: ValueId,
        mutable: bool,
    },
    IncrementRc {
        value: ValueId,
    },
    DecrementRc {
        value: ValueId,
    },
    IfElse {
        then_condition: ValueId,
        then_value: ValueId,
        else_condition: ValueId,
        else_value: ValueId,
    },
    MakeArray {
        elements: Vector<ValueId>,
        typ: Type,
    },
    Noop,

}

Instructions are used to perform tasks. The instructions that the IR is able to specify are listed below.

Variants

Binary(Binary)

Binary Operations like +, -, *, /, ==, !=

Cast(ValueId, NumericType)

Converts Value into the given NumericType

This operation only changes the type of the value, it does not change the value itself. It is expected that the value can fit into the target type. For instance a value of type u32 casted to u8 must already fit into 8 bits A value of type i8 cannot be casted to ‘i16’ since the value would need to include the sign bit (which is the MSB) Ssa code-gen must ensure that the necessary truncation or sign extension is performed when emitting a Cast instruction.

Not(ValueId)

Computes a bit wise not

Truncate

Truncates value to bit_size

Fields

value: ValueId

bit_size: u32

max_bit_size: u32

Constrain(ValueId, ValueId, Option<ConstrainError>)

Constrains two values to be equal to one another.

ConstrainNotEqual(ValueId, ValueId, Option<ConstrainError>)

Constrains two values to not be equal to one another.

RangeCheck

Range constrain value to max_bit_size

Fields

value: ValueId

max_bit_size: u32

assert_message: Option<String>

Call

Performs a function call with a list of its arguments.

Fields

func: ValueId

arguments: Vec<ValueId>

Allocate

Allocates a region of memory. Note that this is not concerned with the type of memory, the type of element is determined when loading this memory. This is used for representing mutable variables and references.

Load

Loads a value from memory.

Fields

address: ValueId

Store

Writes a value to memory.

Fields

address: ValueId

value: ValueId

EnableSideEffectsIf

Provides a context for all instructions that follow up until the next EnableSideEffectsIf is encountered, for stating a condition that determines whether such instructions are allowed to have side-effects.

For example,

EnableSideEffectsIf condition0;
code0;
EnableSideEffectsIf condition1;
code1;

code0 will have side effects iff condition0 evaluates to true

code1 will have side effects iff condition1 evaluates to true

This instruction is only emitted after the cfg flattening pass, and is used to annotate instruction regions with a condition that corresponds to their position in the CFG’s if-branching structure.

Fields

condition: ValueId

ArrayGet

Retrieve a value from an array at the given index

Fields

array: ValueId

index: ValueId

ArraySet

Creates a new array with the new value at the given index. All other elements are identical to those in the given array. This will not modify the original array unless mutable is set. This flag is off by default and only enabled when optimizations determine it is safe.

Fields

array: ValueId

index: ValueId

value: ValueId

mutable: bool

IncrementRc

An instruction to increment the reference count of a value.

This currently only has an effect in Brillig code where array sharing and copy on write is implemented via reference counting. In ACIR code this is done with im::Vector and these IncrementRc instructions are ignored.

Fields

value: ValueId

DecrementRc

An instruction to decrement the reference count of a value.

This currently only has an effect in Brillig code where array sharing and copy on write is implemented via reference counting. In ACIR code this is done with im::Vector and these DecrementRc instructions are ignored.

Fields

value: ValueId

IfElse

Merge two values returned from opposite branches of a conditional into one.

if then_condition {
    then_value
} else {   // else_condition = !then_condition
    else_value
}

Fields

then_condition: ValueId

then_value: ValueId

else_condition: ValueId

else_value: ValueId

MakeArray

Creates a new array or slice.

typ should be an array or slice type with an element type matching each of the elements values’ types.

Fields

elements: Vector<ValueId>

typ: Type

Noop

A No-op instruction. These are intended to replace other instructions in a block’s instructions vector without having to move each instruction afterward.

A No-op has no results and is always removed when Instruction::simplify is called. When replacing another instruction, the instruction’s results should always be mapped to a new value since they will not be able to refer to their original instruction value any more.

Implementations

impl Instruction

pub(crate) fn binary( operator: BinaryOp, lhs: ValueId, rhs: ValueId, ) -> Instruction

Returns a binary instruction with the given operator, lhs, and rhs

pub(crate) fn result_type(&self) -> InstructionResultType

Returns the type that this instruction will return.

pub(crate) fn requires_ctrl_typevars(&self) -> bool

True if this instruction requires specifying the control type variables when inserting this instruction into a DataFlowGraph.

pub(crate) fn requires_acir_gen_predicate(&self, dfg: &DataFlowGraph) -> bool

If true the instruction will depend on enable_side_effects context during acir-gen.

pub(crate) fn has_side_effects(&self, dfg: &DataFlowGraph) -> bool

Indicates if the instruction has a side effect, ie. it can fail, or it interacts with memory.

pub(crate) fn replace_values(&mut self, mapping: &ValueMapping)

Replaces values present in this instruction with other values according to the given mapping.

pub(crate) fn map_values( &self, f: impl FnMut(ValueId) -> ValueId, ) -> Instruction

Maps each ValueId inside this instruction to a new ValueId, returning the new instruction. Note that the returned instruction is fresh and will not have an assigned InstructionId until it is manually inserted in a DataFlowGraph later.

pub(crate) fn map_values_mut(&mut self, f: impl FnMut(ValueId) -> ValueId)

Maps each ValueId inside this instruction to a new ValueId in place.

pub(crate) fn for_each_value<T>(&self, f: impl FnMut(ValueId) -> T)

Applies a function to each input value this instruction holds.