noirc_evaluator::ssa::ir::instruction
Enum Intrinsic

pub enum Intrinsic {

    ArrayLen,
    ArrayAsStrUnchecked,
    AsSlice,
    AssertConstant,
    StaticAssert,
    SlicePushBack,
    SlicePushFront,
    SlicePopBack,
    SlicePopFront,
    SliceInsert,
    SliceRemove,
    ApplyRangeConstraint,
    StrAsBytes,
    ToBits(Endian),
    ToRadix(Endian),
    BlackBox(BlackBoxFunc),
    Hint(Hint),
    AsWitness,
    IsUnconstrained,
    DerivePedersenGenerators,
    FieldLessThan,
    ArrayRefCount,
    SliceRefCount,

}

These are similar to built-ins in other languages. These can be classified under two categories:

Opcodes which the IR knows the target machine has special support for. (LowLevel)

Opcodes which have no function definition in the source code and must be processed by the IR.

Variants

ArrayLen

ArrayLen - returns the length of the input array argument: array (value id) result: length of the array, panic if the input is not an array

ArrayAsStrUnchecked

ArrayAsStrUnchecked - Converts a byte array of type [u8; N] to a string argument: array (value id) result: str

AsSlice

AsSlice argument: value id result: a slice containing the elements of the argument. Panic if the value id does not correspond to an array type

AssertConstant

AssertConstant - Enforce the argument to be a constant value, at compile time. argument: value id result: (), panic if the argument does not resolve to a constant value

StaticAssert

StaticAssert - Enforce the first argument to be true, at compile time arguments: boolean (value id), โฆmessage. The message can be a format string of several arguments result: (), panic if the arguments do not resolve to constant values or if the first one is false.

SlicePushBack

SlicePushBack - Add elements at the end of a slice arguments: slice length, slice contents, โฆelements_to_push result: a slice containing slice contents,..elements_to_push

SlicePushFront

SlicePushFront - Add elements at the start of a slice arguments: slice length, slice contents, โฆelements_to_push result: a slice containing ..elements_to_push, slice contents

SlicePopBack

SlicePopBack - Removes the last element of a slice arguments: slice length, slice contents result: a slice without the last element of slice contents

SlicePopFront

SlicePopFront - Removes the first element of a slice arguments: slice length, slice contents result: a slice without the first element of slice contents

SliceInsert

SliceInsert - Insert elements inside a slice. arguments: slice length, slice contents, insert index, โฆelements_to_insert result: a slice with โฆelements_to_insert inserted at the insert index

SliceRemove

SliceRemove - Removes an element from a slice arguments: slice length, slice contents, remove index result: a slice with without the element at remove index

ApplyRangeConstraint

ApplyRangeConstraint - Enforces the bit size of the first argument via a range check. arguments: value id, bit size (constant) result: applies a range check constraint to the input. It is replaced by a RangeCheck instruction during simplification.

StrAsBytes

StrAsBytes - Convert a str into a byte array of type [u8; N] arguments: value id result: the argument. Internally a str is a byte array.

ToBits(Endian)

ToBits(Endian) - Computes the bit decomposition of the argument. argument: a field element (value id) result: an array whose elements are the bit decomposition of the argument, in the endian order depending on the chosen variant. The type of the result gives the number of limbs to use for the decomposition.

ToRadix(Endian)

ToRadix(Endian) - Decompose the first argument over the radix base arguments: a field element (value id), the radix to use (constant, a power of 2 between 2 and 256) result: an array whose elements are the decomposition of the argument into the radix base, in the endian order depending on the chosen variant. The type of the result gives the number of limbs to use for the decomposition.

BlackBox(BlackBoxFunc)

BlackBox(BlackBoxFunc) - Calls a blackbox function. More details can be found here: [acvm-repo::acir::::circuit::opcodes::BlackBoxFuncCall]

Hint(Hint)

Hint(Hint) - Avoid its arguments to be removed by DIE. arguments: โฆ value id result: the arguments. Hint does not layout any constraint but avoid its arguments to be simplified out during SSA transformations

AsWitness

AsWitness - Adds a new witness constrained to be equal to the argument arguments: value id result: the argument

IsUnconstrained

IsUnconstrained - Indicates if the execution context is constrained or unconstrained argument: () result: true if execution is under unconstrained context, false else.

DerivePedersenGenerators

DerivePedersenGenerators - Computes the Pedersen generators arguments: domain_separator_string (constant string), starting_index (constant) result: array of elliptic curve points (Grumpkin) containing the generators. The type of the result gives the number of generators to compute.

FieldLessThan

FieldLessThan - Compare the arguments: lhs < rhs arguments: lhs, rhs. Field elements result: true if lhs mod p < rhs mod p (p being the field characteristic), false else

ArrayRefCount

ArrayRefCount - Gives the reference count of the array argument: array (value id) result: reference count of array. In unconstrained context, the reference count is stored alongside the array. in constrained context, it will be 0.

SliceRefCount

SliceRefCount - Gives the reference count of the slice arguments: slice length, slice contents (value id) result: reference count of slice. In unconstrained context, the reference count is stored alongside the slice. in constrained context, it will be 0.

Implementations

impl Intrinsic

pub(crate) fn has_side_effects(&self) -> bool

Returns whether the Intrinsic has side effects.

If there are no side effects then the Intrinsic can be removed if the result is unused.

An example of a side effect is increasing the reference count of an array, but functions which can fail due to implicit constraints are also considered to have a side effect.

pub(crate) fn purity(&self) -> Purity

pub(crate) fn lookup(name: &str) -> Option<Intrinsic>

Lookup an Intrinsic by name and return it if found. If there is no such intrinsic by that name, None is returned.