noirc_evaluator::ssa::ir

Module printer

This file is for pretty-printing the SSA IR in a human-readable form for debugging.

Structs

Printer

Functions

display_array_offset 🔒

display_binary 🔒

display_block 🔒

Display a single block. This will not display the block’s successors.

display_constrain_error 🔒

display_databus 🔒

display_function 🔒

Helper function for Function’s Display impl to pretty-print the function with the given formatter.

display_instruction 🔒

Display an arbitrary instruction

display_instruction_buffer 🔒

display_instruction_inner 🔒

display_terminator 🔒

Display a terminator instruction

is_printable_byte

number 🔒

Formats the given number assuming it has the given type. Unsigned types and field element types will be formatter as-is, while signed types will be formatted as positive or negative numbers depending on where the number falls in the range given by the type’s bit size.

result_types 🔒

try_byte_array_to_string 🔒

try_to_extract_string_from_error_payload 🔒

Tries to extract a constant string from an error payload.

value 🔒

Specialize displaying value ids so that if they refer to a numeric constant or a function we print those directly.

value_list 🔒

Display each value separated by a comma

value_list_with_types 🔒

Display each value along with its type. E.g. v0: Field, v1: u64, v2: u1

write_location_information 🔒