noirc_evaluator::ssa::opt

Module array_get

Replaces array_get with known indices with values from previous instructions such as array_set or make_array.

Given these two instructions:

v1 = array_set v0, index 0, value: 42
v2 = array_get v1, index 0 -> Field

because we get from v1 at index 0, but v1 is the result of setting the value “42” at index 0, we can conclude that v2 will be “42”, and so this SSA pass will do that. However, this is only safe to do if the array_set happened under the same side effects variable as the array_get. For example, in this case:

enable_side_effects v100
v1 = array_set v0, index 0, value: 42
enable_side_effects v200
v2 = array_get v1, index 0 -> Field

it would be wrong to replace v2 with “42” as the previous array_set might not have been executed.

In this case:

v1 = array_set v0, index 0, value: 42
v2 = array_set v1, index 1, value: 15
v3 = array_get v2, index 0 -> Field

for v3 the optimization will try to find a previous array_set with the same index (index 0). It will first find v2. Because it’s an array_set of a different known index, it will then find v1 and apply the same optimization as before. Note that it’s safe to skip v2 and look at v1 even if v2 was under a different side effects var because it doesn’t affect the index used in v3.

In this case:

v1 = array_set v0, index 0, value: 42
v2 = array_set v1, index v88, value: 15
v3 = array_get v2, index 0 -> Field

for v3 the optimization will find v2. Because the set index is unknown, and it might be zero, the optimization can’t deduce anything so it won’t do anything.

Another case where the optimization applies is when it finds a make_array:

v1 = make_array [Field 10, Field 20] : [Field; 2]
v2 = array_get v1, index 0 -> Field

In this case v2 will be replaced with Field 10. A make_array could also be reached after passing through other array_set with a different index, as previously shown.

Finally, the optimization might also reach to params:

b0(v1: [Field; 2]):
  v2 = array_set v1, index 1, value: 42
  v3 = array_get v2, index 0 -> Field

In this case v3 will be replaced with array_get v1, index 0, directly getting from v1 instead of from v2, because v2 is the same as v1 except for what’s in index 1, but v3 is getting from index 0.

This module also provides a try_optimize_array_get_from_previous_instructions function that is used in other SSA-related optimizations. For example, whenever an array_get is inserted into a DFG: in this case a previous array_set with the same index as the array_get cannot be used because we don’t know under which side effects var it happens. However, array_set with a different known index can be skipped through to eventually reach a make_array or param.

Structs

ArrayGetOptimizationSideEffects 🔒

Side effects information to be able to optimize array_get more efficiently.

Enums

ArrayGetOptimizationResult 🔒

The result of the array_get optimization.

Functions

try_optimize_array_get_from_previous_instructions 🔒

Tries to replace an array_get instructions with values from previous instructions. See the array_get module for more information.