noirc_evaluator::ssa::opt

Module array_set

Optimizes array_set instructions by turning then into make_array instructions when they set a known index on a previous make_array instruction, as long as both instructions are under the same side-effects predicate or the side effects var for the array_set is true.

If the predicates are different, but the element type is numeric, it will use the ValueMerger to merge the items.

For example, this:

v0 = make_array [Field 2, Field 3] : [Field; 2]
v1 = array_set v0, index u32 0, value Field 4

will change into this:

v0 = make_array [Field 2, Field 3] : [Field; 2]
v1 = make_array [Field 4, Field 3] : [Field; 2]

but this:

enable_side_effects v0
v1 = make_array [Field 2, Field 3] : [Field; 2]
enable_side_effects v2
v3 = array_set v0, index u32 0, value Field 4

will change into this:

enable_side_effects v0
v1 = make_array [Field 2, Field 3] : [Field; 2]
enable_side_effects v2
v3 = not v2
v4 = cast v2 as Field
v5 = cast v3 as Field
v6 = mul v4, Field 4
v7 = mul v5, Field 2
v8 = add v6, v7
v9 = make_array [v8, Field 3] : [Field; 2]

and this:

enable_side_effects v0
v1 = make_array [Field 1, Field 2] : [Field; 2]
v2 = make_array [Field 3, Field 4] : [Field; 2]
v3 = make_array [v1, v2] : [[Field; 2]; 2]
enable_side_effects v4
v5 = make_array [Field 5, Field 6] : [Field; 2]
v6 = array_set v3, index u32 0, value v5

will remain unchanged.

Functions

array_set_optimization_pre_check 🔒

Pre-check condition for Function::array_set_optimization.

fold_array_set_into_make_array 🔒

Decide whether we can turn an array_set into a make_array, returning: