noirc_evaluator::ssa::opt::array_set

Function fold_array_set_into_make_array

fn fold_array_set_into_make_array(
    dfg: &mut DataFlowGraph,
    block_id: BasicBlockId,
    call_stack_id: CallStackId,
    array_id: ValueId,
    value: ValueId,
    index: u32,
    side_effects_var: ValueId,
    make_array_predicates: &HashMap<Id<Instruction>, ValueId>,
) -> Option<(Vector<ValueId>, bool)>

Decide whether we can turn an array_set into a make_array, returning:

None to keep the array_set

Some(elements, insert_predicate) to replace it with a make_array with the updated elements;

If insert_predicate is true then we should keep tracking the side effect variable of the new make_array. Otherwise the elements are a result of merging the items at the index under different predicates, and the items in the make_array can be a mix of various side effects, and tracking must be stopped for it.