noirc_evaluator::ssa::opt
Module assert_constant

Functions

append_foreign_call_param 🔒

check_instruction 🔒

During the loop unrolling pass we also evaluate calls to assert_constant. This is done in this pass because loop unrolling is the only pass that will error if a value (the loop bounds) are not known constants.

evaluate_assert_constant 🔒

Evaluate a call to assert_constant, returning an error if any of the elements are not constants. If all of the elements are constants, Ok(false) is returned. This signifies a success but also that the instruction need not be reinserted into the block being unrolled since it has already been evaluated.

evaluate_static_assert 🔒

Evaluate a call to static_assert, returning an error if the value is false or not constant (see assert_constant).