noirc_evaluator::ssa::opt

Module basic_conditional

An SSA pass that operates on Brillig functions This optimization pass identifies simple conditional control flow patterns in unconstrained code and flattens them to reduce the number of basic blocks and improve performance.

e.g: if c {a} else {b} would be flattened to c*(a-b)+b A simple conditional pattern is defined as a conditional sub-graph of the form jmpif c: A, else B, where A and B are basic blocks which join on the same successor. This exclude the graph from having any nested conditional or loop statements. Performance improvement is based on a simple execution cost metric

This pass does not have any pre/post conditions.

Structs

BasicConditional 🔒

Functions

block_cost 🔒

Computes a cost estimate for the execution of a basic block returns u32::MAX if the block has side-effect instructions WARNING: these are estimates of the runtime cost of each instruction, 1 being the cost of the simplest instruction. These numbers can be improved.

flatten_function 🔒

Identifies all simple conditionals in the function and flattens them

flatten_multiple 🔒

Flattens multiple basic conditionals within a function.

is_conditional 🔒

Returns the blocks of the simple conditional sub-graph of the CFG whose input block is the entry. Returns None if the input block is not the entry block of a simple conditional. A simple conditional is an if-then(-else) statement where branches are ‘small’ basic blocks. ‘Small’ basic blocks means that we expect their execution cost to be small.