noirc_evaluator::ssa::opt
Module brillig_array_get_and_set

This SSA pass adjusts constant indexes of array operations inside Brillig functions to avoid performing an extra binary operation.

For example, if we have this SSA:

b0(v0: [Field; 10]):
  v1 = array_get v0, index u32 3 -> Field

Brillig would have to fetch the element at index 3 of the array. However, in the Brillig runtime arrays are represented as [RC, …items], where RC holds the reference count of the array. That means that the final index that needs to be retrieved is 4, not 3. With the above operation the final Brillig code would have to add 1 to 3 to get the desired element.

So, this pass will transform the above SSA into this:

b0(v0: [Field; 10]):
  v1 = array_get v0, index u32 4 minus 1 -> Field

The “minus 1” part is just there so that readers can understand that the index was offset and that the actual element index is 3. On the Brillig side, array operations with constant indexes are always assumed to have already been shifted.

Now the index to retrieve is 4 and there’s no need to offset it in Brillig, avoiding one addition.

In the case of slices, they are represented as Brillig vectors as [RC, Size, Capacity, …items], thus the items pointer instead starts at three rather than one. So for a slice this pass will transform this:

b0(v0: [Field]):
  v1 = array_get v0, index u32 3 -> Field

to this:

b0(v0: [Field]):
  v1 = array_get v0, index u32 6 minus 3 -> Field

This pass must be the very last, just before generating ACIR and Brillig opcodes from the SSA. Shifting indexes can break some of the assumptions of earlier compiler operations, so this is done outside the normal SSA pipeline, as an internal step, and must be run only once.

The main motivation behind this pass is to make it possible to reuse the same constants across Brillig opcodes without having to re-allocate another register for them every time they appear. For this to happen the constants need to be part of the DFG, where they can be hoisted and deduplicated across instructions. Doing this during Brillig codegen would be too late, as at that time we have a read-only DFG and we would be forced to generate more Brillig opcodes.

Functions

compute_offset_index 🔒

Given an array or slice value and a constant index, returns an offset (shifted) index.