noirc_evaluator::ssa::opt::brillig_entry_points
Function collect_callsites_to_rewrite

fn collect_callsites_to_rewrite(
    function: &Function,
    entry_point: Id<Function>,
    function_per_entry: &mut FxHashMap<Id<Function>, FxHashMap<Id<Function>, Id<Function>>>,
    calls_to_update: &FxHashMap<(Id<Function>, Id<Function>), Id<Function>>,
) -> FxHashSet<CallToUpdate>

Go through the supplied function and based upon the call sites set in the calls_to_update map build the set of call sites that should be rewritten. Upon finding call sites that should be rewritten this method will also update the mapping of old functions to new functions in the supplied NewCallSitesMap.