noirc_evaluator::ssa::opt

Module check_u128_mul_overflow

An SSA pass that operates on ACIR functions that checks that multiplying two u128 doesn’t overflow because both operands are greater or equal than 2^64. If both are, then the result is surely greater or equal than 2^128 so it would overflow. The operands can still overflow if just one of them is less than 2^64, but in that case the result will be less than 2^192 so it fits in a Field value, and acir will check that it fits in a u128.

In Brillig an overflow check is automatically performed on unsigned binary operations so this SSA pass has no effect for Brillig functions.

Statics

MAX_NON_OVERFLOWING_CONST_ARG 🔒

MAX_NON_OVERFLOWING_CONST_ARG is expected to be [p/U], where U=U128::max() and p is the field modulus.

Functions

check_u128_mul_overflow 🔒

max_non_overflowing_const_arg 🔒