noirc_evaluator::ssa::opt
Module constant_folding

The goal of the constant folding optimization pass is to propagate any constants forwards into later Instructions to maximize the impact of compile-time simplifications.

The pass works as follows:

Re-insert each instruction in order to apply the instruction simplification performed by the DataFlowGraph automatically as new instructions are pushed.

Check whether any input values have been constrained to be equal to a value of a simpler form by a constrain instruction. If so, replace the input value with the simpler form.

Check whether the instruction can_be_deduplicated by duplicate instruction earlier in the same block.

These operations are done in parallel so that they can each benefit from each other without the need for multiple passes.

This is the only pass which removes duplicated pure Instructions however and so is needed when different blocks are merged, i.e. after the flatten_cfg pass.

Structs

BrilligInfo 🔒

Context 🔒

ResultCache 🔒

Records the results of all duplicate Instructions along with the blocks in which they sit.

SimplificationCache 🔒

Records a simplified equivalents of an Instruction in the blocks where the constraint that advised the simplification has been encountered.

Enums

CacheResult 🔒

EvaluationResult 🔒

Result of trying to evaluate an instruction (any instruction) in this pass.

Functions

can_be_deduplicated 🔒

Indicates if the instruction can be safely replaced with the results of another instruction with the same inputs. If deduplicate_with_predicate is set, we assume we’re deduplicating with the instruction and its predicate, rather than just the instruction. Setting this means instructions that rely on predicates can be deduplicated as well.

simplify 🔒

Check if one expression is simpler than the other. Returns Some((complex, simple)) if a simplification was found, otherwise None. Expects the ValueIds to be fully resolved.

type_to_brillig_parameter 🔒

Similar to FunctionContext::ssa_type_to_parameter but never panics and disallows reference types.

value_id_to_calldata 🔒

Type Aliases

ConstraintSimplificationCache 🔒

HashMap from (side_effects_enabled_var, Instruction) to a simplified expression that it can be replaced with based on constraints that testify to their equivalence, stored together with the set of blocks at which this constraint has been observed.

InstructionResultCache 🔒

HashMap from (Instruction, side_effects_enabled_var) to the results of the instruction. Stored as a two-level map to avoid cloning Instructions during the .get call.