noirc_evaluator::ssa::opt::constant_folding
Function can_be_deduplicated

fn can_be_deduplicated(
    instruction: &Instruction,
    dfg: &DataFlowGraph,
) -> CanBeDeduplicated

Indicates if the instruction can be safely replaced with the results of another instruction with the same inputs. If deduplicate_with_predicate is set, we assume we’re deduplicating with the instruction and its predicate, rather than just the instruction. Setting this means instructions that rely on predicates can be deduplicated as well.

Some instructions get the predicate attached to their inputs by handle_instruction_side_effects in flatten_cfg. These can be deduplicated because they implicitly depend on the predicate, not only when the caller uses the predicate variable as a key to cache results. However, to avoid tight coupling between passes, we make the deduplication conditional on whether the caller wants the predicate to be taken into account or not.