noirc_evaluator::ssa::opt::constant_folding
Module interpret

Enums

EvaluationResult 🔒

Result of trying to evaluate an instruction (any instruction) in this pass.

Functions

const_ir_value_to_interpreter_value 🔒

Converts a constant SSA value into an interpreter value for execution.

evaluate_const_argument_call 🔒

Tries to evaluate an instruction if it’s a call where all its arguments are constant. We do this by interpreting the function’s SSA to calculate the result.

interpreter_value_to_ir_value 🔒

Converts a constant interpreter value back into an SSA constant.

try_interpret_call 🔒

Checks if the given instruction is a call to a function with all constant arguments. If so, we can try to evaluate that function and replace the results with the evaluation results.