noirc_evaluator::ssa::opt::constant_folding::interpret
Function try_interpret_call

pub(super) fn try_interpret_call(
    instruction: &Instruction,
    block: BasicBlockId,
    dfg: &mut DataFlowGraph,
    interpreter: Option<&mut Interpreter<'_, Empty>>,
) -> Option<Vec<ValueId>>

Checks if the given instruction is a call to a function with all constant arguments. If so, we can try to evaluate that function and replace the results with the evaluation results.