noirc_evaluator::ssa::opt::constant_folding

Module simplification_cache

Structs

ConstraintSimplificationCache 🔒

HashMap from (side_effects_enabled_var, Instruction) to a simplified expression that it can be replaced with based on constraints that testify to their equivalence, stored together with the set of blocks at which this constraint has been observed.

SimplificationCache 🔒

Records a simplified equivalents of an Instruction in the blocks where the constraint that advised the simplification has been encountered.

Functions

simplify 🔒

Check if one expression is simpler than the other. Returns Some((complex, simple)) if a simplification was found, otherwise None. Expects the ValueIds to be fully resolved.