noirc_evaluator::ssa::opt::constant_folding
Type Alias ConstraintSimplificationCache

type ConstraintSimplificationCache = FxHashMap<ValueId, FxHashMap<ValueId, SimplificationCache>>;

HashMap from (side_effects_enabled_var, Instruction) to a simplified expression that it can be replaced with based on constraints that testify to their equivalence, stored together with the set of blocks at which this constraint has been observed.

Only blocks dominated by one in the cache should have access to this information, otherwise we create a sort of time paradox where we replace an instruction with a constant we believe it should equal to, without ever actually producing and asserting the value.

Aliased Type

struct ConstraintSimplificationCache { /* private fields */ }