noirc_evaluator::ssa::opt::constant_folding
Type Alias InstructionResultCache

type InstructionResultCache = FxHashMap<Instruction, FxHashMap<Option<ValueId>, ResultCache>>;

HashMap from (Instruction, side_effects_enabled_var) to the results of the instruction. Stored as a two-level map to avoid cloning Instructions during the .get call.

The side_effects_enabled_var is optional because we only use them when Instruction::requires_acir_gen_predicate is true and the constraint information is also taken into account.

In addition to each result, the original BasicBlockId is stored as well. This allows us to deduplicate instructions across blocks as long as the new block dominates the original.

Aliased Type

struct InstructionResultCache { /* private fields */ }