noirc_evaluator::ssa::opt
Module defunctionalize

This module defines the defunctionalization pass for the SSA IR. The purpose of this pass is to transforms all functions used as values into constant numbers (fields) that represent the function id. That way all calls with a non-literal target can be replaced with a call to an apply function. The apply function is a dispatch function that takes the function id as a parameter and dispatches to the correct target.

Structs

ApplyFunction 🔒

Represents an ‘apply’ function created by this pass to dispatch higher order functions to. Pseudocode of an apply function is given below:

DefunctionalizationContext 🔒

Performs defunctionalization on all functions This is done by changing all functions as value to be a number (FieldElement) And creating apply functions that dispatch to the correct target by runtime comparisons with constants

Functions

build_return_block 🔒

If no previous return target exists, it will create a final return, otherwise returns the existing return block to jump to.

create_apply_function 🔒

Creates an apply function for the given signature and variants

create_apply_functions 🔒

find_dynamic_dispatches 🔒

Finds all dynamic dispatch signatures in the given function

find_functions_as_values 🔒

Finds all literal functions used as values in the given function

find_variants 🔒

Collects all functions used as values that can be called by their signatures

function_id_to_field 🔒

map_function_to_field 🔒

Try to map the given function literal to a field, returning Some(field) on success. Returns none if the given value was not a function or doesn’t need to be mapped.

remove_first_class_functions_in_instruction 🔒

Replace any first class functions used in an instruction with a field value. This applies to any function used anywhere else other than the function position of a call instruction. Returns true if the instruction was modified

Type Aliases

ApplyFunctions 🔒

Variants 🔒