noirc_evaluator::ssa::opt::defunctionalize
Function create_apply_functions

fn create_apply_functions(
    ssa: &mut Ssa,
    variants_map: BTreeMap<(Signature, RuntimeType), Vec<(Id<Function>, RuntimeType)>>,
) -> (FxHashMap<(Signature, RuntimeType), ApplyFunction>, FxHashMap<Id<Function>, Purity>)

Creates all apply functions needed for dispatch of function values.

This function maintains the grouping set in Variants, meaning an apply function is grouped by functions that share a signature and runtime. An apply function is only created if there are multiple function variants for a specific (Signature, RuntimeType) group. Otherwise, if there is a single variant that function is simply reused.

If there are no variants a dummy function is created. A dummy function acts as a safe no-op to continue compilation even though there are no variants for a first-class function call. For more information you can reference create_dummy_function.

Arguments

ssa: A mutable reference to the full Ssa structure containing all functions.

variants_map: Variants

Returns

ApplyFunctions keyed by each function’s signature before functions are changed into field types. The inner apply function itself will have its defunctionalized type, with function values represented as field values.

[HashMap<FunctionId, Purity>] with purities that must be set to all functions in the SSA, as this function might have created dummy pure functions.