noirc_evaluator::ssa::opt::defunctionalize

Function find_variants

fn find_variants(
    ssa: &Ssa,
) -> BTreeMap<(Signature, RuntimeType), Vec<(Id<Function>, RuntimeType)>>

Collects all functions used as values that can be called by their signatures.

Groups all FunctionIds used as values by their Signature and caller RuntimeType, producing a mapping from these tuples to the list of target functions that may be dynamically dispatched.

Arguments

ssa: The full Ssa structure

Returns

Variants that should then be used to generate apply functions for dispatching