noirc_evaluator::ssa::opt::die

Module array_oob_checks

Functions

apply_side_effects 🔒

Given lhs and rhs values, if there’s a side effects condition this will return (lhs * condition, rhs * condition), otherwise just (lhs, rhs)

handle_array_get_group 🔒

Handle the case when an ArrayGet is potentially out-of-bounds and the array contains composite types by figuring out whether all ArrayGet of different parts of the complex item are unused, and if so then insert a single constraint to replace all of them.

should_insert_oob_check 🔒

Array get/set has explicit out of bounds (OOB) checks laid down in the Brillig runtime. These checks are not laid down in the ACIR runtime as that runtime maps the SSA to a memory model where OOB accesses will be prevented. Essentially all array ops in ACIR will have a side effect where they check for the index being OOB. However, in order to maintain parity between the Brillig and ACIR runtimes, if we have an unused array operation we need insert an OOB check so that the side effects ordering remains correct.