noirc_evaluator::ssa::opt::die
Module prune_dead_parameters

This module implements the pruning of unused block parameters from functions.

Blocks can accept parameters that are passed via terminator instructions (e.g., jmp). During the compilation pipeline, it’s common for some of these parameters to become unused. This module eliminates those unused parameters and correspondingly adjusts the terminators of predecessor blocks. This work is ultimately done to reduce code size.

How this pass works:

Iterates through all blocks in post-order (to ensure predecessors are visited after successors).

Detects and removes unused block parameters, except for those on entry blocks of program entry point functions (e.g., main, foldable functions). Brillig entry points (except main) are not counted here as their signature does not determine the ABI and is safe to update.

Clears the list of unused block parameters after removing them from the block.

Updates the corresponding argument lists in predecessor terminator instructions to keep them aligned with the new parameter lists.

Entry block parameters of non-program entry point functions may be pruned if they are unused, and the corresponding call instructions are rewritten to remove the dead arguments.

Entry point functions often correspond to ABI inputs and must remain to preserve the program’s external interface.

Preconditions:

This pass should be run after Dead Instruction Elimination (DIE) so that parameter liveness is up-to-date.

Panics

Return blocks are not expected to have successors, so encountering one as a predecessor is treated as an internal compiler error (ICE).

Example:

Before pruning:

b0():
  jmp b1(Field 1, Field 2, Field 3)
b1(v0: Field, v1: Field, v2: Field):
  return v1

After pruning:

b0():
  jmp b1(Field 2)
b1(v0: Field):
  return v0

Entry block callsite update example

Original function f1:

brillig(inline) fn f1:
  b0(v0: Field, v1: Field):
    jmp b1(Field 1)
  b1(v2: Field):
    return v2

Call to f1 in another function:

v0 = call f1(Field 5, Field 10) -> Field

After pruning::

brillig(inline) fn f1:
  b0():
    jmp b1(Field 1)
  b1(v2: Field):
    return v2

Correspondingly, the callsite is rewritten to remove the now unused arguments:

v0 = call f1() -> Field