noirc_evaluator::ssa::opt::die
Struct Context

struct Context {
    used_values: FxHashSet<ValueId>,
    instructions_to_remove: FxHashSet<Id<Instruction>>,
    rc_instructions: Vec<(Id<Instruction>, BasicBlockId)>,
    flattened: bool,
    parameter_keep_list: FxHashMap<BasicBlockId, Vec<bool>>,
}

Per function context for tracking unused values and which instructions to remove.

Fields
used_values: FxHashSet<ValueId>instructions_to_remove: FxHashSet<Id<Instruction>>rc_instructions: Vec<(Id<Instruction>, BasicBlockId)>

IncrementRc & DecrementRc instructions must be revisited after the main DIE pass since they technically contain side-effects but we still want to remove them if their value parameter is not used elsewhere.

flattened: bool

The elimination of certain unused instructions assumes that the DIE pass runs after the flattening of the CFG, but if that’s not the case then we should not eliminate them just yet.

parameter_keep_list: FxHashMap<BasicBlockId, Vec<bool>>

A per-block list indicating which block parameters are still considered alive.

Each entry maps a BasicBlockId to a Vec<bool>, where the ith boolean corresponds to the ith parameter of that block. A value of true means the parameter is used and should be preserved. A value of false means it is unused and can be pruned.

This keep list is used during terminator analysis to avoid incorrectly marking values as used simply because they appear as terminator arguments. Only parameters marked as live here should result in values being marked as used in terminator arguments.

Implementations

impl Context

pub(super) fn replace_array_instructions_with_out_of_bounds_checks( &mut self, function: &mut Function, block_id: BasicBlockId, possible_index_out_of_bounds_indexes: &mut Vec<usize>, ) -> bool

Replaces unused ArrayGet/ArraySet instructions with out of bounds checks. Returns true if at least one check was inserted. Because some ArrayGet might happen in groups (for composite types), if just some of the instructions in a group are used but not all of them, no check is inserted, so this method might return false.

impl Context

fn remove_unused_instructions_in_block( &mut self, function: &mut Function, block_id: BasicBlockId, insert_out_of_bounds_checks: bool, ) -> bool

Steps backwards through the instruction of the given block, amassing a set of used values as it goes, and at the same time marking instructions for removal if they haven’t appeared in the set thus far.

It is not only safe to mark instructions for removal as we go because no instruction result value can be referenced before the occurrence of the instruction that produced it, and we are iterating backwards. It is also important to identify instructions that can be removed as we go, such that we know not to include its referenced values in the used values set. This allows DIE to identify whole chains of unused instructions. (If the values referenced by an unused instruction were considered to be used, only the head of such chains would be removed.)

If insert_out_of_bounds_checks is true and there are unused ArrayGet/ArraySet that might be out of bounds, this method will insert out of bounds checks instead of removing unused instructions and return true. The idea then is to later call this function again with insert_out_of_bounds_checks set to false to effectively remove unused instructions but leave the out of bounds checks.

fn is_unused( &self, instruction_id: Id<Instruction>, function: &Function, ) -> bool

Returns true if an instruction can be removed.

An instruction can be removed as long as it has no side-effects, and none of its result values have been referenced.

fn mark_terminator_values_as_used( &mut self, function: &Function, block: &BasicBlock, )

Adds values referenced by the terminator to the set of used values.

fn mark_used_instruction_results( &mut self, dfg: &DataFlowGraph, value_id: ValueId, )

Inspects a value and marks all instruction results as used.

fn remove_rc_instructions(&self, dfg: &mut DataFlowGraph)

Go through the RC instructions collected when we figured out which values were unused; for each RC that refers to an unused value, remove the RC as well.

fn is_inc_dec_instruction_on_known_array( instruction: &Instruction, dfg: &DataFlowGraph, ) -> bool

True if this is a Instruction::IncrementRc or Instruction::DecrementRc operating on an array directly from a Instruction::MakeArray or an intrinsic known to return a fresh array.