noirc_evaluator::ssa::opt
Module evaluate_static_assert_and_assert_constant

A simple SSA pass to go through each instruction and evaluate:

each call to assert_constant, issuing an error if any arguments to the function are not constants

each call to static_assert, issuing an error if the assertion does not hold, the value to test is not a constant, or the message is dynamic.

Note that this pass must be placed directly before loop unrolling to be useful. Any optimization passes between this and loop unrolling will cause the constants that this pass sees to be potentially different than the constants seen by loop unrolling. Furthermore, this pass cannot be a part of loop unrolling since we must go through every instruction to find all references to assert_constant while loop unrolling only touches blocks with loops in them.

Functions

append_foreign_call_param 🔒

check_instruction 🔒

During the loop unrolling pass we also evaluate calls to assert_constant. This is done in this pass because loop unrolling is the only pass that will error if a value (the loop bounds) are not known constants.

evaluate_assert_constant 🔒

Evaluate a call to assert_constant, returning an error if any of the elements are not constants. If all of the elements are constants, Ok(false) is returned. This signifies a success but also that the instruction need not be reinserted into the block being unrolled since it has already been evaluated.

evaluate_static_assert 🔒

Evaluate a call to static_assert, returning an error if the value is false or not constant (see assert_constant).

get_blocks_within_empty_loop 🔒

Returns all of a function’s block that are part of empty loops.