noirc_evaluator::ssa::opt

Module expand_signed_math

An SSA pass for ACIR functions that transforms “less than”, “div” and “mod” operation on signed integers into equivalent sequences of operations that rely on unsigned integers.

The purpose of this pass is to avoid ACIR having to handle signed integers “less than”, “div” and “mod” operations (for simplicity), while also allowing further optimizations to be done during subsequent SSA passes on the expanded instructions.

Structs

Context 🔒

Functions

expand_signed_math_post_check 🔒

Post-check condition for Function::expand_signed_math.