noirc_evaluator::ssa::opt
Module flatten_cfg

The flatten cfg optimization pass “flattens” the entire control flow graph into a single block. This includes branches in the CFG with non-constant conditions. Flattening these requires special handling for operations with side-effects and can lead to a loss of information since the jmpif will no longer be in the program. As a result, this pass should usually be towards or at the end of the optimization passes. Furthermore, this pass assumes that no loops are present in the program and will assume that a jmpif is a branch point and will attempt to merge both blocks. No actual looping will occur.

This pass is also known to produce some extra instructions which may go unused (usually ‘Not’) while merging branches. These extra instructions can be cleaned up by a later dead instruction elimination (DIE) pass.

Though CFG information is lost during this pass, some key information is retained in the form of EnableSideEffectsIf instructions. Each time the flattening pass enters and exits a branch of a jmpif, an instruction is inserted to capture a condition that is analogous to the activeness of the program point. For example:

b0(v0: u1): jmpif v0, then: b1, else: b2 b1(): v1 = call f0 jmp b3(v1) … blocks b2 & b3 …

Would brace the call instruction as such: enable_side_effects v0 v1 = call f0 enable_side_effects u1 1

(Note: we restore to “true” to indicate that this program point is not nested within any other branches.)

When we are flattening a block that was reached via a jmpif with a non-constant condition c, the following transformations of certain instructions within the block are expected:

A constraint is multiplied by the condition and changes the constraint to an equality with c:

constrain v0

v1 = mul v0, c v2 = eq v1, c constrain v2

If we reach the end block of the branch created by the jmpif instruction, its block parameters will be merged. To merge the jmp arguments of the then and else branches, the formula c * then_arg + !c * else_arg is used for each argument.

b0(v0: u1, v1: Field, v2: Field): jmpif v0, then: b1, else: b2 b1(): jmp b3(v1) b2(): jmp b3(v2) b3(v3: Field): … b3 instructions …

b0(v0: u1, v1: Field, v2: Field): v3 = mul v0, v1 v4 = not v0 v5 = mul v4, v2 v6 = add v3, v5 … b3 instructions …

After being stored to in at least one predecessor of a block with multiple predecessors, the value of a memory address is the value it had in both branches combined via c * a + !c * b. Note that the following example is simplified to remove extra load instructions and combine the separate merged stores for each branch into one store. See the next example for a non-simplified version with address offsets.

b0(v0: u1): v1 = allocate 1 Field jmpif v0, then: b1, else: b2 b1(): store v1, Field 5 … b1 instructions … jmp b3 b2(): store v1, Field 7 … b2 instructions … jmp b3 b3(): … b3 instructions …

b0(): v1 = allocate 1 Field store v1, Field 5 … b1 instructions … store v1, Field 7 … b2 instructions … v2 = mul v0, Field 5 v3 = not v0 v4 = mul v3, Field 7 v5 = add v2, v4 store v1, v5 … b3 instructions …

Note that if the ValueId of the address stored to is not the same, two merging store instructions will be made - one to each address. This is the case even if both addresses refer to the same address internally. This can happen when they are equivalent offsets:

b0(v0: u1, v1: ref) jmpif v0, then: b1, else: b2 b1(): v2 = add v1, Field 1 store Field 11 in v2 … b1 instructions … b2(): v3 = add v1, Field 1 store Field 12 in v3 … b2 instructions …

In this example, both store instructions store to an offset of 1 from v1, but because the ValueIds differ (v2 and v3), two store instructions will be created:

b0(v0: u1, v1: ref) v2 = add v1, Field 1 v3 = load v2 (new load) store Field 11 in v2 … b1 instructions … v4 = not v0 (new not) v5 = add v1, Field 1 v6 = load v5 (new load) store Field 12 in v5 … b2 instructions … v7 = mul v0, Field 11 v8 = mul v4, v3 v9 = add v7, v8 store v9 at v2 (new store) v10 = mul v0, v6 v11 = mul v4, Field 12 v12 = add v10, v11 store v12 at v5 (new store)

Modules

branch_analysis 🔒

This is an algorithm for identifying branch starts and ends.

value_merger 🔒

Structs

ConditionalBranch 🔒

ConditionalContext 🔒

Context 🔒

Functions

flatten_cfg_post_check 🔒

Post-check condition for Ssa::flatten_cfg.

flatten_cfg_pre_check 🔒

Pre-check condition for Ssa::flatten_cfg.

flatten_function_cfg 🔒

Flattens the control flow graph of the function such that it is left with a single block containing all instructions and no more control-flow.