noirc_evaluator::ssa::opt::flatten_cfg::branch_analysis
Struct Context

struct Context<'cfg> {
    branch_ends: FxHashMap<BasicBlockId, BasicBlockId>,
    branch_ends_pending: FxHashMap<BasicBlockId, BasicBlockId>,
    branch_parents: FxHashMap<BasicBlockId, BasicBlockId>,
    stack: Vec<(BasicBlockId, BasicBlockId, BasicBlockId)>,
    cfg: &'cfg ControlFlowGraph,
}

Fields
branch_ends: FxHashMap<BasicBlockId, BasicBlockId>branch_ends_pending: FxHashMap<BasicBlockId, BasicBlockId>branch_parents: FxHashMap<BasicBlockId, BasicBlockId>stack: Vec<(BasicBlockId, BasicBlockId, BasicBlockId)>cfg: &'cfg ControlFlowGraph
Implementations

impl<'cfg> Context<'cfg>

fn new(cfg: &'cfg ControlFlowGraph) -> Self

fn find_join_point_of_branches( &mut self, start: BasicBlockId, successors: impl Iterator<Item = BasicBlockId>, ) -> BasicBlockId

fn complete_parents(&mut self, branch: BasicBlockId, join: BasicBlockId)

Try to complete as many of the parent levels as we can, after finding the end point of a branch.

Say we have this CFG:

                   b11
                  /   \
      b5        b9     b13
     /  \      /  \   /   \
   b1    b7--b8    b12     b15
   / \  /      \          /   \
 b0   b6        b10----b14    b4
   \                          /
   b2-----------------------b3

At some point during the algorithm:

We process b8 we see that on the b10 side it’s followed by the join point b15 and on the b9 side it’s another branch.

We take note of b15 as the pending ending for b8, and queue b9 for processing.

When we process b9, we get the join point b13 on both b11 and b12, so we “immediately” know that b9 ends at b13.

At that point we look for the next point after b13, which is the join point b15.

The parent of b9 is b8, which already has b15 as a pending end, which checks out, so we can “complete” b8.

Then we follow up the chain: the parent level of b8 is b0; b15 is followed by b4; b0 already has b4 as pending end, so we can “complete” that as well.

fn push_branch( &mut self, parent: BasicBlockId, branch: BasicBlockId, left: BasicBlockId, right: BasicBlockId, )

Push a branch to the stack for exploration, remembering what the parent level branch was.

fn maybe_join( &mut self, branch: BasicBlockId, left: &Point, right: &Point, ) -> Option<BasicBlockId>

Check if the left and right branches joined. If so, mark the branch as completed, and return Some join point, otherwise return None. Panics if both points are joins, but have different values.

fn maybe_join_pending( &mut self, parent: BasicBlockId, join: BasicBlockId, ) -> bool

Try to join a pending branch, once the next join point is found:

if this is the first time we encounter this, mark it as pending, and return false,

if this is the second time, then we mark it as completed, and return true.

Panics if the join point does not match the existing one.

fn must_join( &mut self, start: BasicBlockId, left: BasicBlockId, right: BasicBlockId, )

Check that the left and right join points are the same, and mark this as the join point for the start block.

fn find_next_point(&self, block: BasicBlockId, skip: bool) -> Point

Starting with the current block, find the next join or branching point, skipping over blocks with single successors and predecessors.

If a block is both a join and a branch block, it is returned as a join.

When skip is true, we don’t consider the current block for an immediate join point, although it can be returned as a branch.