noirc_evaluator::ssa::opt::flatten_cfg
Struct Context

pub(crate) struct Context<'f> {
    pub(crate) inserter: FunctionInserter<'f>,
    cfg: ControlFlowGraph,
    pub(crate) target_block: Id<BasicBlock>,
    branch_ends: FxHashMap<Id<BasicBlock>, Id<BasicBlock>>,
    condition_stack: Vec<ConditionalContext>,
    arguments_stack: Vec<Vec<Id<Value>>>,
    local_allocations: FxHashSet<Id<Value>>,
    not_instructions: FxHashMap<Id<Value>, Id<Value>>,
    pub(crate) no_predicate: bool,
}

Fields
inserter: FunctionInserter<'f>cfg: ControlFlowGraph

This ControlFlowGraph is the graph from before the function was modified by this flattening pass.

target_block: Id<BasicBlock>

Target block of the flattening

branch_ends: FxHashMap<Id<BasicBlock>, Id<BasicBlock>>

Maps start of branch -> end of branch

condition_stack: Vec<ConditionalContext>

A stack of each jmpif condition that was taken to reach a particular point in the program. When two branches are merged back into one, this constitutes a join point, and is analogous to the rest of the program after an if statement. When such a join point / end block is found, the top of this conditions stack is popped since we are no longer under that condition. If we are under multiple conditions (a nested if), the topmost condition is the most recent condition combined with all previous conditions via And instructions.

arguments_stack: Vec<Vec<Id<Value>>>

Stack of block arguments When processing a block, we pop this stack to get its arguments and at the end we push the arguments for his successor

local_allocations: FxHashSet<Id<Value>>

Stores all allocations local to the current branch.

Since these branches are local to the current branch (i.e. only defined within one branch of an if expression), they should not be merged with their previous value or stored value in the other branch since there is no such value.

The ValueId here is that which is returned by the allocate instruction.

not_instructions: FxHashMap<Id<Value>, Id<Value>>

A map from cond to Not(cond)

Not instructions are inserted constantly by this pass and this map helps keep us from unnecessarily inserting extra instructions, and keeps ids unique which helps simplifications.

no_predicate: bool

Flag to tell the context to not issue ‘enable_side_effect’ instructions during flattening. This should be set to true only by flatten_single(), when no instruction is known to fail.

Implementations

impl Context<'_>

fn flatten_single_conditional( &mut self, conditional: &BasicConditional, no_predicates: &mut HashMap<Id<Function>, bool>, )

fn map_block_with_mapping( mapping: HashMap<Id<Value>, Id<Value>>, func: &mut Function, block: Id<BasicBlock>, )

impl<'f> Context<'f>

pub(crate) fn new( function: &'f mut Function, cfg: ControlFlowGraph, branch_ends: HashMap<Id<BasicBlock>, Id<BasicBlock>>, target_block: Id<BasicBlock>, ) -> Self

pub(crate) fn flatten(&mut self, no_predicates: &HashMap<Id<Function>, bool>)

fn link_condition(&mut self, condition: Id<Value>) -> Id<Value>

Returns the updated condition so that it is ‘AND-ed’ with the previous condition (if any)

fn get_last_condition(&self) -> Option<Id<Value>>

Returns the current condition

fn is_no_predicate( &self, no_predicates: &HashMap<Id<Function>, bool>, instruction: &Id<Instruction>, ) -> bool

Use the provided map to say if the instruction is a call to a no_predicates function

pub(crate) fn inline_block( &mut self, block: Id<BasicBlock>, no_predicates: &HashMap<Id<Function>, bool>, )

pub(crate) fn handle_terminator( &mut self, block: Id<BasicBlock>, work_list: &[Id<BasicBlock>], ) -> Vec<Id<BasicBlock>>

Returns the list of blocks that need to be processed after the given block For a normal block, it would be its successor For blocks related to a conditional statement, we ensure to process the ‘then-branch’, then the ‘else-branch’ (if it exists), and finally the end block

fn if_start( &mut self, condition: &Id<Value>, then_destination: &Id<BasicBlock>, else_destination: &Id<BasicBlock>, if_entry: &Id<BasicBlock>, call_stack: CallStackId, ) -> Vec<Id<BasicBlock>>

Process a conditional statement

fn then_stop(&mut self, block: &Id<BasicBlock>) -> Vec<Id<BasicBlock>>

Switch context to the ‘else-branch’

fn not_instruction( &mut self, condition: Id<Value>, call_stack: CallStackId, ) -> Id<Value>

fn else_stop(&mut self, block: &Id<BasicBlock>) -> Vec<Id<BasicBlock>>

Process the ‘exit’ block of a conditional statement

fn inline_branch_end( &mut self, destination: Id<BasicBlock>, cond_context: ConditionalContext, ) -> Id<BasicBlock>

Inline the ending block of a branch, the point where all blocks from a jmpif instruction join back together. In particular this function must handle merging block arguments from all of the join point’s predecessors, and it must handle any differing side effects from each branch.

Afterwards, continues inlining recursively until it finds the next end block or finds the end of the function.

Returns the final block that was inlined.

fn insert_instruction( &mut self, instruction: Instruction, call_stack: CallStackId, ) -> Id<Value>

Insert a new instruction into the target block. Unlike push_instruction, this function will not map any ValueIds. within the given instruction, nor will it modify self.values in any way.

fn insert_instruction_with_typevars( &mut self, instruction: Instruction, ctrl_typevars: Option<Vec<Type>>, call_stack: CallStackId, ) -> InsertInstructionResult<'_>

Inserts a new instruction into the target block, using the given control type variables to specify result types if needed. Unlike push_instruction, this function will not map any ValueIds. within the given instruction, nor will it modify self.values in any way.

fn insert_current_side_effects_enabled(&mut self)

Checks the branch condition on the top of the stack and uses it to build and insert an EnableSideEffectsIf instruction into the target block.

If the stack is empty, a “true” u1 constant is taken to be the active condition. This is necessary for re-enabling side-effects when re-emerging to a branch depth of 0.

fn push_instruction(&mut self, id: Id<Instruction>)

Push the given instruction to the end of the target block of the current function.

Note that each ValueId of the instruction will be mapped via self.inserter.resolve. As a result, the instruction that will be pushed will actually be a new instruction with a different InstructionId from the original. The results of the given instruction will also be mapped to the results of the new instruction.

previous_allocate_result should only be set to the result of an allocate instruction if that instruction was the instruction immediately previous to this one - if there are any instructions in between it should be None.

fn handle_instruction_side_effects( &mut self, instruction: Instruction, call_stack: CallStackId, ) -> Instruction

If we are currently in a branch, we need to modify constrain instructions to multiply them by the branch’s condition (see optimization #1 in the module comment).

fn cast_condition_to_value_type( &mut self, condition: Id<Value>, value: Id<Value>, call_stack: CallStackId, ) -> Id<Value>

fn mul_by_condition( &mut self, value: Id<Value>, condition: Id<Value>, call_stack: CallStackId, ) -> Id<Value>

fn apply_predicate_to_msm_argument( &mut self, argument: Id<Value>, predicate: Id<Value>, call_stack: CallStackId, ) -> (Vector<Id<Value>>, Type)

When a MSM is done under a predicate, we need to apply the predicate to the is_infinity property of the input points in order to ensure that the points will be on the curve no matter what.

fn var_or_one( &mut self, var: Id<Value>, condition: Id<Value>, call_stack: CallStackId, ) -> Id<Value>