noirc_evaluator::ssa::opt

Module inlining

This module defines the function inlining pass for the SSA IR. The purpose of this pass is to inline the instructions of each function call within the function caller. If all function calls are known, there will only be a single function remaining when the pass finishes.

Re-exports

pub use inline_info::MAX_SIMPLE_FUNCTION_WEIGHT;

Modules

inline_info 🔒

Structs

InlineContext 🔒

The context for the function inlining pass.

PerFunctionContext 🔒

The per-function inlining context contains information that is only valid for one function. For example, each function has its own DataFlowGraph, and thus each function needs a translation layer to translate between BlockId to BlockId for the current function and the function to inline into. The same goes for ValueIds, InstructionIds, and for storing other data like parameter to argument mappings.

Constants

RECURSION_LIMIT 🔒

An arbitrary limit to the maximum number of recursive call frames at any point in time.