noirc_evaluator::ssa::opt::inlining::inline_info
Function compute_function_should_be_inlined

fn compute_function_should_be_inlined(
    ssa: &Ssa,
    inline_no_predicates_functions: bool,
    small_function_max_instructions: i64,
    aggressiveness: i64,
    times_called: &HashMap<Id<Function>, usize>,
    inline_infos: &mut BTreeMap<Id<Function>, InlineInfo>,
    call_graph: &CallGraph,
    index: NodeIndex,
)

Determines whether a function should be inlined.

Inlining is determined by the following:

the function is not recursive

the cost of inlining outweighs the cost of not doing so

The total weight of a function and its cost are computed in this method. The total weight is calculated by taking the function’s own weight and multiplying it by the weight of each callee. We then determine the cost of inlining to be the times a function has been called multiplied by its total weight.

To determine the cost of retaining a function we first need the function interface cost, computed in compute_function_interface_cost. The cost of retaining of a function is then (times a function has been called) * (interface cost) + total weight.

A function’s net cost is then (cost of inlining - cost of retaining). The net cost is then compared against the inliner aggressiveness setting. If the net cost is less than the aggressiveness, we inline the function (granted there are not other restrictions such as recursion).