noirc_evaluator::ssa::opt::inlining::inline_info

Function compute_inline_infos

pub(crate) fn compute_inline_infos(
    ssa: &Ssa,
    call_graph: &CallGraph,
    inline_no_predicates_functions: bool,
    small_function_max_instructions: usize,
    aggressiveness: i64,
) -> BTreeMap<Id<Function>, InlineInfo>

The functions we should inline into (and that should be left in the final program) are:

main

Any Brillig function called from Acir

Some Brillig functions depending on aggressiveness and some metrics

Any Acir functions with a [fold inline type][noirc_frontend::monomorphization::ast::InlineType],

The returned InlineInfos won’t have every function in it, only the ones which the algorithm visited.