noirc_evaluator::ssa::opt::inlining
Struct PerFunctionContext

struct PerFunctionContext<'function> {
    entry_function: &'function Function,
    source_function: &'function Function,
    context: &'function mut InlineContext,
    values: HashMap<ValueId, ValueId>,
    blocks: HashMap<BasicBlockId, BasicBlockId>,
    inlining_entry: bool,
}

The per-function inlining context contains information that is only valid for one function. For example, each function has its own DataFlowGraph, and thus each function needs a translation layer to translate between BlockId to BlockId for the current function and the function to inline into. The same goes for ValueIds, InstructionIds, and for storing other data like parameter to argument mappings.

Fields
entry_function: &'function Function

The function that we are inlining calls into.

source_function: &'function Function

The source function is the function we’re currently inlining into the function being built.

context: &'function mut InlineContext

The shared inlining context for all functions. This notably contains the FunctionBuilder used to build the function we’re inlining into.

values: HashMap<ValueId, ValueId>

Maps ValueIds in the function being inlined to the new ValueIds to use in the function being inlined into. This mapping also contains the mapping from parameter values to argument values.

blocks: HashMap<BasicBlockId, BasicBlockId>

Maps blocks in the source function to blocks in the function being inlined into, where each mapping is from the start of a source block to an inlined block in which the analogous program point occurs.

Note that the starts of multiple source blocks can map into a single inlined block. Conversely the whole of a source block is not guaranteed to map into a single inlined block.

inlining_entry: bool

True if we’re currently working on the entry point function.

Implementations

impl<'function> PerFunctionContext<'function>

fn new( context: &'function mut InlineContext, entry_function: &'function Function, source_function: &'function Function, ) -> Self

Create a new PerFunctionContext from the source function. The value and block mappings for this context are initially empty except for containing the mapping between parameters in the source_function and the arguments of the destination function.

fn translate_value(&mut self, id: ValueId) -> ValueId

Translates a ValueId from the function being inlined to a ValueId of the function being inlined into. Note that this expects value ids for all Value::Instruction and Value::Param values are already handled as a result of previous inlining of instructions and blocks respectively. If these assertions trigger it means a value is being used before the instruction or block that defines the value is inserted.

fn translate_block( &mut self, source_block: BasicBlockId, block_queue: &mut VisitOnceDeque, ) -> BasicBlockId

Translates the program point representing the start of the given source_block to the inlined block in which the analogous program point occurs. (Once inlined, the source block’s analogous program region may span multiple inlined blocks.)

If the block isn’t already known, this will insert a new block into the target function with the same parameter types as the source block.

fn translate_block_parameters( &mut self, source_block: BasicBlockId, target_block: BasicBlockId, )

Copy block parameters from source_block into target_block.

fn get_function(&mut self, id: ValueId) -> Option<Id<Function>>

Try to retrieve the function referred to by the given Id. Expects that the given ValueId belongs to the source_function.

Returns None if the id is not known to refer to a function.

fn inline_blocks( &mut self, ssa: &Ssa, should_inline_call: &impl Fn(&Function) -> bool, ) -> Result<Vec<ValueId>, RuntimeError>

Inline all reachable blocks within the source_function into the destination function.

fn handle_function_returns( &mut self, returns: Vec<(BasicBlockId, Vec<ValueId>)>, ) -> Vec<ValueId>

Handle inlining a function’s possibly multiple return instructions. If there is only 1 return we can just continue inserting into that block. If there are multiple, we’ll need to create a join block to jump to with each value.

fn inline_block_instructions( &mut self, ssa: &Ssa, block_id: BasicBlockId, should_inline_call: &impl Fn(&Function) -> bool, ) -> Result<(), RuntimeError>

Inline each instruction in the given block into the function being inlined into. This may recurse if it finds another function to inline if a call instruction is within this block.

fn validate_callee( &self, callee: &Function, call_stack: Vec<Location>, ) -> Result<(), RuntimeError>

Extra error check where given a caller’s runtime its callee runtime is valid. We determine validity as the following (where we have caller -> callee). Valid:

ACIR -> ACIR

ACIR -> Brillig

Brillig -> Brillig

Invalid:

Brillig -> ACIR

Whether a valid callee should be inlined is determined separately by the inline info computation.

fn inline_function( &mut self, ssa: &Ssa, call_id: Id<Instruction>, function: Id<Function>, arguments: &[ValueId], should_inline_call: &impl Fn(&Function) -> bool, ) -> Result<(), RuntimeError>

Inline a function call and remember the inlined return values in the values map

fn push_instruction(&mut self, id: Id<Instruction>)

Push the given instruction from the source_function into the current block of the function being inlined into.

fn insert_new_instruction_results( values: &mut HashMap<ValueId, ValueId>, old_results: &[ValueId], new_results: InsertInstructionResult<'_>, )

Modify the values HashMap to remember the mapping between an instruction result’s previous ValueId (from the source_function) and its new ValueId in the destination function.

fn handle_terminator_instruction( &mut self, block_id: BasicBlockId, block_queue: &mut VisitOnceDeque, ) -> Option<(BasicBlockId, Vec<ValueId>)>

Handle the given terminator instruction from the given source function block. This will push any new blocks to the destination function as needed, add them to the block queue, and set the terminator instruction for the current block.

If the terminator instruction was a Return, this will return the block this instruction was in as well as the values that were returned.

fn extend_call_stack(&mut self, call_stack: CallStackId)