noirc_evaluator::ssa::opt
Module loop_invariant

The loop invariant code motion (LICM) pass moves code from inside a loop to before the loop if that code will always have the same result on every iteration of the loop.

To identify a loop invariant, check whether all of an instruction’s values are:

Outside of the loop

Constant

Already marked as loop invariants

We also check that we are not hoisting instructions with side effects. However, there are certain instructions whose side effects are only activated under a predicate (e.g. an array out of bounds error on a dynamic index). Thus, we also track the control dependence of loop blocks to determine whether these “pure with predicate instructions” can be hoisted. We use post-dominance frontiers to determine control dependence.

Let’s look at definition 3 from the following paper: Jeanne Ferrante, Karl J. Ottenstein, and Joe D. Warren. 1987. The program dependence graph and its use in optimization. ACM Trans. Program. Lang. Syst. 9, 3 (July 1987), 319–349. https://doi.org/10.1145/24039.24041

Let G be a control flow graph. Let X and Y be nodes in G. Y is
control dependent on X iff
(1) there exists a directed path P from X to Y with any 2 in P (excluding X
and Y) post-dominated by Y and
(2) X is not post-dominated by Y.

Verifying these conditions for every loop block would be quite inefficient. For example, let’s say we just want to check whether a given loop block is control dependent at all after the loop preheader. We would have to to verify the conditions above for every block between the loop preheader and the given loop block. This is n^2 complexity in the worst case. To optimize the control dependence checks, we can use post-dominance frontiers (PDF).

From Cooper, Keith D. et al. “A Simple, Fast Dominance Algorithm.” (1999).

A dominance frontier is the set of all CFG nodes, y, such that
b dominates a predecessor of y but does not strictly dominate y.

Reversing this for post-dominance we can see that the conditions for control dependence are the same as those for post-dominance frontiers. Thus, we rewrite our control dependence condition as Y is control dependent on X iff X is in PDF(Y).

We then can store the PDFs for every block as part of the context of this pass, and use it for checking control dependence. Using PDFs gets us from a worst case n^2 complexity to a worst case n.

Structs

LoopInvariantContext 🔒

Functions

can_be_hoisted 🔒

Indicates if the instruction can be safely hoisted out of a loop. If hoist_with_predicate is set, we assume we’re hoisting the instruction and its predicate, rather than just the instruction. Setting this means instructions that rely on predicates can be hoisted as well.