noirc_evaluator::ssa::opt::loop_invariant

Function can_be_hoisted

fn can_be_hoisted(
    instruction: &Instruction,
    dfg: &DataFlowGraph,
) -> CanBeHoistedResult

Indicates if the instruction can be safely hoisted out of a loop.

If it returns WithPredicate, the instruction can only be hoisted together with the predicate it relies on.

Preconditions

Certain instructions can be hoisted because they implicitly depend on a predicate. However, to avoid tight coupling between passes, we make the hoisting conditional on whether the caller wants the predicate to be taken into account or not.

Even if we know the predicate is the same for an instruction’s block and a loop’s header block, the caller of this methods needs to be careful as a loop may still never be executed. This is because an loop with dynamic bounds may never execute its loop body. If the instruction were to trigger a failure, our program may fail inadvertently. If we know a loop’s upper bound is greater than its lower bound we can hoist these instructions, but it is left to the caller of this method to account for this case.

This differs from can_be_deduplicated as that method assumes there is a matching instruction with the same inputs. Hoisting is for lone instructions, meaning a mislabeled hoist could cause unexpected failures if the instruction was never meant to be executed.