noirc_evaluator::ssa::opt::loop_invariant
Struct LoopInvariantContext

struct LoopInvariantContext<'f> {
    inserter: FunctionInserter<'f>,
    defined_in_loop: FxHashSet<Id<Value>>,
    loop_invariants: FxHashSet<Id<Value>>,
    current_induction_variables: FxHashMap<Id<Value>, (FieldElement, FieldElement)>,
    outer_induction_variables: FxHashMap<Id<Value>, (FieldElement, FieldElement)>,
    current_pre_header: Option<Id<BasicBlock>>,
    cfg: ControlFlowGraph,
    current_block_control_dependent: bool,
    post_dom_frontiers: FxHashMap<Id<BasicBlock>, FxHashSet<Id<BasicBlock>>>,
}

Fields
inserter: FunctionInserter<'f>defined_in_loop: FxHashSet<Id<Value>>loop_invariants: FxHashSet<Id<Value>>current_induction_variables: FxHashMap<Id<Value>, (FieldElement, FieldElement)>outer_induction_variables: FxHashMap<Id<Value>, (FieldElement, FieldElement)>current_pre_header: Option<Id<BasicBlock>>cfg: ControlFlowGraphcurrent_block_control_dependent: boolpost_dom_frontiers: FxHashMap<Id<BasicBlock>, FxHashSet<Id<BasicBlock>>>
Implementations

impl<'f> LoopInvariantContext<'f>

fn new(function: &'f mut Function) -> Self

fn pre_header(&self) -> Id<BasicBlock>

fn hoist_loop_invariants(&mut self, loop_: &Loop)

fn is_control_dependent_post_pre_header( &mut self, loop_: &Loop, block: Id<BasicBlock>, )

Checks whether a block is control dependent on any blocks after the given loop’s header.

fn is_control_dependent( &mut self, parent_block: Id<BasicBlock>, block: Id<BasicBlock>, ) -> bool

Checks whether a block is control dependent on a parent_block Uses post-dominance frontiers to determine control dependence. Reference the doc comments at the top of the this module for more information regarding post-dominance frontiers and control dependence.

fn set_values_defined_in_loop(&mut self, loop_: &Loop)

Gather the variables declared within the loop

fn extend_values_defined_in_loop_and_invariants( &mut self, instruction_id: Id<Instruction>, hoist_invariant: bool, )

Update any values defined in the loop and loop invariants after a analyzing and re-inserting a loop’s instruction.

fn can_hoist_invariant(&mut self, instruction_id: Id<Instruction>) -> bool

fn set_induction_var_bounds(&mut self, loop_: &Loop, current_loop: bool)

Keep track of a loop induction variable and respective upper bound. In the case of a nested loop, this will be used by later loops to determine whether they have operations reliant upon the maximum induction variable. When within the current loop, the known upper bound can be used to simplify instructions, such as transforming a checked add to an unchecked add.

fn can_be_hoisted_from_loop_bounds(&self, instruction: &Instruction) -> bool

Certain instructions can take advantage of that our induction variable has a fixed minimum/maximum.

For example, an array access can usually only be safely deduplicated when we have a constant index that is below the length of the array. Checking an array get where the index is the loop’s induction variable on its own would determine that the instruction is not safe for hoisting. However, if we know that the induction variable’s upper bound will always be in bounds of the array we can safely hoist the array access.

fn transform_to_unchecked_from_loop_bounds( &mut self, instruction_id: Id<Instruction>, )

Binary operations can take advantage of that our induction variable has a fixed minimum/maximum, to be transformed from a checked operation to an unchecked operation.

Checked operations require more bytecode and thus we aim to minimize their usage wherever possible.

For example, if one side of an add/mul operation is a constant and the other is an induction variable with a known upper bound, we know whether that binary operation will ever overflow. If we determine that an overflow is not possible we can convert the checked operation to unchecked.

fn can_evaluate_binary_op( &self, binary: &Binary, induction_vars: &HashMap<Id<Value>, (FieldElement, FieldElement)>, ) -> bool

Checks whether a binary operation can be evaluated using the bounds of a given loop induction variables.

If it cannot be evaluated, it means that we either have a dynamic loop bound or that the operation can potentially overflow during a given loop iteration.

fn map_dependent_instructions(&mut self)

Loop invariant hoisting only operates over loop instructions. The FunctionInserter is used for mapping old values to new values after re-inserting loop invariant instructions. However, there may be instructions which are not within loops that are still reliant upon the instruction results altered during the pass. This method re-inserts all instructions so that all instructions have correct new value IDs based upon the FunctionInserter internal map. Leaving out this mapping could lead to instructions with values that do not exist.