noirc_evaluator::ssa::opt::mem2reg
Module block

Structs

Block 🔒

A Block acts as a per-block context for the mem2reg pass. Most notably, it contains the current alias set thought to track each reference value if known, and it contains the expected ReferenceValue for each ValueId. When a block is finished, the final values of these are expected to match the values held by each ValueId at the very end of a block.

Enums

Expression 🔒

An Expression here is used to represent a canonical key into the aliases map since otherwise two dereferences of the same address will be given different ValueIds.