noirc_evaluator::ssa::opt::mem2reg::block
Struct Block

pub(super) struct Block {
    pub(super) expressions: OrdMap<ValueId, Expression>,
    pub(super) aliases: OrdMap<Expression, AliasSet>,
    pub(super) references: OrdMap<ValueId, ValueId>,
    pub(super) last_stores: OrdMap<ValueId, Id<Instruction>>,
    pub(super) last_loads: OrdMap<ValueId, Id<Instruction>>,
}

A Block acts as a per-block context for the mem2reg pass. Most notably, it contains the current alias set thought to track each reference value if known, and it contains the expected ReferenceValue for each ValueId. When a block is finished, the final values of these are expected to match the values held by each ValueId at the very end of a block.

Fields
expressions: OrdMap<ValueId, Expression>

Maps a ValueId to the Expression it represents. Multiple ValueIds can map to the same Expression, e.g. dereferences to the same allocation.

aliases: OrdMap<Expression, AliasSet>

Each expression is tracked as to how many aliases it may have. If there is only 1, we can attempt to optimize out any known loads to that alias. Note that โaliasโ here includes the original reference as well.

references: OrdMap<ValueId, ValueId>

Each allocate instruction result (and some reference block parameters) will map to a Reference value which tracks whether the last value stored to the reference is known.

last_stores: OrdMap<ValueId, Id<Instruction>>

The last instance of a Store instruction to each address in this block

last_loads: OrdMap<ValueId, Id<Instruction>>
Implementations

impl Block

pub(super) fn get_known_value(&self, address: ValueId) -> Option<ValueId>

If the given reference id points to a known value, return the value

pub(super) fn set_known_value(&mut self, address: ValueId, value: ValueId)

If the given address is known, set its value to value.

pub(super) fn set_unknown(&mut self, address: ValueId)

fn set_value(&mut self, address: ValueId, value: Option<ValueId>)

fn set_reference_value(&mut self, address: ValueId, value: Option<ValueId>)

fn invalidate_all_references(&mut self)

pub(super) fn unify(self, other: &Self) -> Self

pub(super) fn remember_dereference( &mut self, function: &Function, address: ValueId, result: ValueId, )

Remember that result is the result of dereferencing address. This is important to track aliasing when references are stored within other references.

fn keep_last_stores_for(&mut self, address: ValueId, function: &Function)

Forget the last store to an address and all of its aliases, to eliminate them from the candidates for removal at the end.

Note that this only affects this block: when we merge blocks we clear the last stores anyway, we donโt inherit them from predecessors, so if one block stores to an address and a descendant block loads it, this mechanism does not affect the candidacy of the last store in the predecessor block.

fn keep_last_store(&mut self, address: ValueId, function: &Function)

Forget the last store to an address, to remove it from the set of instructions which are candidates for removal at the end. Also marks the values in the last store as used, now that we know we want to keep them.

pub(super) fn mark_value_used(&mut self, value: ValueId, function: &Function)

Mark a value (for example an address we loaded) as used by forgetting the last store instruction, which removes it from the candidates for removal.

pub(super) fn collect_all_aliases( &self, values: impl IntoIterator<Item = ValueId>, ) -> AliasSet

Collect all aliases used by the given value list

pub(super) fn get_aliases_for_value(&self, value: ValueId) -> Cow<'_, AliasSet>

pub(super) fn set_last_load( &mut self, address: ValueId, instruction: Id<Instruction>, )

pub(super) fn keep_last_load_for(&mut self, address: ValueId)