noirc_evaluator::ssa::opt::mem2reg
Struct PerFunctionContext

struct PerFunctionContext<'f> {
    cfg: ControlFlowGraph,
    post_order: PostOrder,
    blocks: BTreeMap<BasicBlockId, Block>,
    inserter: FunctionInserter<'f>,
    instructions_to_remove: FxHashSet<Id<Instruction>>,
    instructions_analyzed: FxHashSet<Id<Instruction>>,
    last_loads: FxHashSet<ValueId>,
    instruction_input_references: FxHashSet<ValueId>,
    aliased_references: FxHashMap<ValueId, FxHashSet<Id<Instruction>>>,
    loop_aliases: FxHashSet<ValueId>,
}

Fields
cfg: ControlFlowGraphpost_order: PostOrderblocks: BTreeMap<BasicBlockId, Block>inserter: FunctionInserter<'f>instructions_to_remove: FxHashSet<Id<Instruction>>

Load and Store instructions that should be removed at the end of the pass.

We avoid removing individual instructions as we go since removing elements from the middle of Vecs many times will be slower than a single call to retain.

instructions_analyzed: FxHashSet<Id<Instruction>>

All instructions analyzed so far in the function. Anything new can be reinserted, while things that appear repeatedly have to be cloned.

last_loads: FxHashSet<ValueId>

Track a value’s last load across all blocks. If a value is not used in anymore loads we can remove the last store to that value.

instruction_input_references: FxHashSet<ValueId>

Track whether a reference was passed into another instruction (e.g. Call) This is needed to determine whether we can remove a store.

aliased_references: FxHashMap<ValueId, FxHashSet<Id<Instruction>>>

Track whether a reference has been aliased, and store the respective instruction that aliased that reference. If that store has been set for removal, we can also remove this instruction.

loop_aliases: FxHashSet<ValueId>

Loop carried aliases: references that may have aliases due to stores within loops. Contains values of references that should not have their stores removed.

Implementations

impl<'f> PerFunctionContext<'f>

fn new(function: &'f mut Function, loop_aliases: HashSet<ValueId>) -> Self

fn has_loop_carried_aliases(&self, address: ValueId) -> bool

Check if an address has loop carried aliases.

This is important for preventing incorrect store removal. If store v3 at v2 occurs in a loop, then v2 may point to v3 in future iterations. A subsequent store X at v3 should not have its previous store removed, as that store may be read through v2 (which now aliases v3) in a future loop iteration.

We only check the address, not the value being stored. The decision of whether to remove a previous store to address depends solely on whether address might be accessed through an alias, not on whether the value being stored has aliases.

fn mem2reg(&mut self)

Apply the mem2reg pass to the given function.

This function is expected to be the same one that the internal cfg, post_order, and dom_tree were created from.

fn is_store_alias_used( &self, store_address: &ValueId, block: &Block, all_terminator_values: &HashSet<ValueId>, per_func_block_params: &HashSet<ValueId>, ) -> bool

fn reference_parameters(&self) -> BTreeSet<ValueId>

Collect the input parameters of the function which are of reference type. All references are mutable, so these inputs are shared with the function caller and thus stores should not be eliminated, even if the blocks in this function don’t use them anywhere.

fn find_starting_references(&mut self, block: BasicBlockId) -> Block

The value of each reference at the start of the given block is the unification of the value of the same reference at the end of its predecessor blocks.

fn analyze_block(&mut self, block: BasicBlockId, references: Block)

Analyze a block with the given starting reference values.

This will remove any known loads in the block and track the value of references as they are stored to. When this function is finished, the value of each reference at the end of this block will be remembered in self.blocks.

fn add_aliases_for_reference_parameters( &self, block: BasicBlockId, references: &mut Block, )

Go through each parameter and register that all reference parameters of the same type are possibly aliased to each other. If there are parameters with nested references (arrays of references or references containing other references) we give up and assume all parameter references are AliasSet::unknown().

fn remove_stores_that_do_not_alias_parameters_or_returns( &mut self, references: &Block, )

Add all instructions in last_stores to self.instructions_to_remove which do not possibly alias any parameters of the given function.

fn analyze_instruction( &mut self, block_id: BasicBlockId, references: &mut Block, instruction_id: Id<Instruction>, )

fn analyze_possibly_simplified_instruction( &mut self, references: &mut Block, instruction: Id<Instruction>, simplified: bool, )

fn add_array_aliases(&self, elements: &Vector<ValueId>, aliases: &mut AliasSet)

In order to handle nested arrays we need to recursively search whether there are any references contained within an array’s elements.

fn set_aliases( &self, references: &mut Block, address: ValueId, new_aliases: AliasSet, )

fn clear_aliases(&self, references: &mut Block, value: ValueId)

fn mark_all_unknown(&self, values: &[ValueId], references: &mut Block)

fn remove_instructions(&mut self)

Remove any instructions in self.instructions_to_remove from the current function. This is expected to contain any loads which were replaced and any stores which are no longer needed.

fn update_data_bus(&mut self)

fn handle_terminator(&mut self, block: BasicBlockId, references: &mut Block)