noirc_evaluator::ssa::opt

Module mem2reg_simple

Mem2reg algorithm adapted from the paper: https://bernsteinbear.com/assets/img/bebenita-ssa.pdf

The goal is for this new, simpler to eventually replace our existing mem2reg algorithm in ssa/opt/mem2reg.rs. The pre-existing pass however can optimize in more/different cases than this pass. For example, it can still optimize out stores/loads in some cases even when the reference is aliased. That other pass has a larger surface area for bugs though and this one is simpler so the goal is to replace the old pass with this one plus any other, separate passes needed for the features unhandled here (such as alias analysis).

Structs

BlockState 🔒

Contains the starting & ending values of each variable in one block

BlockStates 🔒

Contains the starting & ending values of each variable in each block

Constants

MAX_BLOCK_SPAN_PRE_FLATTENING 🔒

Maximum number of blocks a variable’s declaration can dominate before we skip promoting it in the pre-flattening pass.

MAX_VARIABLES_OPTIMIZED 🔒

Arbitrary limit for maximum variables optimized by this pass in each function.

Functions

abstract_interpret_block 🔒

Abstractly interpret a block, collecting the value of each reference at the end of a block. Any references not included in the result are assumed to be unchanged.

add_block_params_and_find_exit_states 🔒

Find the starting & ending states of each variable in each block.

add_terminator_arguments 🔒

Link entry & exit states by adding terminator arguments for every variable stored to.

add_visible_variables_as_block_parameters 🔒

Filter variables, returning only those that are visible at the start of the given block. Since we do not consider variables within a block, the visible variables at the start of a block are any variables whose declaration block dominates the given block.

collect_all_eligible_variables 🔒

Return a map from each variable to the block it was declared in. This only includes variables that are eligible for mem2reg optimization, i.e. those that are allocated but never used in a first-class manner.

commit 🔒

Commit to all changes made by the pass:

get_terminator_args 🔒

Get the terminator arguments for block block jumping to block jmp_target. The jmp_target is relevant if block terminates in a jmpif terminator and may jmp to multiple blocks. Panics if the given block does not have block arguments.

get_terminator_args_mut 🔒

Get the terminator arguments for block block jumping to block jmp_target. The jmp_target is relevant if block terminates in a jmpif terminator and may jmp to multiple blocks. Panics if the given block does not have block arguments.

keep_argument_mask 🔒

Return a mask indicating whether to keep or remove the corresponding parameter.

remove_params_from_block_with_identical_terminator_args 🔒

Removes block parameters whose arguments (from predecessors’ terminators) are all identical

remove_params_from_blocks_with_identical_terminator_args 🔒

For every block, remove any block parameters whose arguments (from predecessors’ terminators) are all identical

retain_items_from_mask 🔒

For each index i of items, keep items[i] iff mask[i]

Type Aliases

StateVec 🔒

Mapping from a variable to its value at a point in time.