noirc_evaluator::ssa::opt
Module remove_enable_side_effects

The goal of the “remove enable side effects” optimization pass is to delay any Instruction::EnableSideEffectsIf instructions in ACIR functions such that they cover the minimum number of instructions possible.

The pass works as follows:

Insert instructions until an Instruction::EnableSideEffectsIf is encountered, save this InstructionId.

Continue inserting instructions until either

Another Instruction::EnableSideEffectsIf is encountered, if so then drop the previous InstructionId in favour of this one.

An Instruction that is affected by the side-effects variable is encountered, if so then insert the currently saved Instruction::EnableSideEffectsIf before the Instruction. Continue inserting instructions until the next Instruction::EnableSideEffectsIf is encountered.

The pass will also remove redundant Instruction::EnableSideEffectsIf instructions, for example if two consecutive Instruction::EnableSideEffectsIf instructions have the same condition.

This pass doesn’t run in Brillig functions as Instruction::EnableSideEffectsIf is not allowed in Brillig functions.

Preconditions:

This pass must run once ACIR functions only have one basic block.

Functions

remove_enable_side_effects_pre_check 🔒

Check that the CFG has been flattened.

should_insert_side_effects_before_instruction 🔒

Decide we should insert any pending side effect variable before we insert a particular instruction into the SSA.