noirc_evaluator::ssa::opt
Module remove_if_else

This file contains the SSA remove_if_else pass - a required pass for ACIR to remove any remaining Instruction::IfElse in the singular program-function, and replace them with arithmetic operations using the then_condition.

ACIR/Brillig differences within this pass:

This pass is strictly ACIR-only and never mutates brillig functions.

Conditions:

Precondition: Flatten CFG has been performed which should result in the function having only one basic block.

Precondition: then_value and else_value of Instruction::IfElse return arrays or slices. Numeric values should be handled previously by the flattening pass. Reference or function values are not handled by remove if-else and will cause an error.

Postcondition: A program without any IfElse instructions.

Relevance to other passes:

Flattening inserts Instruction::IfElse to merge array or slice values from an if-expression’s “then” and “else” branches. Instruction::IfElse with numeric values are directly handled during flattening, via instruction simplification, and will cause a panic in the remove_if_else pass.

Defunctionalize removes first-class function values from the program which eliminates the need for remove-if-else to handle Instruction::IfElse returning function values.

Implementation details & examples: IfElse instructions choose between its two operand values, then_value and else_value, based on the then_condition:

 if then_condition {
     then_value
 } else {
     else_value
 }

These instructions are inserted during the flatten cfg pass, which convert conditional control flow at the basic block level into simple ternary operations returning a value, using these IfElse instructions, and leaving only one basic block. The flatten cfg pass directly handles numeric values and issues Instruction::IfElse only for arrays and slices. The remove-if-else pass is used for array and slices in order to track their lengths, depending on existing slice intrinsics which modify slices, or the array set instructions. The Instruction::IfElse is removed using a ValueMerger which operates recursively for nested arrays/slices.

For example, this code:

fn main(x: bool, mut y: [u32; 2]) {
    if x {
         y[0] = 1;
    } else {
         y[0] = 2;
    }

    assert(y[0] == 3);
 }

will be translated into this code, where the IfElse instruction: v9 = if v0 then v5 else (if v6) v8 is using array v5 from then branch, and array v8 from the else branch:

acir(inline) predicate_pure fn main f0 {
  b0(v0: u1, v1: [u32; 2]):
    v2 = allocate -> &mut [u32; 2]
    enable_side_effects v0
    v5 = array_set v1, index u32 0, value u32 1
    v6 = not v0
    enable_side_effects v6
    v8 = array_set v1, index u32 0, value u32 2
    v9 = if v0 then v5 else (if v6) v8
    enable_side_effects u1 1
    v11 = array_get v9, index u32 0 -> u32
    constrain v11 == u32 3
    return
}

The IfElse instruction is then replaced by these instruction during the remove if-else pass:

v13 = cast v0 as u32
v14 = cast v6 as u32
v15 = unchecked_mul v14, u32 2
v16 = unchecked_add v13, v15
v17 = array_get v5, index u32 1 -> u32
v18 = array_get v8, index u32 1 -> u32
v19 = cast v0 as u32
v20 = cast v6 as u32
v21 = unchecked_mul v19, v17
v22 = unchecked_mul v20, v18
v23 = unchecked_add v21, v22
v24 = make_array [v16, v23] : [u32; 2]

The result of the removed IfElse instruction, array v24, is a merge of each of the elements of v5 and v8. The elements at index 0 are replaced by their known value, instead of doing an additional array get. Operations with the conditions are unchecked operations, because the conditions are 0 or 1, so it cannot overflow.

For slices the logic is similar except that slice lengths need to be tracked in order to know the length of the merged slice resulting in a make_array instruction. This length will be the maximum length of the two input slices. Note that the actual length of the merged slice should have been merged during flattening.

Structs

Context 🔒

Enums

SizeChange 🔒

Functions

remove_if_else_post_check 🔒

Post-check condition for Function::remove_if_else.

remove_if_else_pre_check 🔒