noirc_evaluator::ssa::opt
Module remove_unreachable_functions

Removes any unreachable functions from the code. These can result from optimizations making existing functions unreachable, e.g. if false { foo() }, or even from monomorphizing an unconstrained version of a constrained function where the original constrained version ends up never being used.

This pass identifies all unreachable functions and prunes them from the function set. Reachability is defined as:

A function is reachable if it is an entry point (e.g., main)

A function is reachable if it is called from another reachable function

A function is reachable if it is stored in a reference (e.g., in a Store instruction) from another reachable function. Even if not immediately called, it may later be dynamically loaded and invoked. This marking is conservative but ensures correctness. We should instead rely on mem2reg for resolving loads/stores.

A function is reachable if it is used in a block terminator (e.g., returned from a function)

The pass builds a call graph based upon the definition of reachability above. It then identifies all entry points and uses the CallGraph::reachable_from utility to mark all transitively reachable functions. It then discards the rest.

This pass helps shrink the SSA before compilation stages like inlining and dead code elimination.

Functions

reachable_functions 🔒

Identifies all reachable function IDs within the provided Ssa. This includes:

remove_unreachable_functions_post_check 🔒

Post-check condition for Ssa::remove_unreachable_functions.

used_functions 🔒

Identifies all reachable function IDs within a given function. This includes: