noirc_evaluator::ssa::opt::remove_unreachable_functions
Function remove_unreachable_functions_post_check

fn remove_unreachable_functions_post_check(ssa: &Ssa)

Post-check condition for Ssa::remove_unreachable_functions.

Succeeds if:

ssa contains no unreachable functions, i.e., all functions are reachable from the entry points.

Otherwise panics.

Note

We reuse the same logic for checking reachability as in the main pass, so this function will not catch any bugs in the pass itself, but it will ensure that the pass is idempotent.