noirc_evaluator::ssa::opt
Module remove_unreachable_instructions

This module defines an SSA pass to remove instructions that are unreachable. For example, if an instruction in a block is constrain u1 0 == u1 1, any subsequent instructions in that block will never be executed. This pass then removes those subsequent instructions and replaces the block’s terminator with a special unreachable value.

This pass might also add constrain checks after existing instructions, for example binary operations that are guaranteed to overflow.

Enums

Reachability 🔒

Functions

binary_operation_always_fails 🔒

zeroed_value 🔒