noirc_evaluator::ssa::opt::simple_optimization
Struct SimpleOptimizationContext

pub(crate) struct SimpleOptimizationContext<'dfg, 'mapping> {
    pub(crate) block_id: BasicBlockId,
    pub(crate) instruction_id: Id<Instruction>,
    pub(crate) call_stack_id: CallStackId,
    pub(crate) dfg: &'dfg mut DataFlowGraph,
    pub(crate) enable_side_effects: ValueId,
    values_to_replace: &'mapping mut ValueMapping,
    insert_current_instruction_at_callback_end: bool,
    orig_instruction_hash: u64,
}

Fields
block_id: BasicBlockIdinstruction_id: Id<Instruction>call_stack_id: CallStackIddfg: &'dfg mut DataFlowGraphenable_side_effects: ValueIdvalues_to_replace: &'mapping mut ValueMappinginsert_current_instruction_at_callback_end: boolorig_instruction_hash: u64
Implementations

impl SimpleOptimizationContext<'_, '_>

pub(crate) fn instruction(&self) -> &Instruction

Returns the current instruction being visited.

The instruction has already had its values updated with any replacements to be done.

pub(crate) fn replace_value(&mut self, from: ValueId, to: ValueId)

Instructs this context to replace a value with another value. The value will be replaced in all subsequent instructions.

fn has_instruction_changed(&self) -> bool

Check if the instruction has changed relative to its original contents, e.g. because any of its values have been replaced.

pub(crate) fn insert_current_instruction(&mut self)

Instructs this context to insert the current instruction right away, as opposed to doing this at the end of mutate’s block (unless remove_current_instruction is called).

If the instruction or its values has changed relative to their original content, we attempt to simplify the instruction before re-inserting it into the block.

pub(crate) fn remove_current_instruction(&mut self)

Instructs this context to remove the current instruction from its block.

pub(crate) fn insert_instruction( &mut self, instruction: Instruction, ctrl_typevars: Option<Vec<Type>>, ) -> InsertInstructionResult<'_>

Inserts an instruction in the current block right away.

pub(crate) fn insert_instruction_by_id( &mut self, instruction_id: Id<Instruction>, )

Inserts an instruction by id in the current block right away.

pub(crate) fn replace_current_instruction_with( &mut self, instruction: Instruction, )

Replaces the current instruction with another one.