noirc_evaluator::ssa::opt
Module simplify_cfg

This file contains the simplify cfg pass of the SSA IR.

This is a rather simple pass that is expected to be cheap to perform. It:

Removes blocks with no predecessors

Inlines a block into its sole predecessor if that predecessor only has one successor.

Removes any block arguments for blocks with only a single predecessor.

Removes any blocks which have no instructions other than a single terminating jmp.

Replaces any jmpifs with constant conditions with jmps. If this causes the block to have only 1 successor then (2) also will be applied.

Replacing any jmpifs with a negated condition with a jmpif with a un-negated condition and reversed branches.

Replaces any jmpif whose branches converge. Convergence is defined as two paths that ultimately jump to the same block through empty jump chains. If the branches converge, the jmpif is unnecessary and can be replaced with a simple jmp.

Currently only 1 is unimplemented.

Functions

check_for_constant_jmpif 🔒

Optimize a jmpif into a jmp if the condition is known

check_for_converging_jmpif 🔒

Attempts to simplify a jmpif terminator if both branches converge.

check_for_double_jmp 🔒

Optimize a jmp to a block which immediately jmps elsewhere to just jmp to the second block.

check_for_negated_jmpif_condition 🔒

Optimize a jmpif on a negated condition by swapping the branches.

remove_block_parameters 🔒

If the given block has block parameters, replace them with the jump arguments from the predecessor.

resolve_jmp_chain 🔒

Follow a chain of empty blocks to find the real destination.

simplify_current_block 🔒

A helper function to simplify the current block based on information on its successors.

try_inline_into_predecessor 🔒

Try to inline a block into its predecessor, returning true if successful.

try_inline_successor 🔒

Try to inline a block into its predecessor, returning true if successful.