noirc_evaluator::ssa::opt::simplify_cfg

Function remove_block_parameters

fn remove_block_parameters(
    function: &mut Function,
    block: BasicBlockId,
    predecessor: BasicBlockId,
    values_to_replace: &mut ValueMapping,
)

If the given block has block parameters, replace them with the jump arguments from the predecessor.

Currently, if this function is needed, try_inline_into_predecessor will also always apply, although in the future it is possible for only this function to apply if jmpif instructions with block arguments are ever added.