noirc_evaluator::ssa::opt::unrolling

Function get_induction_variable

fn get_induction_variable(
    dfg: &DataFlowGraph,
    block: BasicBlockId,
) -> Result<ValueId, CallStack>

Return the induction value of the current iteration of the loop, from the given block’s jmp arguments.

Expects the current block to terminate in jmp h(N) where h is the loop header and N is a Field value. Returns an Err if this isn’t the case.

Consider the following example:

main():
  v0 = ... start ...
  v1 = ... end ...
  jmp loop_entry(v0)
loop_entry(i: Field):
  ...

We’re looking for the terminating jump of the main predecessor of loop_entry.