noirc_evaluator::ssa::opt::unrolling
Struct Loop

pub(crate) struct Loop {
    pub(crate) header: BasicBlockId,
    pub(crate) back_edge_start: BasicBlockId,
    pub(crate) blocks: BTreeSet<BasicBlockId>,
}

Describe the blocks that constitute up a loop.

Fields
header: BasicBlockId

The header block of a loop is the block which dominates all the other blocks in the loop.

back_edge_start: BasicBlockId

The start of the back_edge n -> d is the block n at the end of the loop that jumps back to the header block d which restarts the loop.

blocks: BTreeSet<BasicBlockId>

All the blocks contained within the loop, including header and back_edge_start.

Implementations

impl Loop

fn get_induction_variable(&self, function: &Function) -> Option<ValueId>

Find the value that controls whether to perform a loop iteration. This is going to be the block parameter of the loop header.

Consider the following example of a for i in 0..4 loop:

brillig(inline) fn main f0 {
  b0(v0: u32):
    ...
    jmp b1(u32 0)
  b1(v1: u32):                  // Loop header
    v5 = lt v1, u32 4           // Upper bound
    jmpif v5 then: b3, else: b2

In the example above, v1 is the induction variable.

There is an example in the tests where a loop does not have an induction variable, but rather loads a reference in the header, in which case this will return None.

pub(super) fn is_fully_executed(&self, cfg: &ControlFlowGraph) -> bool

Check if the loop will be fully executed, that is, there is no early break in it.

Our SSA code generation restricts loops to having one exit block. If the exit block has only one predecessor, that means there is no break in the loop.

If a loop can have several exit blocks, we would need to update this function.

If the loop header doesn’t lead to an exit block, then it must be a loop or while, rather than a for loop. Even if such blocks don’t have a break (e.g. they are infinite), we don’t consider them fully executed.

impl Loop

pub(crate) fn find_blocks_in_loop( header: BasicBlockId, back_edge_start: BasicBlockId, cfg: &ControlFlowGraph, ) -> Self

Return each block that is in a loop starting in the given header block. Expects back_edge_start -> header to be the back edge of the loop.

fn has_const_back_edge_induction_value(&self, dfg: &DataFlowGraph) -> bool

Check that the loop does not end with a constant value passed to the header from the back-edge, which would result in a loop we would never finish unrolling.

This can happen if constant folding replaces a variable with a constant it is constrained to equal (which doesn’t even have to fall into the loop bounds).

For example:

brillig(inline) predicate_pure fn main f0 {
  b0():
    jmp b1(u32 1)                // Pre-header
  b1(v0: u32):                   // Header
    v3 = lt v0, u32 20
    jmpif v3 then: b2, else: b3
  b2():                          // Back edge
    constrain v0 == u32 1        // Constrain the induction variable to a known value.
    jmp b1(u32 2)                // `v1 = unchecked_add v0, u32 1; jmp b1(v1)` replaced by `jmp b1 (1+1)`
  b3():
    return
}

fn has_const_zero_jump_condition(&self, dfg: &DataFlowGraph) -> bool

Check if the loop header has a constant zero jump condition, which indicates an empty loop.

This can happen if a jump condition has been simplified out.

For example:

brillig(inline) predicate_pure fn main f0 {
  b0():
    jmp b1(u32 10)               // Pre-header
  b1(v0: u32):                   // Header
    // v3 = lt v0, u32 0         // Simplified to `u1 0`
    jmpif u1 0 then: b2, else: b3
  b2():                          // Back edge
    v1 = unchecked_add v0, u32 1 // Increment induction value
    jmp b1(v1)
  b3():
    return
}

fn get_const_lower_bound( &self, dfg: &DataFlowGraph, pre_header: BasicBlockId, ) -> Option<IntegerConstant>

Find the lower bound of the loop in the pre-header and return it if it’s a numeric constant, which it will be if the previous SSA steps managed to inline it.

Consider the following example of a for i in 0..4 loop:

brillig(inline) fn main f0 {
  b0(v0: u32):                  // Pre-header
    ...
    jmp b1(u32 0)               // Lower-bound
  b1(v1: u32):                  // Induction variable
    v5 = lt v1, u32 4
    jmpif v5 then: b3, else: b2

fn get_const_upper_bound( &self, dfg: &DataFlowGraph, pre_header: BasicBlockId, ) -> Option<IntegerConstant>

Find the upper bound of the loop in the loop header and return it if it’s a numeric constant, which it will be if the previous SSA steps managed to inline it.

Consider the following example of a for i in 0..4 loop:

brillig(inline) fn main f0 {
  b0(v0: u32):
    ...
    jmp b1(u32 0)
  b1(v1: u32):                  // Loop header
    v5 = lt v1, u32 4           // Upper bound
    jmpif v5 then: b3, else: b2

pub(super) fn get_const_bounds( &self, dfg: &DataFlowGraph, pre_header: BasicBlockId, ) -> Option<(IntegerConstant, IntegerConstant)>

Get the lower and upper bounds of the loop if both are constant numeric values.

fn unroll( &self, function: &mut Function, cfg: &ControlFlowGraph, ) -> Result<(), CallStack>

Unroll a single loop in the function. Returns Ok(()) if it succeeded, Err(callstack) if it failed, where the callstack indicates the location of the instruction that could not be processed, or empty if such information was not available.

Consider this example:

main():
  v0 = 0
  v1 = 2
  jmp loop_entry(v0)
loop_entry(i: Field):
  v2 = lt i, v1
  jmpif v2, then: loop_body, else: loop_end

The first step is to unroll the header by recognizing that jump condition is a constant, which means it will go to loop_body:

main():
  v0 = 0
  v1 = 2
  v2 = lt v0, v1
  // jmpif v2, then: loop_body, else: loop_end
  jmp dest: loop_body

Following that we unroll the loop body, which is the next source, replace the induction variable with the new value created in the body, and have another go at the header.

main():
  v0 = 0
  v1 = 2
  v2 = lt v0, v1
  v3 = ... body ...
  v4 = add v0, u32 1
  jmp loop_entry(v4)

At the end we reach a point where the condition evaluates to 0 and we jump to the end.

main():
  v0 = 0
  v1 = 2
  v2 = lt v0, v1
  v3 = ... body ...
  v4 = add u32 1, v0
  v5 = lt v4, v1
  v6 = ... body ...
  v7 = add v4, u32 1
  v8 = lt v7, v1
  jmp loop_end

When e.g. v8 = lt v7, v1 cannot be evaluated to a constant, the loop signals by returning Err that a few SSA passes are required to evaluate and simplify these values.

pub(super) fn get_pre_header( &self, function: &Function, cfg: &ControlFlowGraph, ) -> Result<BasicBlockId, CallStack>

The loop pre-header is the block that comes before the loop begins. Generally a header block is expected to have 2 predecessors: the pre-header and the final block of the loop which jumps back to the beginning. Other predecessors can come from break or continue.

fn unroll_header<'a>( &'a self, function: &'a mut Function, unroll_into: BasicBlockId, induction_value: ValueId, ) -> Result<Option<(LoopIteration<'a>, BasicBlockId)>, CallStack>

Unrolls the header block of the loop. This is the block that dominates all other blocks in the loop and contains the jmpif instruction that lets us know if we should continue looping. Returns Some((iteration context, loop_header_id)) if we should perform another iteration.

fn find_pre_header_reference_values( &self, function: &Function, cfg: &ControlFlowGraph, ) -> Option<HashSet<ValueId>>

Find all reference values which were allocated before the pre-header.

These are accessible inside the loop body, and they can be involved in load/store operations that could be eliminated if we unrolled the body into the pre-header.

Consider this loop:

let mut sum = 0;
let mut arr = &[];
for i in 0..3 {
    sum = sum + i;
    arr.push_back(sum)
}
sum

The SSA has a load+store for the sum and a load+push for the arr:

b0(v0: u32):
  v2 = allocate -> &mut u32     // reference allocated for `sum`
  store u32 0 at v2             // initial value for `sum`
  v4 = allocate -> &mut u32     // reference allocated for the length of `arr`
  store u32 0 at v4             // initial length of `arr`
  inc_rc [] of u32              // storage for `arr`
  v6 = allocate -> &mut [u32]   // reference allocated to point at the storage of `arr`
  store [] of u32 at v6         // initial value for the storage of `arr`
  jmp b1(u32 0)                 // start looping from 0
b1(v1: u32):                    // `i` induction variable
  v8 = lt v1, u32 3             // loop until 3
  jmpif v8 then: b3, else: b2
b3():
  v11 = load v2 -> u32          // load `sum`
  v12 = add v11, v1             // add `i` to `sum`
  store v12 at v2               // store updated `sum`
  v13 = load v4 -> u32          // load length of `arr`
  v14 = load v6 -> [u32]        // load storage of `arr`
  v16, v17 = call vector_push_back(v13, v14, v12) -> (u32, [u32]) // builtin to push, will store to storage and length references
  v19 = add v1, u32 1           // increase `arr`
  jmp b1(v19)                   // back-edge of the loop
b2():                           // after the loop
  v9 = load v2 -> u32           // read final value of `sum`

We won’t always find load and store ops (e.g. the push above doesn’t come with a store), but it’s likely that mem2reg could eliminate a lot of the loads we can find, so we can use this as an approximation of the gains we would see.

fn count_loads_and_stores( &self, function: &Function, refs: &HashSet<ValueId>, ) -> (usize, usize)

Count the number of load and store instructions of specific variables in the loop.

Returns (loads, stores) in case we want to differentiate in the estimates.

fn count_all_instructions(&self, function: &Function) -> usize

Count the number of instructions in the loop, including the terminating jumps.

fn count_induction_increments(&self, function: &Function) -> usize

Count the number of increments to the induction variable. It should be one, but it can be duplicated. The increment should be in the block where the back-edge was found.

fn is_small_loop(&self, function: &Function, cfg: &ControlFlowGraph) -> bool

Decide if this loop is small enough that it can be inlined in a way that the number of unrolled instructions times the number of iterations would result in smaller bytecode than if we keep the loops with their overheads.

fn boilerplate_stats( &self, function: &Function, cfg: &ControlFlowGraph, ) -> Option<BoilerplateStats>

Collect boilerplate stats if we can figure out the upper and lower bounds of the loop, and the loop doesn’t have multiple back-edges from breaks and continues.