noirc_evaluator::ssa::parser::into_ssa
Struct Translator

struct Translator {
    builder: FunctionBuilder,
    functions: HashMap<String, Id<Function>>,
    blocks: HashMap<Id<Function>, HashMap<String, BasicBlockId>>,
    variables: HashMap<Id<Function>, HashMap<String, ValueId>>,
    globals_function: Function,
    global_types: Vec<Type>,
    global_values: HashMap<String, ValueId>,
    globals_graph: Arc<GlobalsGraph>,
    error_selector_counter: u64,
    purities: Arc<FxHashMap<Id<Function>, Purity>>,
}

Fields
builder: FunctionBuilderfunctions: HashMap<String, Id<Function>>

Maps internal function names (e.g. “f1”) to their IDs

blocks: HashMap<Id<Function>, HashMap<String, BasicBlockId>>

Maps block names to their IDs

variables: HashMap<Id<Function>, HashMap<String, ValueId>>

Maps variable names to their IDs.

This is necessary because the SSA we parse might have undergone some passes already which replaced some of the original IDs. The translator will recreate the SSA step by step, which can result in a new ID layout.

globals_function: Function

The function that will hold the actual SSA globals.

global_types: Vec<Type>

The types of globals in the parsed SSA, in the order they were defined.

global_values: HashMap<String, ValueId>

Maps names (e.g. “g0”) in the parsed SSA to global IDs.

globals_graph: Arc<GlobalsGraph>error_selector_counter: u64purities: Arc<FxHashMap<Id<Function>, Purity>>
Implementations

impl Translator

fn translate( parsed_ssa: ParsedSsa, simplify: bool, validate: bool, ) -> Result<Ssa, SsaError>

fn new(parsed_ssa: &mut ParsedSsa, simplify: bool) -> Result<Self, SsaError>

fn translate_non_main_function( &mut self, function: ParsedFunction, ) -> Result<(), SsaError>

fn translate_function_body( &mut self, function: ParsedFunction, ) -> Result<(), SsaError>

fn translate_function_data_bus( &mut self, parsed_data_bus: ParsedDataBus, ) -> Result<(), SsaError>

fn compute_blocks_order( &self, entry_block_id: BasicBlockId, parsed_blocks_by_id: &HashMap<BasicBlockId, ParsedBlock>, ) -> Result<Vec<BasicBlockId>, SsaError>

Computes the order in which blocks should be translated. The order will be according to the block terminators, starting from the entry block. This is needed because a variable in a block might refer to a variable that syntactically happens afterwards, but logically happens before.

fn translate_block(&mut self, block: ParsedBlock) -> Result<(), SsaError>

fn translate_instruction( &mut self, instruction: ParsedInstruction, ) -> Result<(), SsaError>

fn translate_values( &mut self, values: Vec<ParsedValue>, ) -> Result<Vec<ValueId>, SsaError>

fn translate_value(&mut self, value: ParsedValue) -> Result<ValueId, SsaError>

fn translate_globals( &mut self, globals: Vec<ParsedGlobal>, ) -> Result<(), SsaError>

fn translate_global(&mut self, global: ParsedGlobal) -> Result<(), SsaError>

fn define_variable( &mut self, identifier: Identifier, value_id: ValueId, ) -> Result<(), SsaError>

fn lookup_variable(&self, identifier: &Identifier) -> Result<ValueId, SsaError>

fn define_global( &mut self, identifier: Identifier, value_id: ValueId, ) -> Result<(), SsaError>

fn lookup_global(&self, identifier: Identifier) -> Result<ValueId, SsaError>

fn lookup_block( &self, identifier: &Identifier, ) -> Result<BasicBlockId, SsaError>

fn lookup_function( &self, identifier: &Identifier, ) -> Result<Id<Function>, SsaError>

fn lookup_call_function( &mut self, function: Identifier, ) -> Result<ValueId, SsaError>

fn finish(self) -> Ssa

fn current_function_id(&self) -> Id<Function>

fn set_offset( &mut self, target: &Identifier, offset: ArrayOffset, ) -> Result<(), SsaError>

If any array instruction has an offset, mark the DFG as using offsets in general.