noirc_evaluator::ssa::ssa_gen
Module context

Structs

FunctionContext 🔒

The FunctionContext is the main context object for translating a function into SSA form during the SSA-gen pass.

Loop 🔒

SharedContext 🔒

Shared context for all functions during ssa codegen. This is the only object that is shared across all threads when generating ssa in multiple threads.

Enums

LValue 🔒

Used to remember the results of each step of extracting a value from an ast::LValue

Functions

convert_operator 🔒

Converts the given operator to the appropriate BinaryOp. Take care when using this to insert a binary instruction: this requires checking operator_requires_not and operator_requires_swapped_operands to represent the full operation correctly.

operator_requires_not 🔒

True if the given operator cannot be encoded directly and needs to be represented as !(some other operator)

operator_requires_swapped_operands 🔒

True if the given operator cannot be encoded directly and needs to have its lhs and rhs swapped to be represented with another operator. Example: (a > b) needs to be represented as (b < a)

Type Aliases

FunctionQueue 🔒

The queue of functions remaining to compile