noirc_evaluator::ssa::ssa_gen::context
Struct FunctionContext

pub(super) struct FunctionContext<'a> {
    definitions: FxHashMap<LocalId, Tree<Value>>,
    pub(super) builder: FunctionBuilder,
    shared_context: &'a SharedContext,
    loops: Vec<Loop>,
}

The FunctionContext is the main context object for translating a function into SSA form during the SSA-gen pass.

This context can be used to build any amount of functions, so long as it is cleared out in between each function via calling self.new_function().

If compiling many functions across multiple threads, there should be a separate FunctionContext for each thread. Each FunctionContext can communicate via the SharedContext field which as its name suggests is the only part of the context that needs to be shared between threads.

Fields
definitions: FxHashMap<LocalId, Tree<Value>>builder: FunctionBuildershared_context: &'a SharedContextloops: Vec<Loop>

Contains any loops we’re currently in the middle of translating. These are ordered such that an inner loop is at the end of the vector and outer loops are at the beginning. When a loop is finished, it is popped.

Implementations

impl<'a> FunctionContext<'a>

pub(super) fn new( function_name: String, parameters: &Parameters, runtime: RuntimeType, shared_context: &'a SharedContext, globals: GlobalsGraph, ) -> Self

Create a new FunctionContext to compile the first function in the shared_context’s function queue.

This will pop from the function queue, so it is expected the shared_context’s function queue is non-empty at the time of calling this function. This can be ensured by calling shared_context.get_or_queue_function(function_to_queue) before calling this constructor.

function_name and parameters are expected to be the name and parameters of the function this constructor will pop from the function queue.

pub(super) fn new_function(&mut self, id: Id<Function>, func: &Function)

Finish building the current function and switch to building a new function with the given name, id, and parameters.

Note that the previous function cannot be resumed after calling this. Developers should avoid calling new_function until the previous function is completely finished with ssa-gen.

fn add_parameters_to_scope(&mut self, parameters: &Parameters)

Add each parameter to the current scope, and return the list of parameter types.

The returned parameter type list will be flattened, so any struct parameters will be returned as one entry for each field (recursively).

fn add_parameter_to_scope( &mut self, parameter_id: LocalId, parameter_type: &Type, mutable: bool, )

Adds a “single” parameter to scope.

Single is in quotes here because in the case of tuple parameters, the tuple is flattened into a new parameter for each field recursively.

pub(super) fn new_mutable_variable(&mut self, value_to_store: ValueId) -> Value

Allocate a single slot of memory and store into it the given initial value of the variable. Always returns a Value::Mutable wrapping the allocate instruction.

pub(super) fn map_type<T>(typ: &Type, f: impl FnMut(Type) -> T) -> Tree<T>

Maps the given type to a Tree of the result type.

This can be used to (for example) flatten a tuple type, creating and returning a new parameter for each field type.

fn map_type_helper<T>(typ: &Type, f: &mut dyn FnMut(Type) -> T) -> Tree<T>

pub(super) fn convert_type(typ: &Type) -> Tree<Type>

Convert a monomorphized type to an SSA type, preserving the structure of any tuples within.

pub(super) fn convert_non_tuple_type(typ: &Type) -> Type

Converts a non-tuple type into an SSA type. Panics if a tuple type is passed.

This function is needed since this SSA IR has no concept of tuples and thus no type for them. Use convert_type if tuple types need to be handled correctly.

pub(super) fn unit_value() -> Tree<Value>

Returns the unit value, represented as an empty tree of values

pub(super) fn checked_numeric_constant( &mut self, value: SignedField, numeric_type: NumericType, ) -> Result<ValueId, RuntimeError>

Insert a numeric constant into the current function

Unlike FunctionBuilder::numeric_constant, this version checks the given constant is within the range of the given type. This is needed for user provided values where otherwise values like 2^128 can be assigned to a u8 without error or wrapping.

pub(super) fn insert_binary( &mut self, lhs: ValueId, operator: BinaryOpKind, rhs: ValueId, location: Location, ) -> Tree<Value>

Insert a binary instruction at the end of the current block. Converts the form of the binary instruction as necessary (e.g. swapping arguments, inserting a not) to represent it in the IR. For example, (a <= b) is represented as !(b < a)

pub(super) fn insert_call( &mut self, function: ValueId, arguments: Vec<ValueId>, result_type: &Type, location: Location, ) -> Tree<Value>

Inserts a call instruction at the end of the current block and returns the results of the call.

Compared to self.builder.insert_call, this version will reshape the returned Vec<ValueId> back into a Values tree of the proper shape.

pub(super) fn insert_safe_cast( &mut self, value: ValueId, typ: NumericType, location: Location, ) -> ValueId

Inserts a cast instruction at the end of the current block and returns the results of the cast.

Compared to self.builder.insert_cast, this version will automatically truncate value to be a valid typ.

fn sign_extend( &mut self, value: ValueId, incoming_type_size: u32, target_type_size: u32, location: Location, ) -> ValueId

During casting signed values, if target size is bigger, we do a sign extension:

When the value is negative, it is represented in 2-complement form; 2^s-v, where s is the incoming bit size and v is the absolute value. Sign extension in this case will give 2^t-v, where t is the target bit size. So we simply convert 2^s-v into 2^t-v by adding 2^t-2^s to the value when the value is negative.

Casting s-bits signed v0 to t-bits will add the following instructions:

v1 = cast v0 to 's-bits unsigned'
v2 = lt v1, 2**(s-1)
v3 = not(v1)
v4 = cast v3 to 't-bits unsigned'
v5 = v3 * (2**t - 2**s)
v6 = cast v1 to 't-bits unsigned'
return v6 + v5

Return an unsigned value that we can cast back to the signed type if we want, or keep it as it is, if we did the sign extension as part of casting e.g. i8 to u64.

pub(super) fn make_offset( &mut self, address: ValueId, offset: u128, unchecked: bool, ) -> ValueId

Create a const offset of an address for an array load or store

pub(super) fn make_array_index(&mut self, index: ValueId) -> ValueId

Array indexes are u32. This function casts values used as indexes to u32.

pub(super) fn define(&mut self, id: LocalId, value: Tree<Value>)

Define a local variable to be some Values that can later be retrieved by calling self.lookup(id)

pub(super) fn lookup(&self, id: LocalId) -> Tree<Value>

Looks up the value of a given local variable. Expects the variable to have been previously defined or panics otherwise.

pub(super) fn lookup_global(&self, id: GlobalId) -> Tree<Value>

pub(super) fn get_field(tuple: Tree<Value>, field_index: usize) -> Tree<Value>

Extract the given field of the tuple. Panics if the given Values is not a Tree::Branch or does not have enough fields.

pub(super) fn get_field_ref( tuple: &Tree<Value>, field_index: usize, ) -> &Tree<Value>

Extract the given field of the tuple by reference. Panics if the given Values is not a Tree::Branch or does not have enough fields.

pub(super) fn replace_field( tuple: Tree<Value>, field_index: usize, new_value: Tree<Value>, ) -> Tree<Value>

Replace the given field of the tuple with a new one. Panics if the given Values is not a Tree::Branch or does not have enough fields.

pub(super) fn get_or_queue_function(&mut self, id: FuncId) -> Tree<Value>

Retrieves the given function, adding it to the function queue if it is not yet compiled.

pub(super) fn extract_current_value( &mut self, lvalue: &LValue, ) -> Result<LValue, RuntimeError>

Extracts the current value out of an LValue.

Goal: Handle the case of assigning to nested expressions such as foo.bar[i1].baz[i2] = e while also noting that assigning to arrays will create a new array rather than mutate the original.

Method: First extract_current_value must recurse on the lvalue to extract the current value contained:

v0 = foo.bar                 ; allocate instruction for bar
v1 = load v0                 ; loading the bar array
v2 = add i1, baz_index       ; field offset for index i1, field baz
v3 = array_get v1, index v2  ; foo.bar[i1].baz

Method (part 2): Then, assign_new_value will recurse in the opposite direction to construct the larger value as needed until we can store to the nearest allocation.

v4 = array_set v3, index i2, e ; finally create a new array setting the desired value v5 = array_set v1, index v2, v4 ; now must also create the new bar array store v5 in v0 ; and store the result in the only mutable reference

The returned LValueRef tracks the current value at each step of the lvalue. This is later used by assign_new_value to construct a new updated value that can be assigned to an allocation within the LValueRef::Ident.

This is operationally equivalent to extract_current_value_recursive, but splitting these into two separate functions avoids cloning the outermost Values returned by the recursive version, as it is only needed for recursion.

fn dereference_lvalue( &mut self, values: &Tree<Value>, element_type: &Type, ) -> Tree<Value>

fn ident_lvalue(&self, ident: &Ident) -> (Tree<Value>, bool)

Compile the given identifier as a reference - ie. avoid calling .eval(). Returns the variable’s value and whether the variable is mutable.

fn index_lvalue( &mut self, array: &LValue, index: &Expression, location: &Location, ) -> Result<(ValueId, ValueId, LValue, Option<ValueId>), RuntimeError>

Compile the given array[index] expression as a reference. This will return a triple of (array, index, lvalue_ref, Option<length>) where the lvalue_ref records the structure of the lvalue expression for use by assign_new_value. The optional length is for indexing vectors rather than arrays since vectors are represented as a tuple in the form: (length, vector contents).

fn extract_current_value_recursive( &mut self, lvalue: &LValue, ) -> Result<(Tree<Value>, LValue), RuntimeError>

pub(super) fn assign_new_value( &mut self, lvalue: LValue, new_value: Tree<Value>, )

Assigns a new value to the given LValue. The LValue can be created via a previous call to extract_current_value. This method recurs on the given LValue to create a new value to assign an allocation instruction within an LValue::Ident or LValue::Dereference - see the comment on extract_current_value for more details.

fn assign_lvalue_index( &mut self, new_value: Tree<Value>, array: ValueId, index: ValueId, location: Location, ) -> ValueId

fn element_size(&self, array: ValueId) -> FieldElement

fn assign(&mut self, lhs: Tree<Value>, rhs: Tree<Value>)

Given an lhs containing only references, create a store instruction to store each value of rhs into its corresponding value in lhs.

pub(crate) fn enter_loop(&mut self, loop_: Loop)

pub(crate) fn exit_loop(&mut self)

pub(crate) fn current_loop(&self) -> Loop

impl FunctionContext<'_>

fn codegen_function_body( &mut self, body: &Expression, ) -> Result<(), RuntimeError>

Codegen a function’s body and set its return value to that of its last parameter. For functions returning nothing, this will be an empty list.

fn codegen_expression( &mut self, expr: &Expression, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_non_tuple_expression( &mut self, expr: &Expression, ) -> Result<ValueId, RuntimeError>

Codegen any non-tuple expression so that we can unwrap the Values tree to return a single value for use with most SSA instructions.

fn codegen_ident_reference(&mut self, ident: &Ident) -> Tree<Value>

Codegen a reference to an ident. The only difference between this and codegen_ident is that if the variable is mutable as in let mut var = ...; the Value::Mutable will be returned directly instead of being automatically loaded from. This is needed when taking the reference of a variable to reassign to it. Note that mutable references let x = &mut ...; do not require this since they are not automatically loaded from and must be explicitly dereferenced.

fn codegen_ident(&mut self, ident: &Ident) -> Tree<Value>

Codegen an identifier, automatically loading its value if it is mutable.

fn codegen_literal( &mut self, literal: &Literal, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_array_elements( &mut self, elements: &[Expression], ) -> Result<Vec<Tree<Value>>, RuntimeError>

fn codegen_string(&mut self, string: &str) -> Tree<Value>

fn codegen_array_checked( &mut self, elements: Vec<Tree<Value>>, typ: Type, ) -> Result<Tree<Value>, RuntimeError>

The bool aspect of each array element indicates whether the element is an array constant or not. If it is, we avoid incrementing the reference count because we consider the constant to be moved into this larger array constant.

fn codegen_array( &mut self, elements: Vec<Tree<Value>>, typ: Type, ) -> Tree<Value>

Codegen an array by allocating enough space for each element and inserting separate store instructions until each element is stored. The store instructions will be separated by add instructions to calculate the new offset address to store to next.

In the case of arrays of structs, the structs are flattened such that each field will be stored next to the other fields in memory. So an array such as [(1, 2), (3, 4)] is stored the same as the array [1, 2, 3, 4].

The bool aspect of each array element indicates whether the element is an array constant or not. If it is, we avoid incrementing the reference count because we consider the constant to be moved into this larger array constant.

The value returned from this function is always that of the allocate instruction.

fn codegen_block( &mut self, block: &[Expression], ) -> Result<Tree<Value>, RuntimeError>

fn codegen_unary(&mut self, unary: &Unary) -> Result<Tree<Value>, RuntimeError>

fn dereference( &mut self, values: &Tree<Value>, element_type: &Type, ) -> Tree<Value>

fn codegen_reference( &mut self, expr: &Expression, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_binary( &mut self, binary: &Binary, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_index(&mut self, index: &Index) -> Result<Tree<Value>, RuntimeError>

fn codegen_array_index( &mut self, array: ValueId, index: ValueId, element_type: &Type, location: Location, length: Option<ValueId>, ) -> Result<Tree<Value>, RuntimeError>

This is broken off from codegen_index so that it can also be used to codegen a LValue::Index.

Set load_result to true to load from each relevant index of the array (it may be multiple in the case of tuples). Set it to false to instead return a reference to each element, for use with the store instruction.

fn codegen_access_check(&mut self, index: ValueId, length: ValueId)

Prepare an array or vector access. Check that the index being used to access an array/vector element is less than the (potentially dynamic) array/vector length.

fn codegen_cast(&mut self, cast: &Cast) -> Result<Tree<Value>, RuntimeError>

fn codegen_for(&mut self, for_expr: &For) -> Result<Tree<Value>, RuntimeError>

Codegens a for loop, creating three new blocks in the process. The return value of a for loop is always a unit literal.

For example, the loop for i in start .. end { body } is codegen’d as:

  v0 = ... codegen start ...
  v1 = ... codegen end ...
  br loop_entry(v0)
loop_entry(i: Field):
  v2 = lt i v1
  jmpif v2, then: loop_body, else: loop_end
loop_body():
  v3 = ... codegen body ...
  v4 = add 1, i
  br loop_entry(v4)
loop_end():
  ... This is the current insert point after codegen_for finishes ...

fn codegen_loop( &mut self, block: &Expression, ) -> Result<Tree<Value>, RuntimeError>

Codegens a loop, creating two new blocks in the process. The return value of a loop is always a unit literal.

For example, the loop loop { body } is codegen’d as:

  br loop_body()
loop_body():
  v3 = ... codegen body ...
  br loop_body()
loop_end():
  ... This is the current insert point after codegen_for finishes ...

fn codegen_while(&mut self, while_: &While) -> Result<Tree<Value>, RuntimeError>

Codegens a while loop, creating three new blocks in the process. The return value of a while is always a unit literal.

For example, the loop while cond { body } is codegen’d as:

  jmp while_entry()
while_entry:
  v0 = ... codegen cond ...
  jmpif v0, then: while_body, else: while_end
while_body():
  v3 = ... codegen body ...
  jmp while_entry()
while_end():
  ... This is the current insert point after codegen_while finishes ...

fn codegen_if(&mut self, if_expr: &If) -> Result<Tree<Value>, RuntimeError>

Codegens an if expression, handling the case of what to do if there is no ‘else’.

For example, the expression if cond { a } else { b } is codegen’d as:

  v0 = ... codegen cond ...
  jmpif v0, then: then_block, else: else_block
then_block():
  v1 = ... codegen a ...
  br end_if(v1)
else_block():
  v2 = ... codegen b ...
  br end_if(v2)
end_if(v3: ?):  // Type of v3 matches the type of a and b
  ... This is the current insert point after codegen_if finishes ...

As another example, the expression if cond { a } is codegen’d as:

  v0 = ... codegen cond ...
  jmpif v0, then: then_block, else: end_if
then_block:
  v1 = ... codegen a ...
  br end_if()
end_if:  // No block parameter is needed. Without an else, the unit value is always returned.
  ... This is the current insert point after codegen_if finishes ...

fn try_codegen_constant_if( &mut self, condition: ValueId, if_expr: &If, ) -> Option<Result<Tree<Value>, RuntimeError>>

If the condition is known, skip codegen for the then/else branch and only compile the relevant branch.

fn codegen_match( &mut self, match_expr: &Match, ) -> Result<Tree<Value>, RuntimeError>

fn variant_index_value( &mut self, constructor: &Constructor, typ: NumericType, ) -> Result<ValueId, RuntimeError>

fn no_match( &mut self, variable: Tree<Value>, case: &MatchCase, ) -> Result<Tree<Value>, RuntimeError>

fn enum_tag(&mut self, enum_value: &Tree<Value>) -> ValueId

Extracts the tag value from an enum. Assumes enums are represented as a tuple where the tag is always the first field of the tuple.

If the enum is only a single Leaf value, this expects the enum to consist only of the tag value.

fn bind_case_arguments(&mut self, enum_value: Tree<Value>, case: &MatchCase)

Bind the given variable ids to each argument of the given enum, using the variant at the given variant index. Note that this function makes assumptions that the representation of an enum is:

( tag_value, (field0_0, .. field0_N), // fields of variant 0, (field1_0, .. field1_N), // fields of variant 1, .., (fieldM_0, .. fieldM_N), // fields of variant N, )

fn bind_enum_case_arguments( &mut self, enum_value: Tree<Value>, case: &MatchCase, )

fn bind_tuple_or_struct_case_arguments( &mut self, struct_value: Tree<Value>, case: &MatchCase, )

fn codegen_tuple( &mut self, tuple: &[Expression], ) -> Result<Tree<Value>, RuntimeError>

fn codegen_extract_tuple_field( &mut self, tuple: &Expression, field_index: usize, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_call(&mut self, call: &Call) -> Result<Tree<Value>, RuntimeError>

Generate SSA for a function call. Note that calls to built-in functions and intrinsics are also represented by the function call instruction.

fn codegen_intrinsic_call_checks( &mut self, function: ValueId, arguments: &[ValueId], location: Location, )

fn codegen_let(&mut self, let_expr: &Let) -> Result<Tree<Value>, RuntimeError>

Generate SSA for the given variable. If the variable is immutable, no special handling is necessary and we can return the given ValueId directly. If it is mutable, we’ll need to allocate space for the value and store the initial value before returning the allocate instruction.

fn codegen_constrain( &mut self, expr: &Expression, location: Location, assert_payload: &Option<Box<(Expression, HirType)>>, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_constrain_error( &mut self, assert_message: &Option<Box<(Expression, HirType)>>, ) -> Result<Option<ConstrainError>, RuntimeError>

fn codegen_assign( &mut self, assign: &Assign, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_semi( &mut self, expr: &Expression, ) -> Result<Tree<Value>, RuntimeError>

fn codegen_break(&mut self) -> Result<Tree<Value>, RuntimeError>

fn codegen_continue(&mut self) -> Result<Tree<Value>, RuntimeError>

fn codegen_clone( &mut self, expr: &Expression, ) -> Result<Tree<Value>, RuntimeError>

Evaluate the given expression, increment the reference count of each array within, and return the evaluated expression.

fn codegen_drop( &mut self, expr: &Expression, ) -> Result<Tree<Value>, RuntimeError>

Evaluate the given expression, decrement the reference count of each array within, and return unit.

fn codegen_unless_break_or_continue<T, F>( &mut self, result: Result<T, RuntimeError>, f: F, ) -> Result<(), RuntimeError>
where F: FnOnce(&mut Self, T),