noirc_evaluator::ssa::ssa_gen::program
Struct Ssa

pub struct Ssa {
    pub functions: BTreeMap<Id<Function>, Function>,
    pub main_id: Id<Function>,
    pub next_id: AtomicCounter<Function>,
    entry_point_to_generated_index: BTreeMap<Id<Function>, u32>,
    pub error_selector_to_type: BTreeMap<ErrorSelector, Type>,
}

Contains the entire SSA representation of the program.

Fields
functions: BTreeMap<Id<Function>, Function>main_id: Id<Function>next_id: AtomicCounter<Function>entry_point_to_generated_index: BTreeMap<Id<Function>, u32>

Maps SSA entry point function ID -> Final generated ACIR artifact index. There can be functions specified in SSA which do not act as ACIR entry points. This mapping is necessary to use the correct function pointer for an ACIR call, as the final program artifact will be a list of only entry point functions.

error_selector_to_type: BTreeMap<ErrorSelector, Type>
Implementations

impl Ssa

pub fn into_acir( self, brillig: &Brillig, brillig_options: &BrilligOptions, ) -> Result<Artifacts, RuntimeError>

impl Ssa

pub fn to_brillig(&self, options: &BrilligOptions) -> Brillig

Compile Brillig functions and ACIR functions reachable from them.

impl Ssa

pub(crate) fn check_for_underconstrained_values(&mut self) -> Vec<SsaReport>

This function provides an SSA pass that detects if the final function has any subgraphs independent from inputs and outputs. If this is the case, then part of the final circuit can be completely replaced by any other passing circuit, since there are no constraints ensuring connections. Go through each top-level non-Brillig function and detect if it has independent subgraphs

pub(crate) fn check_for_missing_brillig_constraints( &mut self, enable_lookback: bool, ) -> Vec<SsaReport>

Detect Brillig calls left unconstrained with manual asserts and return a vector of bug reports if any have been found

impl Ssa

pub fn interpret(&self, args: Vec<Value>) -> IResults

pub fn interpret_with_options<W: Write>( &self, args: Vec<Value>, options: InterpreterOptions, output: W, ) -> IResults

fn interpret_function<W: Write>( &self, function: Id<Function>, args: Vec<Value>, options: InterpreterOptions, output: W, ) -> IResults

impl Ssa

pub fn print_without_locations(&self) -> Printer<'_>

pub fn print_with<'local>( &'local self, files: Option<&'local FileManager>, ) -> Printer<'local>

impl Ssa

pub(crate) fn array_set_optimization(self) -> Self

Finds the last instruction that writes to an array and modifies it to do an in-place mutation instead of making a copy if there are no potential shared references to it.

impl Ssa

pub(crate) fn as_vector_optimization(self) -> Self

A simple SSA pass to find any calls to Intrinsic::AsVector and replacing any references to the length of the resulting vector with the length of the array from which it was generated.

This allows the length of a vector generated from an array to be used in locations where a constant value is necessary when the value of the array is unknown.

Note that this pass must be placed before loop unrolling to be useful.

impl Ssa

pub(crate) fn flatten_basic_conditionals(self) -> Ssa

Apply the basic_conditional pass to all functions of the program. It first retrieve the no_predicates attribute of each function which will be used during the flattening.

impl Ssa

pub(crate) fn brillig_array_get_and_set(self) -> Ssa

See brillig_array_get_and_set module for more information.

impl Ssa

pub(crate) fn brillig_entry_point_analysis(self) -> Ssa

impl Ssa

pub(crate) fn check_u128_mul_overflow(self) -> Ssa

See check_u128_mul_overflow module for more information.

impl Ssa

pub(crate) fn checked_to_unchecked(self) -> Ssa

See checked_to_unchecked module for more information.

impl Ssa

pub(crate) fn fold_constants(self, max_iter: usize) -> Ssa

Performs constant folding on each instruction.

It will not look at constraints to inform simplifications based on the stated equivalence of two instructions.

See constant_folding module for more information.

pub(crate) fn fold_constants_using_constraints(self, max_iter: usize) -> Ssa

Performs constant folding on each instruction.

Also uses constraint information to inform more optimizations.

See constant_folding module for more information.

pub fn fold_constants_with_brillig(self, max_iter: usize) -> Ssa

Performs constant folding on each instruction while also replacing calls to brillig functions with all constant arguments by trying to evaluate those calls.

impl Ssa

pub(crate) fn defunctionalize(self) -> Ssa

See defunctionalize module for more information.

impl Ssa

pub(crate) fn prune_dead_parameters( self, unused_parameters: &HashMap<Id<Function>, HashMap<BasicBlockId, Vec<ValueId>>>, ) -> Self

See prune_dead_parameters module for more information.

fn rewrite_calls_to_pruned_entry_blocks( &mut self, call_graph: &CallGraph, entry_block_keep_lists: &HashMap<Id<Function>, Vec<bool>>, )

Helper to rewrite Call instructions targeting functions whose entry block parameters were pruned.

impl Ssa

pub fn dead_instruction_elimination(self) -> Ssa

Performs Dead Instruction Elimination (DIE) to remove any instructions with unused results.

This step should come after the flattening of the CFG and mem2reg.

pub(crate) fn dead_instruction_elimination_pre_flattening(self) -> Ssa

The elimination of certain unused instructions assumes that the DIE pass runs after the flattening of the CFG, but if that’s not the case then we should not eliminate them just yet.

fn dead_instruction_elimination_with_pruning(self, flattened: bool) -> Ssa

fn dead_instruction_elimination_inner(self, flattened: bool) -> (Ssa, DIEResult)

pub(crate) fn dead_instruction_elimination_post_check(&self, flattened: bool)

Sanity check on the final SSA, panicking if the assumptions don’t hold.

Done as a separate step so that we can put it after other passes which provide concrete feedback about where the problem with the Noir code might be, such as dynamic indexing of arrays with references in ACIR. We can look up the callstack of the offending instruction here as well, it’s just not clear what error message to return, besides the fact that mem2reg was unable to eliminate something.

impl Ssa

pub(crate) fn evaluate_static_assert_and_assert_constant( self, ) -> Result<Ssa, RuntimeError>

See evaluate_static_assert_and_assert_constant module for more information.

impl Ssa

pub(crate) fn expand_signed_checks(self) -> Ssa

Expands signed arithmetic operations to include explicit overflow checks.

See expand_signed_checks module for more information.

impl Ssa

pub(crate) fn expand_signed_math(self) -> Ssa

Expands signed “less than”, “div” and “mod” operations in ACIR to be done using unsigned operations.

See expand_signed_math module for more information.

impl Ssa

pub(crate) fn flatten_cfg(self) -> Ssa

Flattens the control flow graph of main such that the function is left with a single block containing all instructions and no more control-flow.

This pass will modify any instructions with side effects in particular, often multiplying them by jump conditions to maintain correctness even when all branches of a jmpif are inlined. For more information, see the module-level comment at the top of this file.

impl Ssa

pub(crate) fn inline_simple_functions(self: Ssa) -> Result<Ssa, RuntimeError>

See the inline_simple_functions module for more information.

impl Ssa

pub(crate) fn inline_functions( self, aggressiveness: i64, small_function_max_instructions: usize, ) -> Result<Ssa, RuntimeError>

Inline all functions within the IR.

In the case of recursive Acir functions, this will attempt to recursively inline until the RECURSION_LIMIT is reached.

Functions are recursively inlined into main until either we finish inlining all functions or we encounter a function whose function id is not known. When the later happens, the call instruction is kept in addition to the function it refers to. The function it refers to is kept unmodified without any inlining changes. This is because if the function’s id later becomes known by a later pass, we would need to re-run all of inlining anyway to inline it, so we might as well save the work for later instead of performing it twice.

There are some attributes that allow inlining a function at a different step of codegen. Currently this is just InlineType::NoPredicates for which we have a flag indicating whether treating that inline functions. The default is to treat these functions as entry points.

This step should run after runtime separation, since it relies on the runtime of the called functions being final.

pub(crate) fn inline_functions_with_no_predicates( self, aggressiveness: i64, small_function_max_instructions: usize, ) -> Result<Ssa, RuntimeError>

Run the inlining pass where functions marked with InlineType::NoPredicates as not entry points

pub(crate) fn inline_until_fixed_point( self, aggressiveness: i64, small_function_max_instructions: usize, inline_no_predicates_functions: bool, ) -> Result<Ssa, RuntimeError>

Inline functions repeatedly until no new functions are inlined.

fn inline_functions_inner( self, inline_infos: &BTreeMap<Id<Function>, InlineInfo>, bottom_up: &[Id<Function>], ) -> Result<Ssa, RuntimeError>

Inline entry points in the order of appearance in inline_infos, assuming it goes in bottom-up order.

impl Ssa

pub(crate) fn loop_invariant_code_motion(self) -> Ssa

See loop_invariant module for more information.

impl Ssa

pub(crate) fn make_constrain_not_equal(self) -> Ssa

Replaces Instruction::Constrain asserting two values are not equal with Instruction::ConstrainNotEqual.

See the make_constrain_not_equal module for more information.

impl Ssa

pub(crate) fn mem2reg(self) -> Ssa

Attempts to remove any load instructions that recover values that are already available in scope, and attempts to remove stores that are subsequently redundant.

impl Ssa

pub fn normalize_ids(&mut self)

This is a debugging pass which re-inserts each instruction and block in a fresh DFG context for each function so that ValueIds, BasicBlockIds, and FunctionIds are always identical for the same SSA code.

During normal compilation this is often not the case since prior passes may increase the ID counter so that later passes start at different offsets, even if they contain the same SSA code.

impl Ssa

pub(crate) fn preprocess_functions( self, aggressiveness: i64, small_function_max_instructions: usize, ) -> Result<Ssa, RuntimeError>

Run pre-processing steps on functions in isolation.

impl Ssa

pub(crate) fn purity_analysis(self) -> Ssa

Analyze the purity of each function and tag each function call with that function’s purity. This is purely an analysis pass on its own but can help future optimizations.

There is no constraint on when this pass needs to be run, but it is generally more beneficial to perform this pass before inlining or loop unrolling so that it can:

Run faster by processing fewer instructions.

Be run earlier in the pass list so that more passes afterward can use the results of this pass.

Performing this pass after defunctionalization may also help more function calls be identified as calling known pure functions.

impl Ssa

pub(crate) fn remove_paired_rc(self) -> Ssa

This pass removes inc_rc and dec_rc instructions as long as there are no array_set instructions to an array of the same type in between.

Note that this pass is very conservative since the array_set instruction does not need to be to the same array. This is because the given array may alias another array (e.g. function parameters or a loaded array from a reference).

impl Ssa

pub(crate) fn remove_bit_shifts(self) -> Ssa

Go through every ACIR function replacing bit shifts with more primitive arithmetic operations,

impl Ssa

pub(crate) fn remove_enable_side_effects(self) -> Ssa

See remove_enable_side_effects module for more information.

impl Ssa

pub(crate) fn remove_if_else(self) -> RtResult<Ssa>

Replaces all Instruction::IfElse instructions with the result of a value merger of the then and else values. The specifics of the value merger depends on the type but is expected to be an equivalent value to the IfElse. For example, on integers, the merger will be: then_condition * then_value + !then_condition * else_value which should zero out the branch that was not taken.

In general this is not possible for all types - notably references - which is why the Noir frontend does not allow references to be returned from if expressions.

Also note that Instruction::IfElse are first inserted after the flattening pass, so before then this pass will have no effect.

impl Ssa

pub(crate) fn remove_truncate_after_range_check(self) -> Self

This SSA pass removes truncate instructions that happen on values that have a range_check on them, where the checked range is less or equal than the bits to truncate (the truncate isn’t needed then as it won’t change the underlying value).

impl Ssa

pub(crate) fn remove_unreachable_functions(self) -> Self

See remove_unreachable module for more information.

impl Ssa

pub(crate) fn remove_unreachable_instructions(self) -> Ssa

impl Ssa

pub(crate) fn remove_unused_instructions(self) -> Self

Remove instructions in all functions which aren’t used in any blocks.

impl Ssa

pub(crate) fn simplify_cfg(self) -> Self

See simplify_cfg module for more information

impl Ssa

pub(crate) fn unroll_loops_iteratively( self, max_bytecode_increase_percent: Option<i32>, ) -> Result<Ssa, RuntimeError>

Loop unrolling can return errors, since ACIR functions need to be fully unrolled. This meta-pass will keep trying to unroll loops and simplifying the SSA until no more errors are found.

The max_bytecode_incr_pct, when given, is used to limit the growth of the Brillig bytecode size after unrolling small loops to some percentage of the original loop. For example a value of 150 would mean the new loop can be 150% (ie. 2.5 times) larger than the original loop. It will still contain fewer SSA instructions, but that can still result in more Brillig opcodes.

impl Ssa

pub fn from_str(src: &str) -> Result<Ssa, SsaErrorWithSource>

Creates an Ssa object from the given string.

pub fn from_str_no_validation(src: &str) -> Result<Ssa, SsaErrorWithSource>

Creates an Ssa object from the given string without running SSA validation

pub fn from_str_simplifying(src: &str) -> Result<Ssa, SsaErrorWithSource>

Creates an Ssa object from the given string but trying to simplify each parsed instruction as it’s inserted into the final SSA.

fn from_str_impl( src: &str, simplify: bool, validate: bool, ) -> Result<Ssa, SsaErrorWithSource>

impl Ssa

pub fn new( functions: Vec<Function>, error_types: BTreeMap<ErrorSelector, HirType>, ) -> Self

Create a new Ssa object from the given SSA functions. The first function in this vector is expected to be the main function.

pub(crate) fn main(&self) -> &Function

Returns the entry-point function of the program

pub(crate) fn add_fn( &mut self, build_with_id: impl FnOnce(Id<Function>) -> Function, ) -> Id<Function>

Adds a new function to the program

pub(crate) fn generate_entry_point_index(self) -> Self

pub(crate) fn get_entry_point_index( &self, func_id: &Id<Function>, ) -> Option<u32>

pub(crate) fn is_entry_point(&self, function: Id<Function>) -> bool

pub(crate) fn used_globals_in_functions( &self, ) -> HashMap<Id<Function>, HashSet<ValueId>>

Collect all the global value IDs used per function.

impl Ssa

pub(crate) fn verify_no_dynamic_indices_to_references(self) -> RtResult<Ssa>

Verifies there are no array_get or array_set instructions remaining with dynamic indices where the element type may contain a reference type. This effectively bans dynamic-indexing of arrays with reference elements since we cannot guarantee we optimize references out of the program in that case.

This pass expects to be run late in the Ssa pipeline such that all array indices used are either constant or derived from an input to the program. E.g. we expect optimizations from inlining, flattening, and mem2reg to already be complete.