noirc_evaluator::ssa::ssa_gen::value
Enum Tree

pub(super) enum Tree<T> {
    Branch(Vec<Tree<T>>),
    Leaf(T),
}

A general Tree structure which is used in the SSA generation pass to represent both values and types which may be tuples.

Since the underlying SSA intermediate representation (IR) does not support tuples directly, theyโre instead represented as Tree::Branch nodes. For example, a single ssa value may be a Tree::Leaf(Value), while a tuple would be a Tree::Branch(values).

Variants

Branch(Vec<Tree<T>>)

Leaf(T)

Implementations

impl<T> Tree<T>

pub(super) fn empty() -> Self

Returns an empty tree node represented by a Branch with no branches

pub(super) fn flatten(self) -> Vec<T>

Flattens the tree into a vector of each leaf value

pub(super) fn count_leaves(&self) -> usize

Returns the total amount of leaves in this tree

pub(super) fn for_each(self, f: impl FnMut(T))

Iterates over each Leaf node, calling f on each value within.

fn for_each_helper(self, f: &mut impl FnMut(T))

pub(super) fn map<U>(self, f: impl FnMut(T) -> Tree<U>) -> Tree<U>

Calls the given function on each leaf node, mapping this tree into a new one.

Because the given function returns a Tree<U> rather than a U, it is possible to use this function to turn Leaf nodes into either other Leaf nodes or even Branch nodes.

fn map_helper<U>(self, f: &mut impl FnMut(T) -> Tree<U>) -> Tree<U>

pub(super) fn map_both<U, R>( &self, other: Tree<U>, f: impl FnMut(T, U) -> Tree<R>, ) -> Tree<R>
where T: Debug + Clone, U: Debug,

Map two trees alongside each other. This asserts each tree has the same internal structure.

fn map_both_helper<U, R>( &self, other: Tree<U>, f: &mut impl FnMut(T, U) -> Tree<R>, ) -> Tree<R>
where T: Debug + Clone, U: Debug,

pub(super) fn into_leaf(self) -> T

Unwraps this Tree into the value of the leaf node. Panics if this Tree is a Branch

impl Tree<Type>

pub(super) fn size_of_type(&self) -> usize

Returns the size of the type in terms of the number of FieldElements it contains. Non-field types like functions and references are also counted as 1 FieldElement.

impl Tree<Value>

pub(super) fn into_value_list( self, ctx: &mut FunctionContext<'_>, ) -> Vec<IrValueId>

Flattens and evaluates this Tree<Value> into a list of ir values for return statements, branching instructions, or function parameters.