Crate noirc_evaluator

Re-exports

pub use ssa::create_program;

pub use ssa::ir::instruction::ErrorType;

Modules

acir

The acir module contains all the logic necessary for noirc’s ACIR-gen pass which generates the output ACIR program.

brillig

The brillig module contains all logic necessary for noirc’s Brillig-gen pass for generating Brillig bytecode from Ssa.

errors

Noir Evaluator has two types of errors

ssa

SSA stands for Single Static Assignment The IR presented in this module will already be in SSA form and will be used to apply conventional optimizations like Common Subexpression elimination and constant folding.

Macros

assert_ssa_snapshot

Compare the textural representation of the SSA after normalizing its IDs to a snapshot.