noirc_evaluator
Module acir

This file holds the pass to convert from Noir’s SSA IR to ACIR.

Modules

acir_variable 🔒

big_int 🔒

black_box 🔒

brillig_call 🔒

brillig_directive 🔒

generated_acir 🔒

GeneratedAcir is constructed as part of the acir_gen pass to accumulate all of the ACIR program as it is being converted from SSA form.

Structs

AcirDynamicArray 🔒

Context 🔒

Context struct for the acir generation pass. May be similar to the Evaluator struct in the current SSA IR.

SharedContext 🔒

Enums

AcirValue 🔒

Functions

can_omit_element_sizes_array 🔒

Type Aliases

Artifacts 🔒