noirc_evaluator
Module acir

The acir module contains all the logic necessary for noirc’s ACIR-gen pass which generates the output ACIR program.

Usage

ACIR generation is performed by calling the Ssa::into_acir method, providing any necessary brillig bytecode. The compiled program will be returned as an Artifacts type.

Modules

acir_context 🔒

The acir_context module acts as a high-level wrapper around generated_acir, maintaining a small amount of type information from SSA in order to allow efficient ACIR-gen.

arrays 🔒

Array handling in ACIR.

call 🔒

shared_context 🔒

The shared_context module contains any shared state needed while lowering SSA → ACIR. Specifically, it manages Brillig function pointers, stdlib calls, and global usage.

ssa 🔒

types 🔒

Structs

Context 🔒

Context struct for the acir generation pass. May be similar to the Evaluator struct in the current SSA IR.

GeneratedAcir

The output of the Acir-gen pass, which should only be produced for entry point Acir functions

Functions

acir_post_check 🔒

Check post ACIR generation properties

Type Aliases

Artifacts