noirc_evaluator::acir::acir_context::generated_acir
Module brillig_directive

Structs

BrilligStdLib 🔒

Enums

BrilligStdlibFunc

Constants

PLACEHOLDER_BRILLIG_INDEX 🔒

Brillig calls such as for the Brillig std lib are resolved only after code generation is finished. This index should be used when adding a Brillig call during code generation. Code generation should then keep track of that unresolved call opcode which will be resolved with the correct function index after code generation.

Functions

directive_invert 🔒

Generates brillig bytecode which computes the inverse of its input if not null, else zero.

directive_quotient 🔒

Generates brillig bytecode which computes a / b and returns the quotient and remainder.

directive_to_radix 🔒

Generates brillig bytecode which performs a radix-base decomposition of a The brillig inputs are ‘a’, the numbers of limbs and the radix