noirc_evaluator::acir::arrays
Function array_has_constant_element_size

pub(super) fn array_has_constant_element_size(array_typ: &Type) -> Option<u32>

Returns whether the array’s elements have a constant size.

This is useful as it then allows us to calculate the flattened index by multiplying by this constant size.

Returns

If the array’s element types are all the same size then array_has_constant_element_size will return Some(element_size) where element_size is the size of array’s elements. Otherwise returns None.