noirc_evaluator::acir
Function acir_post_check

fn acir_post_check(context: &Context<'_>, acir: &GeneratedAcir<FieldElement>)

Check post ACIR generation properties

No memory opcodes should be laid down that write to the internal type sizes array. See arrays for more information on the type sizes array.