noirc_evaluator
Module brillig

The brillig module contains all logic necessary for noirc’s Brillig-gen pass for generating Brillig bytecode from Ssa.

Usage

Brillig generation is performed by calling the Ssa::to_brillig method. All compiled Brillig artifacts will be returned as the Brillig context structure.

Modules

brillig_check 🔒

Post-codegen checks that can highlight potential issues with the generated bytecode.

brillig_gen 🔒

The code generation logic for converting crate::ssa objects into their respective Brillig artifacts.

brillig_ir

This module is an abstraction layer over Brillig. To allow for separation of concerns, it knows nothing about SSA types, and can therefore be tested independently. brillig_gen is therefore the module which combines both ssa types and types in this module. A similar paradigm can be seen with the acir_ir module.

Structs

Brillig

Context structure for the Brillig pass. It stores Brillig-related data required for Brillig generation.

BrilligOptions

Options that affect Brillig code generation.

Enums

ProcedureId

Procedures are a set of complex operations that are common in the noir language. Extracting them to reusable procedures allows us to reduce the size of the generated Brillig. Procedures receive their arguments on scratch space to avoid stack dumping&restoring.