noirc_evaluator::brillig
Module brillig_check

Post-codegen checks that can highlight potential issues with the generated bytecode.

Structs

AdvisoryCollector 🔒

Collect advisories about the opcodes in a single block.

ReadCollector 🔒

Collect all the relative addresses read in a range.

Enums

OpcodeAdvisory 🔒

Traits

OpcodeAddressVisitor 🔒

Functions

block_opcode_ranges 🔒

Go through the labels in the BrilligContext and collect the opcode locations where each block starts and ends.

opcode_advisories 🔒

Go through opcodes and collect advisories, indicating opcodes which are potentially unused.

show_opcode_advisories 🔒

Display the opcodes and their corresponding advisories.

Type Aliases

OpcodeAdvisories 🔒