noirc_evaluator::brillig::brillig_gen
Module brillig_globals

Codegen for converting SSA globals to Brillig bytecode.

Structs

BrilligGlobals 🔒

Final context structure for generating Brillig globals. It stores globals related data required for the code generation of regular Brillig functions.

BrilligGlobalsArtifact 🔒

A globals artifact containing all information necessary for utilizing globals from SSA during Brillig code generation.

BrilligGlobalsInit 🔒

Initial context structure for generating Brillig globals.

CallMap 🔒

Map entry points to functions reachable from them and vice versa.

ConstantAllocationCache 🔒

Type Aliases

ConstantCounterMap 🔒

Mapping of a constant value to the number of functions in which it occurs.

HoistedConstantsToBrilligGlobals 🔒

Mapping of constant values shared across functions hoisted to the global memory space.

SsaToBrilligGlobals 🔒

Mapping of SSA value ids to their Brillig allocations.