noirc_evaluator::brillig::brillig_ir
Struct BrilligContext

pub(crate) struct BrilligContext<F, Registers> {
    obj: BrilligArtifact<F>,
    registers: Rc<RefCell<Registers>>,
    context_label: Label,
    current_section: usize,
    next_section: usize,
    debug_show: DebugShow,
    can_call_procedures: bool,
    enable_debug_assertions: bool,
    count_arrays_copied: bool,
    globals_memory_size: Option<usize>,
}

Brillig context object that is used while constructing the Brillig bytecode.

Fields
obj: BrilligArtifact<F>registers: Rc<RefCell<Registers>>

Tracks register allocations

context_label: Label

Context label, must be unique with respect to the function being linked.

current_section: usize

Section label, used to separate sections of code within a context.

next_section: usize

Stores the next available section.

debug_show: DebugShow

IR printer

can_call_procedures: bool

Whether this context can call procedures or not. This is used to prevent a procedure from calling another procedure.

enable_debug_assertions: bool

Insert extra assertions that we expect to be true, at the cost of larger bytecode size.

count_arrays_copied: bool

Count the number of arrays that are copied, and output this to stdout

globals_memory_size: Option<usize>
Implementations

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_array_copy_procedure( &mut self, source_array: BrilligArray, destination_array: BrilligArray, )

Call ProcedureId::ArrayCopy.

Conditionally copies a source array to a destination array. If the reference count of the source array is 1, then we can directly copy the pointer of the source array to the destination array. Otherwise a copy is made, and the ref-count of the original is decreased by 1.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn emit_println_of_array_copy_counter(&mut self)

emit: println(f"Total arrays copied: {array_copy_counter}")

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_array_reverse_procedure( &mut self, pointer: MemoryAddress, size: MemoryAddress, )

Copy the arguments to the scratch space and call ProcedureId::ArrayReverse.

Reverses the size number of items pointed to by pointer in-place.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_check_max_stack_depth_procedure(&mut self)

Call ProcedureId::CheckMaxStackDepth.

Ensures that the stack size does not grow beyond its boundaries, where it would encroach on the heap.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_error_with_string_procedure(&mut self, error_string: String)

Reverts with the error selector of the given string This procedure is useful to deduplicate generating code for the same selector.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_mem_copy_procedure( &mut self, source_pointer: MemoryAddress, destination_pointer: MemoryAddress, num_elements_variable: MemoryAddress, )

Copy the input arguments to the ScratchSpace, then emit an opcode to call ProcedureId::MemCopy.

Copies num_elements_variable number of items on the heap from source_pointer to destination_pointer.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_prepare_vector_insert_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, index: SingleAddrVariable, write_pointer: MemoryAddress, item_count: usize, )

Copy the arguments to the ScratchSpace and call ProcedureId::PrepareVectorInsert.

Prepares a the source_vector for a insert operation by making a copy to destination_vector, leaving an item_count hole at the index position, which is returned as the write_pointer.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_prepare_vector_push_procedure( &mut self, source_len: SingleAddrVariable, source_vector: BrilligVector, destination_vector: BrilligVector, write_pointer: MemoryAddress, item_push_count: usize, back: bool, )

Copy arguments to the ScratchSpace and call ProcedureId::PrepareVectorPush.

Prepares a vector for a push operation, allocating a larger vector and copying the source vector into the destination vector if necessary. It returns the destination vector and the write pointer to where to put the new items.

Brillig vectors have an inherent size and capacity, but their semantic length can be different, and it’s passed as a separate variable in SSA. When ACIR flattens vectors of unequal length, it can be different from the size and capacity in the vector data structure itself.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_copy_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, )

Copy arguments to ScratchSpace and call ProcedureId::VectorCopy.

Conditionally copies a source vector to a destination vector. If the reference count of the source vector is 1, then we can directly copy the pointer of the source vector to the destination vector.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_pop_back_procedure( &mut self, source_len: SingleAddrVariable, source_vector: BrilligVector, destination_vector: BrilligVector, read_pointer: MemoryAddress, item_pop_count: usize, )

Copy arguments to ScratchSpace and call ProcedureId::VectorPopBack.

Pops item_pop_count items from the back of a source_vector, returning the new destination_vector and the pointer to the popped items in read_pointer.

The item_pop_count equals the number of variables to pop into in SSA. The source_len is the semantic (not flattened) length of the vector.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_pop_front_procedure( &mut self, source_len: SingleAddrVariable, source_vector: BrilligVector, destination_vector: BrilligVector, item_pop_count: usize, )

Copy arguments to ScratchSpace and call ProcedureId::VectorPopFront.

Pops items from the front of a vector, returning the new vector. The procedure assumes that there are constraints to prevent 0 size vectors from being popped.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn call_vector_remove_procedure( &mut self, source_vector: BrilligVector, destination_vector: BrilligVector, index: SingleAddrVariable, item_count: usize, )

Copy arguments to ScratchSpace and call ProcedureId::VectorRemove.

Removes item_count items from the source_vector at index, returning the new destination_vector. Modifies the source_vector if the reference count is 1.

The procedure assumes that we have constraints on the index being within the semantic length of the vector.

impl<F, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn registers(&self) -> Ref<'_, Registers>

Read-only reference to the registers.

pub(crate) fn registers_mut(&self) -> RefMut<'_, Registers>

Read-write reference to the registers.

pub(crate) fn deallocate_register(&mut self, register: MemoryAddress)

Manually deallocates a register which is no longer in use.

This could be one of the pre-allocated ones which doesn’t have automation.

pub(crate) fn allocate_register( &mut self, ) -> Allocated<MemoryAddress, Registers>

Allocates an unused register.

pub(crate) fn set_allocated_registers( &mut self, allocated_registers: Vec<MemoryAddress>, )

Resets the registers to a new list of allocated ones.

pub(crate) fn allocate_single_addr( &mut self, bit_size: u32, ) -> Allocated<SingleAddrVariable, Registers>

Allocate a SingleAddrVariable.

pub(crate) fn allocate_single_addr_usize( &mut self, ) -> Allocated<SingleAddrVariable, Registers>

Allocate a SingleAddrVariable with the size of a Brillig memory address.

pub(crate) fn allocate_single_addr_bool( &mut self, ) -> Allocated<SingleAddrVariable, Registers>

Allocate a SingleAddrVariable with a size of 1 bit.

pub(crate) fn allocate_single_addr_mem( &mut self, ) -> Allocated<SingleAddrVariable, Registers>

Allocate a SingleAddrVariable with a size of BRILLIG_MEMORY_ADDRESSING_BIT_SIZE bit.

pub(crate) fn allocate_brillig_vector( &mut self, ) -> Allocated<BrilligVector, Registers>

Allocate a pointer for BrilligVector.

This does not include allocating memory for the data on the heap or shaping the meta-data. That is done by BrilligContext::codegen_initialize_vector.

pub(crate) fn allocate_brillig_array( &mut self, size: SemiFlattenedLength, ) -> Allocated<BrilligArray, Registers>

Allocate a pointer for BrilligArray.

This does not include allocating memory for the data on the heap or shaping the meta-data. That is done by BrilligContext::codegen_initialize_array.

pub(crate) fn allocate_heap_vector( &mut self, ) -> Allocated<HeapVector, Registers>

Allocate a [HeapVector].

pub(crate) fn allocate_heap_array( &mut self, size: SemiFlattenedLength, ) -> Allocated<HeapArray, Registers>

Allocate a [HeapArray].

pub(crate) fn make_scratch_registers<const N: usize>( &self, ) -> [MemoryAddress; N]

Create a number of consecutive [MemoryAddress::Direct] addresses at the start of the ScratchSpace.

impl<F> BrilligContext<F, ScratchSpace>

pub(crate) fn allocate_scratch_registers<const N: usize>( &mut self, ) -> [MemoryAddress; N]

Allocate a number of consecutive scratch registers and replace the current allocator with one that has the new registers pre-allocated.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_usize_op_in_place( &mut self, destination: MemoryAddress, op: BrilligBinaryOp, constant: usize, )

Utility method to perform a binary instruction with a constant value in place

pub(crate) fn codegen_usize_op( &mut self, operand: MemoryAddress, destination: MemoryAddress, op: BrilligBinaryOp, constant: usize, )

Utility method to perform a binary instruction with a constant value

pub(crate) fn codegen_increment_array_copy_counter(&mut self)

pub(crate) fn codegen_usize_equals_one( &mut self, operand: SingleAddrVariable, ) -> Allocated<SingleAddrVariable, Registers>

Utility method to check if the value at a memory address equals one.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_call( &mut self, func_id: Id<Function>, arguments: &[BrilligVariable], returns: &[BrilligVariable], )

Generate Brillig opcodes to:

calculate the current stack size

copy the current stack pointer and the call arguments into a new stack frame

update the stack pointer and execute the call

restore the stack pointer and copy the results into the return variables

pub(crate) fn codegen_return(&mut self, return_variables: &[BrilligVariable])

Codegens a return from the current function.

Takes the variables with the addresses of the values that need to be returned. The values are copied to the beginning of the stack space, into an equal number of slots.

Any potential overlap between the source of the return variables and the final destination on the beginning of the stack is handled by Self::codegen_mov_registers_to_registers.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_generic_iteration<T>( &mut self, make_iterator: impl FnOnce(&mut BrilligContext<F, Registers>) -> T, update_iterator: impl FnOnce(&mut BrilligContext<F, Registers>, &T), make_finish_condition: impl FnOnce(&mut BrilligContext<F, Registers>, &T) -> Allocated<SingleAddrVariable, Registers>, on_iteration: impl FnOnce(&mut BrilligContext<F, Registers>, &T), clean_iterator: impl FnOnce(&mut BrilligContext<F, Registers>, T), )

pub(crate) fn codegen_for_loop( &mut self, loop_start: Option<MemoryAddress>, loop_bound: MemoryAddress, step: Option<MemoryAddress>, on_iteration: impl FnOnce(&mut BrilligContext<F, Registers>, SingleAddrVariable), )

This codegen will issue a loop for (let iterator_register = loop_start; i < loop_bound; i += step) The body of the loop should be issued by the caller in the on_iteration closure.

pub(crate) fn codegen_loop( &mut self, iteration_count: MemoryAddress, on_iteration: impl FnOnce(&mut BrilligContext<F, Registers>, SingleAddrVariable), )

This codegen will issue a loop that will iterate from 0 to iteration_count The body of the loop should be issued by the caller in the on_iteration closure.

pub(crate) fn codegen_branch( &mut self, condition: MemoryAddress, f: impl FnMut(&mut BrilligContext<F, Registers>, bool), )

This codegen will issue an if-then branch that will check if the condition is true and if so, perform the instructions given in f(self, true) and otherwise perform the instructions given in f(self, false). A boolean is passed instead of two separate functions to allow the given function to mutably alias its environment.

pub(crate) fn codegen_if( &mut self, condition: MemoryAddress, f: impl FnOnce(&mut BrilligContext<F, Registers>), )

This codegen issues a branch that jumps over the code generated by the given function if the condition is false

pub(crate) fn codegen_if_not( &mut self, condition: MemoryAddress, f: impl FnOnce(&mut BrilligContext<F, Registers>), )

This codegen issues a branch that jumps over the code generated by the given function if the condition is truthy

pub(crate) fn codegen_constrain_with_error_data( &mut self, condition: SingleAddrVariable, error_data_items: Vec<BrilligVariable>, error_data_types: Vec<BrilligParameter>, error_selector: Option<ErrorSelector>, )

Jump to a trap condition if condition is false. The trap will include the given message as error data.

If error_selector is None, an empty error data is generated (no error information).

pub(crate) fn codegen_constrain( &mut self, condition: SingleAddrVariable, assert_message: Option<String>, )

Jump to a trap condition if condition is false, with any assertion message written to the error data.

pub(super) fn flattened_size(param: &BrilligParameter) -> usize

Computes the size of a parameter if it was flattened

pub(super) fn flattened_tuple_size(tuple: &[BrilligParameter]) -> usize

Computes the size of a parameter if it was flattened

pub(crate) fn has_nested_arrays(tuple: &[BrilligParameter]) -> bool

Computes the size of a parameter if it was flattened

pub(super) fn flatten_array( &mut self, item_type: &[BrilligParameter], item_count: SemanticLength, flattened_array_pointer: MemoryAddress, deflattened_items_pointer: MemoryAddress, )

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_truncate( &mut self, destination_of_truncated_value: SingleAddrVariable, value_to_truncate: SingleAddrVariable, bit_size: u32, )

Codegens a truncation of a value to the given bit size

pub(crate) fn codegen_to_radix( &mut self, source_field: SingleAddrVariable, target_array: BrilligArray, radix: SingleAddrVariable, little_endian: bool, output_bits: bool, )

Issues a to_radix instruction. This instruction will write the modulus of the source_field register and the radix register target_array, with the number of limbs given by the size of target_array.

If output_bits is true, it generates bit limbs, otherwise it generates byte limbs. If little_endian is true, then the target_array will contain the results in Little Endian order.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_allocate_immediate_mem( &mut self, pointer_register: MemoryAddress, size: usize, )

Loads the current free memory pointer into pointer_register and increases the free memory pointer by size, thus allocating the required amount of memory starting at the pointer.

pub(crate) fn codegen_allocate_mem( &mut self, pointer_register: MemoryAddress, size_register: MemoryAddress, )

Loads the current free memory pointer into pointer_register and increases the free memory pointer by the value stored at size_register, thus allocating the required amount of memory starting at the pointer.

pub(crate) fn codegen_load_with_offset( &mut self, base_ptr: MemoryAddress, index: SingleAddrVariable, result: MemoryAddress, )

Gets the value stored at base_ptr + index and stores it in result.

pub(crate) fn codegen_store_with_offset( &mut self, base_ptr: MemoryAddress, index: SingleAddrVariable, value: MemoryAddress, )

Stores value at base_ptr + index.

pub(crate) fn codegen_mem_copy( &mut self, source_pointer: MemoryAddress, destination_pointer: MemoryAddress, num_elements_variable: SingleAddrVariable, )

Copies the values of memory pointed by source_pointer with length stored in num_elements_variable after the address pointed by destination_pointer.

pub(crate) fn codegen_mem_copy_from_the_end( &mut self, source_start: MemoryAddress, target_start: MemoryAddress, num_elements_variable: SingleAddrVariable, )

Copies num_elements_variable number of elements from the source_start pointer to the target_start pointer, starting from the end, moving backwards.

By moving back-to-front, it can shift items backwards, modifying a vector in-place to make room in the front.

pub(crate) fn codegen_array_reverse( &mut self, items_pointer: MemoryAddress, size: MemoryAddress, )

Emit opcodes to reverse the order of the size elements pointed by pointer.

It is the responsibility of the caller to ensure that the ref-count of the array is 1.

pub(crate) fn codegen_brillig_array_to_heap_array( &mut self, array: BrilligArray, ) -> Allocated<HeapArray, Registers>

Converts a BrilligArray (pointer to [RC, ...items]) to a [HeapArray] (pointer to [...items]).

pub(crate) fn codegen_brillig_vector_to_heap_vector( &mut self, vector: BrilligVector, ) -> Allocated<HeapVector, Registers>

Converts a BrilligVector (pointer to [RC, size, capacity, ...items]) to a [HeapVector] (two pointers to [...items] and size).

pub(crate) fn variable_to_value_or_array( &mut self, variable: BrilligVariable, ) -> Allocated<ValueOrArray, Registers>

Converts a BrilligVariable to [ValueOrArray].

This can involve allocating new registers and loading values into them from the input data structures.

pub(crate) fn codegen_read_rc( &mut self, pointer: MemoryAddress, ) -> Allocated<SingleAddrVariable, Registers>

Returns a variable holding the ref-count of a given array or vector.

pub(crate) fn codegen_read_vector_rc( &mut self, vector: BrilligVector, ) -> Allocated<SingleAddrVariable, Registers>

Returns a variable holding the ref-count of a given vector.

pub(crate) fn codegen_read_vector_size( &mut self, vector: BrilligVector, ) -> Allocated<SingleAddrVariable, Registers>

Returns a variable holding the size of a given vector.

pub(crate) fn codegen_update_vector_size( &mut self, vector: BrilligVector, new_size: SingleAddrVariable, )

Writes the value of new size to the size pointer of the vector.

pub(crate) fn codegen_read_vector_capacity( &mut self, vector: BrilligVector, ) -> Allocated<SingleAddrVariable, Registers>

Returns a variable holding the capacity of a given vector.

pub(crate) fn codegen_vector_items_pointer( &mut self, vector: BrilligVector, result: MemoryAddress, )

Writes the pointer to the items of a given vector to the result.

pub(crate) fn codegen_make_vector_items_pointer( &mut self, vector: BrilligVector, ) -> Allocated<MemoryAddress, Registers>

Returns a pointer to the items of a given vector.

pub(crate) fn codegen_read_vector_metadata( &mut self, vector: BrilligVector, semantic_length_and_item_size: Option<(MemoryAddress, MemoryAddress)>, ) -> VectorMetaData<Registers>

Reads the metadata of a vector into individual registers and returns them as VectorMetaData.

If the semantic_length_and_item_size is given, then instead of reading the size from the vector data structure, it is calculated as a multiplication of length and item size.

pub(crate) fn codegen_vector_flattened_size( &mut self, destination: MemoryAddress, length: MemoryAddress, item_size: MemoryAddress, )

Generate code to calculate the flattened vector size from its semantic length and the item size.

For example a [(u32, bool)] would have a flattened item size of 2, because each item consists of 2 values. Such a vector with a semantic length of 3 would have a flattened size of 6.

pub(crate) fn codegen_make_array_items_pointer( &mut self, array: BrilligArray, ) -> Allocated<MemoryAddress, Registers>

Returns a pointer to the items of a given array.

pub(crate) fn codegen_make_array_or_vector_items_pointer( &mut self, variable: BrilligVariable, ) -> Allocated<MemoryAddress, Registers>

Returns a pointer to the items of an array or vector.

pub(crate) fn codegen_initialize_array(&mut self, array: BrilligArray)

Initializes an array, allocating memory on the heap to store its representation and initializing the reference counter to 1.

pub(crate) fn codegen_initialize_rc( &mut self, pointer: MemoryAddress, rc: usize, )

Initialize the reference counter for an array or vector. This should only be used internally in the array and vector initialization methods.

The RC is in the 0th slot of the memory allocated for the array or vector.

pub(crate) fn codegen_decrement_rc( &mut self, _pointer: MemoryAddress, _rc: MemoryAddress, )

Decrement the ref-count by 1.

The inputs are:

the pointer to the array/vector

the rc address of the vector where we have the current RC loaded already

pub(crate) fn codegen_increment_rc( &mut self, pointer: MemoryAddress, rc: MemoryAddress, )

Increment the ref-count by 1.

The inputs are:

the pointer to the array/vector

the rc address of the vector where we have the current RC loaded already

the by is a constant by which to increment the RC, typically 1

pub(crate) fn codegen_initialize_vector( &mut self, vector: BrilligVector, size: SingleAddrVariable, capacity: Option<SingleAddrVariable>, )

Initializes a vector, allocating memory on the heap to store its representation and initializing the reference counter, size and capacity.

pub(super) fn codegen_initialize_vector_metadata( &mut self, vector: BrilligVector, size: SingleAddrVariable, capacity: SingleAddrVariable, )

Writes vector metadata (reference count, size, and capacity) into the allocated memory.

Sets the reference count to 1.

pub(crate) fn codegen_initialize_externally_returned_vector( &mut self, vector: BrilligVector, output: &HeapVector, ) -> SingleAddrVariable

Initialize the BrilligVector after the data returned by a foreign call has been written to the heap.

We don’t know the length of a vector returned externally before the call, so we write the size and the data to the free memory pointer.

Here we are adjusting the rest of the meta-data required by the vector structure: basically the RC and the capacity.

Returns the size variable, which we can use to set the semantic length.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn codegen_mov_registers_to_registers( &mut self, sources: &[MemoryAddress], destinations: &[MemoryAddress], )

Map sources to potentially multiple destinations. This function moves values from a set of registers to another set of registers. It assumes that:

every destination needs to be written at most once. Will panic if not.

destinations are relative addresses, starting from 1 and consecutively incremented by one. Will panic if not

sources are any relative addresses, and can also be direct address to the global space. TODO: panic if not

sources and destinations have same length. Will panic if not.

fn process_loop( &mut self, entry: usize, free: &MemoryAddress, num_destinations: &mut [usize], source: &[MemoryAddress], processed: &mut usize, )

Generates mov opcodes corresponding to a loop, given a node from the loop (entry) and a register not in the loop that contains its value (free)

fn perform_movement( &mut self, dest: usize, src: MemoryAddress, num_destinations: &mut [usize], processed: &mut usize, )

Generates a move opcode from ‘src’ to ‘dest’.

impl<F: AcirField + DebugToString> BrilligContext<F, Stack>

pub(crate) fn new_entry_point_artifact( arguments: Vec<BrilligParameter>, return_parameters: Vec<BrilligParameter>, target_function: Id<Function>, globals_init: bool, globals_memory_size: usize, name: &str, options: &BrilligOptions, ) -> (BrilligArtifact<F>, usize)

Creates an entry point artifact that will jump to the function label provided.

fn calldata_start_offset(&self) -> usize

fn return_data_start_offset(&self, calldata_size: usize) -> usize

fn codegen_entry_point( &mut self, arguments: &[BrilligParameter], return_parameters: &[BrilligParameter], ) -> usize

Adds the instructions needed to handle entry point parameters The runtime will leave the parameters in calldata. Arrays will be passed flattened.

Memory layout for entry points: {reserved} {scratch} {globals} {entry point (call data + return data)} {stack} {heap}

Returns

The start of the stack memory region. The start of the stack is determined by the globals compiled as well as the amount of call data and return data. We return this information so that the max stack depth check can check against the appropriate constant represented the max stack pointer we can have in memory.

fn allocate_function_arguments( &mut self, arguments: &[BrilligParameter], ) -> Vec<Allocated<BrilligVariable, Stack>>

fn copy_and_cast_calldata(&mut self, arguments: &[BrilligParameter])

fn deflatten_array( &mut self, item_type: &[BrilligParameter], item_count: SemanticLength, flattened_array_pointer: MemoryAddress, is_vector: bool, ) -> Allocated<MemoryAddress, Stack>

Deflatten an array by recursively allocating nested arrays and copying the plain values. Returns the pointer to the deflattened items.

fn codegen_exit_point( &mut self, arguments: &[BrilligParameter], return_parameters: &[BrilligParameter], )

Adds the instructions needed to handle return parameters The runtime expects the results in a contiguous memory region. Arrays are expected to be returned with all the nested arrays flattened. However, the function called returns variables (that have extra data) and the returned arrays are deflattened.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

Low level instructions of the brillig IR, used by the brillig ir codegens and brillig_gen Printed using debug_slow

pub(crate) fn binary_instruction( &mut self, lhs: SingleAddrVariable, rhs: SingleAddrVariable, result: SingleAddrVariable, operation: BrilligBinaryOp, )

Processes a binary instruction according operation.

This method will compute lhs <operation> rhs and store the result in the result register.

pub(crate) fn not_instruction( &mut self, input: SingleAddrVariable, result: SingleAddrVariable, )

Processes a not instruction.

Not is computed using a subtraction operation as there is no native not instruction in Brillig.

pub(crate) fn memory_op_instruction( &mut self, lhs: MemoryAddress, rhs: MemoryAddress, destination: MemoryAddress, op: BrilligBinaryOp, )

Utility method to perform a binary instruction with a memory address

pub(crate) fn memory_op_inc_by_usize_one(&mut self, address: MemoryAddress)

Increase the value at a pointer in place by 1.

pub(crate) fn conditional_move_instruction( &mut self, condition: SingleAddrVariable, then_address: SingleAddrVariable, else_address: SingleAddrVariable, destination: SingleAddrVariable, )

Insert a conditional move instruction

fn binary( &mut self, lhs: SingleAddrVariable, rhs: SingleAddrVariable, result: SingleAddrVariable, operation: BrilligBinaryOp, )

fn modulo( &mut self, result: SingleAddrVariable, left: SingleAddrVariable, right: SingleAddrVariable, )

Computes left % right by emitting the necessary Brillig opcodes.

This is done by using the following formula:

a % b = a - (b * (a / b))

Brillig does not have an explicit modulo operation, so we must emit multiple opcodes and process it differently to other binary instructions.

fn binary_result_bit_size( operation: BrilligBinaryOp, arguments_bit_size: u32, ) -> u32

pub(crate) fn foreign_call_instruction( &mut self, func_name: String, inputs: &[ValueOrArray], input_value_types: &[HeapValueType], outputs: &[ValueOrArray], output_value_types: &[HeapValueType], )

Push an opcode to processes a foreign call instruction.

Note: the function being called is external and will not be linked during brillig generation.

pub(crate) fn add_external_call_instruction(&mut self, func_id: Id<Function>)

Adds a unresolved external Call instruction to the bytecode. This calls into another function compiled into this brillig artifact.

pub(super) fn add_procedure_call_instruction( &mut self, procedure_id: ProcedureId, )

pub(super) fn add_globals_init_instruction(&mut self, func_id: Id<Function>)

pub(crate) fn jump_instruction(&mut self, target_label: Label)

Adds a unresolved Jump instruction to the bytecode.

pub(crate) fn jump_if_instruction( &mut self, condition: MemoryAddress, target_label: Label, )

Adds a unresolved JumpIf instruction to the bytecode.

fn add_unresolved_jump( &mut self, jmp_instruction: BrilligOpcode<F>, destination: Label, )

Adds a unresolved Jump to the bytecode.

pub(crate) fn enter_context(&mut self, label: Label)

Adds a context and a section label to the next opcode.

Entering a context resets the current section to 0 and the next section to 1.

pub(super) fn enter_section(&mut self, section: usize)

Enter the given section within a basic block.

pub(super) fn reserve_next_section_label(&mut self) -> (usize, Label)

Create, reserve, and return a new section label.

fn compute_section_label(&self, section: usize) -> Label

Internal function used to compute the section labels

fn current_section_label(&self) -> Label

Returns the current section label

pub(crate) fn return_instruction(&mut self)

Emits a return instruction

pub(crate) fn stop_instruction(&mut self, return_data: HeapVector)

Emits a stop instruction with return data

pub(crate) fn black_box_op_instruction(&mut self, op: BlackBoxOp)

Issues a blackbox operation.

pub(crate) fn load_free_memory_pointer_instruction( &mut self, pointer_register: MemoryAddress, )

Push an opcode to load the current value of the free memory pointer into the pointer_register.

pub(crate) fn increase_free_memory_pointer_instruction( &mut self, size_register: MemoryAddress, )

Push an opcode to increase the free memory pointer by the value stored in size_register.

pub(crate) fn store_instruction( &mut self, destination_pointer: MemoryAddress, source: MemoryAddress, )

Push an opcode to store the value at source address into the address under destination_pointer.

pub(crate) fn load_instruction( &mut self, destination: MemoryAddress, source_pointer: MemoryAddress, )

Push an opcode to load a value from the address under source_pointer into the destination address.

pub(crate) fn mov_instruction( &mut self, destination: MemoryAddress, source: MemoryAddress, )

Emits a mov instruction.

Copies the value at source into destination

pub(crate) fn cast_instruction( &mut self, destination: SingleAddrVariable, source: SingleAddrVariable, )

Cast truncates the value to the given bit size and converts the type of the value in memory to that bit size.

pub(crate) fn cast( &mut self, destination: SingleAddrVariable, source: SingleAddrVariable, )

pub(crate) fn const_instruction( &mut self, result: SingleAddrVariable, constant: F, )

Stores the value of constant in the result register.

pub(crate) fn indirect_const_instruction( &mut self, result_pointer: MemoryAddress, bit_size: u32, constant: F, )

Stores the value of constant in the register pointed at by result_pointer.

fn constant( &mut self, result: MemoryAddress, bit_size: u32, constant: F, indirect: bool, )

Pushes a [IndirectConst][BrilligOpcode::IndirectConst] or [Const][BrilligOpcode::Const] opcode.

pub(crate) fn usize_const_instruction( &mut self, result: MemoryAddress, constant: F, )

Stores the value of constant at the result address.

pub(crate) fn make_constant_instruction( &mut self, constant: F, bit_size: u32, ) -> Allocated<SingleAddrVariable, Registers>

Returns a register which holds the value of a constant.

pub(crate) fn make_usize_constant_instruction( &mut self, constant: F, ) -> Allocated<SingleAddrVariable, Registers>

Returns a register which holds the value of an usize constant.

pub(super) fn calldata_copy_instruction( &mut self, destination: MemoryAddress, calldata_size: usize, calldata_offset: usize, )

Pushes a [CalldataCopy][BrilligOpcode::CalldataCopy] opcode to copy the calldata at a specific offset with and size to the destination address.

pub(super) fn trap_instruction(&mut self, revert_data: HeapVector)

Pushes a [Trap][BrilligOpcode::Trap] opcode.

impl<F, R: RegisterAllocator> BrilligContext<F, R>

pub(crate) fn layout(&self) -> LayoutConfig

Memory layout information. See self::registers for more information about the memory layout.

pub(crate) fn enable_debug_assertions(&self) -> bool

Enable the insertion of bytecode with extra assertions during testing.

pub(crate) fn array_copy_counter_address(&self) -> MemoryAddress

Returns the address of the implicit debug variable containing the count of implicitly copied arrays as a result of RC’s copy on write semantics.

pub(crate) fn count_array_copies(&self) -> bool

If this flag is set, compile the array copy counter as a global.

pub(crate) fn set_globals_memory_size(&mut self, new_size: Option<usize>)

Set the globals memory size if it is not already set. If it is already set, this will assert that the two values must be equal.

pub(crate) fn into_artifact(self) -> BrilligArtifact<F>

Returns the artifact, discarding the rest of the context.

pub(crate) fn artifact(&self) -> &BrilligArtifact<F>

Returns the artifact.

pub(crate) fn name(&self) -> &str

impl<F: AcirField + DebugToString> BrilligContext<F, Stack>

Regular brillig context to codegen user defined functions

pub(crate) fn new( function_name: &str, options: &BrilligOptions, ) -> BrilligContext<F, Stack>

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

pub(crate) fn absolute_value( &mut self, num: SingleAddrVariable, absolute_value: SingleAddrVariable, result_is_negative: SingleAddrVariable, )

Splits a two’s complement signed integer in the sign bit and the absolute value. For example, -6 i8 (11111010) is split to 00000110 (6, absolute value) and 1 (is_negative).

pub(crate) fn convert_signed_division( &mut self, left: SingleAddrVariable, right: SingleAddrVariable, result: SingleAddrVariable, )

impl<F: AcirField + DebugToString> BrilligContext<F, ScratchSpace>

Special brillig context to codegen compiler intrinsic shared procedures

pub(crate) fn new_for_procedure( procedure_id: ProcedureId, options: &BrilligOptions, ) -> BrilligContext<F, ScratchSpace>

Create a BrilligContext with a ScratchSpace for passing procedure arguments.

impl<F: AcirField + DebugToString> BrilligContext<F, GlobalSpace>

Special brillig context to codegen global values initialization

pub(crate) fn new_for_global_init( options: &BrilligOptions, entry_point: Id<Function>, ) -> BrilligContext<F, GlobalSpace>

Create a BrilligContext with a GlobalSpace for memory allocations.

pub(crate) fn global_space_size(&self) -> usize

Total size of the global memory space.

impl<F: AcirField + DebugToString, Registers: RegisterAllocator> BrilligContext<F, Registers>

fn push_opcode(&mut self, opcode: BrilligOpcode<F>)

Adds a brillig instruction to the brillig byte code

pub(crate) fn set_call_stack(&mut self, call_stack: CallStackId)

Sets a current call stack that the next pushed opcodes will be associated with.