noirc_evaluator::ssa

Module interpreter

Modules

errors

intrinsics 🔒

value

Macros

apply_fit_binop 🔒

Applies a fallible integer binary operation on Fitted values, promoting values to Field in case there is an overflow, thus turning the operation infallible.

apply_fit_binop_opt 🔒

Applies a fallible integer binary operation on Fitted values, or returns an overflow error.

apply_fit_comparison_op 🔒

Apply a comparison operator on Fitted values, returning a bool.

apply_int_binop 🔒

Applies an infallible integer binary operation to two NumericValues.

apply_int_binop_opt 🔒

Applies a fallible integer binary operation (e.g., checked arithmetic) to two NumericValues.

apply_int_comparison_op 🔒

Structs

CallContext 🔒

Interpreter 🔒

InterpreterOptions

Constants

MAX_INTERPRETER_CALL_STACK_SIZE 🔒

Maximum number of recursive calls allowed at comptime.

Functions

evaluate_binary 🔒

internal 🔒

interpret_field_binary_op 🔒

interpret_u1_binary_op 🔒

truncate_unsigned 🔒

Type Aliases

IResult 🔒

IResults