noirc_evaluator::ssa
Module interpreter

Modules

errors

intrinsics ğŸ”’

value

Macros

apply_fit_binop ğŸ”’

Applies a fallible integer binary operation on Fitted values, promoting values to Field in case there is an overflow, thus turning the operation infallible.

apply_fit_binop_opt ğŸ”’

Applies a fallible integer binary operation on Fitted values, or returns an overflow error.

apply_fit_comparison_op ğŸ”’

Apply a comparison operator on Fitted values, returning a bool.

apply_int_binop ğŸ”’

Applies an infallible integer binary operation to two NumericValues.

apply_int_binop_opt ğŸ”’

Applies a fallible integer binary operation (e.g., checked arithmetic) to two NumericValues.

apply_int_comparison_op ğŸ”’

Structs

CallContext ğŸ”’

Interpreter ğŸ”’

InterpreterOptions

Functions

evaluate_binary ğŸ”’

internal ğŸ”’

interpret_field_binary_op ğŸ”’

interpret_u1_binary_op ğŸ”’

truncate_unsigned ğŸ”’

Type Aliases

IResult ğŸ”’

IResults