noirc_evaluator::ssa::opt

Module array_set

The purpose of the array_set_optimization SSA pass is to mark ArraySet instructions as mutable iff the array is not potentially shared with the callers or callees of the function and won’t be used again in the function itself either. In other words, if this is the last time we use this version of the array, we can mutate it in place, and avoid having to make a copy of it.

This optimization only applies to ACIR. In Brillig we use ref-counting to decide when there are no other references to an array. This pass assumes that Load and Store instructions have been previously removed.

The pass is expected to run at most once, and requires these passes to occur before itself:

unrolling

flattening

removal of if-else instructions

mem2reg and DIE, in order to remove Load/Store instructions

Structs

Context 🔒

Functions

array_set_optimization_post_check 🔒

Post-check condition for Function::array_set_optimization.

array_set_optimization_pre_check 🔒

Pre-check condition for Function::array_set_optimization.

make_mutable 🔒

Make each ArraySet instruction in instructions_to_update mutable.