noirc_evaluator::ssa::opt
Module constant_folding

The goal of the constant folding optimization pass is to propagate any constants forwards into later Instructions to maximize the impact of compile-time simplifications.

The pass works as follows:

Re-insert each instruction in order to apply the instruction simplification performed by the DataFlowGraph automatically as new instructions are pushed.

Check whether any input values have been constrained to be equal to a value of a simpler form by a constrain instruction. If so, replace the input value with the simpler form.

Check whether the instruction can_be_deduplicated by duplicate instruction earlier in the same block.

These operations are done in parallel so that they can each benefit from each other without the need for multiple passes.

This is the only pass which removes duplicated pure Instructions however and so is needed when different blocks are merged, i.e. after the flatten_cfg pass.

Modules

interpret 🔒

result_cache 🔒

simplification_cache 🔒

Structs

Context 🔒

Enums

CanBeDeduplicated 🔒

Constants

DEFAULT_INTERPRETER_STEP_LIMIT 🔒

Maximum number of SSA instructions to execute during inlining a constant Brillig call.

DEFAULT_MAX_ITER

Functions

can_be_deduplicated 🔒

Indicates if the instruction can be safely replaced with the results of another instruction with the same inputs. If deduplicate_with_predicate is set, we assume we’re deduplicating with the instruction and its predicate, rather than just the instruction. Setting this means instructions that rely on predicates can be deduplicated as well.

constant_folding_post_check 🔒

resolve_cache 🔒