noirc_evaluator::ssa::opt
Module defunctionalize

This module defines the defunctionalization pass for the SSA IR. Certain IR targets (e.g., Brillig and ACIR) do not support higher-order functions directly.

The pass eliminates higher-order functions (a function which accepts function values as arguments or returns functions) by transforming functions used as values (i.e., first-class functions) into constant numbers (fields) that represent their function IDs.

Defunctionalization handles higher-order functions by lowering function values into constant identifiers and replacing calls of function values with calls to a single dispatch apply function.

How the pass works:

Every function used as a value (e.g., passed as a parameter) is assigned a unique NumericType::NativeField value. This value now represents the first-class function’s ID.

All call instructions with non-literal targets are replaced by calls to an apply function.

The apply function is a dispatcher. It takes the function ID as its first argument and calls the appropriate function based on that ID.

Pseudocode of an apply function is given below:

fn apply(function_id: Field, arg1: Field, arg2: Field) -> Field {
    match function_id {
        0 -> function0(arg1, arg2),
        1 -> function1(arg1, arg2),
        ...
        N -> functionN(arg1, arg2),
    }
}

After this pass all first-class functions are replaced with numeric IDs and calls are routed via the newly generated apply functions.

Structs

ApplyFunction 🔒

Represents an ‘apply’ function created by this pass to dispatch higher order functions to. Pseudocode of an apply function is given below:

DefunctionalizationContext 🔒

Performs defunctionalization on all functions This is done by changing all functions as value to be a number (FieldElement) And creating apply functions that dispatch to the correct target by runtime comparisons with constants

Functions

create_apply_function 🔒

Creates a single apply function to enable dispatch across multiple function variants that share the same Signature and RuntimeType.

create_apply_functions 🔒

Creates all apply functions needed for dispatch of function values.

create_dummy_function 🔒

Creates a placeholder (dummy) function to replace calls to invalid function references. An example of a possible invalid function reference is an out-of-bounds access on a zero-length function array.

defunctionalize_post_check 🔒

Check post-execution properties:

defunctionalize_pre_check 🔒

Check pre-execution properties.

find_dynamic_dispatches 🔒

Finds all dynamic dispatch signatures in the given function. Note that these are the signatures before function types are mutated to turn into field types.

find_functions_as_values 🔒

Finds all literal functions used as values in the given function

find_variants 🔒

Collects all functions used as values that can be called by their signatures

function_id_to_field 🔒

Transforms a FunctionId into a FieldElement

make_dummy_return_data 🔒

Construct a dummy value to be returned from the placeholder function for calls to invalid lambda references. We need to construct the appropriate value so that our SSA is well formed even if the function pointer has no variants.

map_function_to_field 🔒

Try to map the given function literal to a field, returning Some(field) on success. Returns none if the given value was not a function or doesn’t need to be mapped.

remove_first_class_functions_in_instruction 🔒

Replace any first class functions used in an instruction with a field value. This applies to any function used anywhere else other than the function position of a call instruction. Returns true if the instruction was modified

replacement_type 🔒

Return what type a function value type should be replaced with:

replacement_types 🔒

Take a list of types that might need replacement. Replaces the ones that need replacement, leaving all others as-is. If no type needs replacement, None is returned.

visit_values_other_than_call_target 🔒

Visit all values which are not targets of a Call.

Type Aliases

ApplyFunctions 🔒

All generated apply functions for each grouping of function variants. Each apply function is handles a specific (Signature, RuntimeType) group. Maps (Signature, RuntimeType) -> ApplyFunction

Variants 🔒

All functions used as a value that share the same signature and runtime type Maps (Signature, Caller RuntimeType) -> Vec<(FunctionId, Callee RuntimeType)>