noirc_evaluator::ssa::opt::defunctionalize
Function create_apply_function

fn create_apply_function(
    ssa: &mut Ssa,
    signature: Signature,
    caller_runtime: RuntimeType,
    function_ids: Vec<(Id<Function>, RuntimeType)>,
) -> Id<Function>

Creates a single apply function to enable dispatch across multiple function variants that share the same Signature and RuntimeType.

This function is responsible for generating an entry point that dispatches between several concrete functions at runtime based on a target field value. It builds a sequence of conditional checks (if-else chain) to compare the target against each function’s ID, and calls the matching function.

These apply functions are to be aggressively inlined as it is assumed that they will be optimized away by the constants at the call site.

Arguments

ssa: A mutable reference to the full Ssa structure containing all functions.

signature: The shared Signature of all variants but with each Type::Function replaced with a field type.

caller_runtime: The runtime in which the apply function will be called, used to update inlining policies.

function_ids: A non-empty list of FunctionIds representing concrete functions to dispatch between. This method will panic if function_ids is empty.

Returns

The FunctionId of the new apply function

Panics

If the function_ids argument has fewer than two elements, implying that no apply function is necessary.