noirc_evaluator::ssa::opt::flatten_cfg
Module branch_analysis

This is an algorithm for identifying branch starts and ends.

The algorithm is split into two parts:

The outer part:

An (unrolled) CFG can be though of as a linear sequence of blocks where some nodes split off, but eventually rejoin to a new node and continue the linear sequence.

Follow this sequence in order, and whenever a split is found call find_join_point_of_branches and then recur from the join point it returns until the return instruction is found.

The inner part defined by find_join_point_of_branches:

The recursive version of the algorithm is as follows:

For each of the two branches in a jmpif block:

Check if either has multiple predecessors. If so, it is a join point.

If not, continue to search the linear sequence of successor blocks from that block.

If another split point is found, recur in find_join_point_of_branches

If a block with multiple predecessors is found, return it.

After, we should have identified a join point for both branches. This is expected to be the same block for both and can be returned from here to continue iteration.

The recursive variant can encounter a stack overflow on certain CFG.

The iterative variant goes like this:

Pop the next branch from a stack of branches we need to process.

Find the next join or branch point on both left and right branches.

If the branches rejoin immediately:

Repeatedly find the next point following the join point:

If it’s another join point, it must be for the parent level:

If this is the first time we see a join point for this level, mark it as pending.

If it’s the second time, mark it as the end point, ensuring it matches the pending value.

If a new branch is encountered, push it onto the stack, noting the parent level to return to.

If either branches are followed by further branching:

Push the branching children onto the stack for visiting later, noting to return to the current level.

Mark any children that goes to in a join point as a pending end for this branch.

This algorithm will remember each join point found in find_join_point_of_branches and the resulting map from each split block to each join block is returned.

Structs

Context 🔒

Enums

Point 🔒

The next interesting point following some block.

Functions

find_branch_ends 🔒

Returns a HashMap mapping blocks that start a branch (i.e. blocks terminated with jmpif) to their corresponding blocks that end the branch.