noirc_evaluator::ssa::opt::flatten_cfg
Struct Context

pub(crate) struct Context<'f> {
    pub(crate) inserter: FunctionInserter<'f>,
    cfg: ControlFlowGraph,
    pub(crate) target_block: BasicBlockId,
    branch_ends: FxHashMap<BasicBlockId, BasicBlockId>,
    condition_stack: Vec<ConditionalContext>,
    next_arguments: Option<Vec<ValueId>>,
    local_allocations: FxHashSet<ValueId>,
    not_instructions: FxHashMap<ValueId, ValueId>,
    pub(crate) no_predicate: bool,
}

Fields
inserter: FunctionInserter<'f>cfg: ControlFlowGraph

This ControlFlowGraph is the graph from before the function was modified by this flattening pass.

target_block: BasicBlockId

Target block of the flattening.

branch_ends: FxHashMap<BasicBlockId, BasicBlockId>

Maps start of branch -> end of branch.

condition_stack: Vec<ConditionalContext>

A stack of each jmpif condition that was taken to reach a particular point in the program. When two branches are merged back into one, this constitutes a join point, and is analogous to the rest of the program after an if statement. When such a join point / end block is found, the top of this conditions stack is popped since we are no longer under that condition. If we are under multiple conditions (a nested if), the topmost condition is the most recent condition combined with all previous conditions via And instructions.

next_arguments: Option<Vec<ValueId>>

Arguments prepared by the last inlined block for the next block we are about to process.

local_allocations: FxHashSet<ValueId>

Stores all allocations local to the current branch.

Since these are local to the current branch (i.e. only defined within one branch of an if expression), they should not be merged with their previous value or stored value in the other branch since there is no such value.

The ValueId here is that which is returned by the allocate instruction.

not_instructions: FxHashMap<ValueId, ValueId>

A map from cond to Not(cond).

Not instructions are inserted constantly by this pass and this map helps keep us from unnecessarily inserting extra instructions, and keeps IDs unique which helps simplifications.

no_predicate: bool

Flag to tell the context to not issue ‘enable_side_effect’ instructions during flattening.

It is set with an attribute when defining a function that cannot fail whatsoever to avoid the overhead of handling side effects.

It can also be set to true when no instruction is known to fail.

Implementations

impl Context<'_>

fn flatten_single_conditional( &mut self, conditional: &BasicConditional, no_predicates: &mut HashMap<Id<Function>, bool>, )

Flattens a single basic conditional by inlining the 2 branches into the entry block.

This method transforms a conditional control flow pattern (if-then-else) into straight-line code by merging the entry, then, else, and exit blocks. The conditional logic is converted into predicated operations using cast and multiplication operations to select between branch values. The method is adapted from flatten_cfg, tailored to do flatten only the input conditional.

Parameters

conditional - The basic conditional structure to flatten

no_predicates - Map of function IDs to their no_predicates attribute

Implementation Details

Sets up context state (target_block, no_predicate) to enable proper inlining

Inlines each block’s instructions into the entry block

Handles terminators to manage control flow during inlining

Uses a WorkList to track which blocks need processing

Copies the exit block’s terminator to the entry block after inlining

Restores original context state after completion

fn map_block_with_mapping( mapping: HashMap<ValueId, ValueId>, func: &mut Function, block: BasicBlockId, )

Applies value mappings to all instructions and terminators in a block.

This method rewrites a block by replacing old value IDs with their mapped equivalents according to the provided mapping. This is used to propagate value simplifications from conditional flattening throughout the rest of the function.

Parameters

mapping - HashMap mapping old ValueIds to their simplified/replaced ValueIds

func - The function containing the block to update

block - The BasicBlockId of the block to remap

impl<'f> Context<'f>

pub(crate) fn new( function: &'f mut Function, cfg: ControlFlowGraph, branch_ends: HashMap<BasicBlockId, BasicBlockId>, target_block: BasicBlockId, ) -> Self

pub(crate) fn flatten(&mut self, no_predicates: &HashMap<Id<Function>, bool>)

Flatten the CFG by inlining all instructions from the queued blocks until all blocks have been flattened.

We follow the terminator of each block to determine which blocks to process next:

If the terminator is a ‘JumpIf’, we assume we are entering a conditional statement and add the start blocks of the ‘then_branch’, ‘else_branch’ and the ‘exit’ block to the queue.

Other blocks will have only one successor, so we will process them iteratively, until we reach one block already in the queue, added when entering a conditional statement, i.e. the ‘else_branch’ or the ‘exit’. In that case we switch to the next block in the queue, instead of the successor.

This process ensures that the blocks are always processed in this order:

if_entry -> then_branch -> else_branch -> exit

In case of nested if statements, for instance in the ‘then_branch’, it will be:

if_entry -> then_branch -> if_entry_2 -> then_branch_2 -> exit_2 -> else_branch -> exit

Information about the nested if statements is stored in the ‘condition_stack’ which is popped/pushed when entering/leaving a conditional statement.

fn link_condition(&mut self, condition: ValueId) -> ValueId

Returns the updated condition so that it is ‘AND-ed’ with the previous condition (if any)

fn get_last_condition(&self) -> Option<ValueId>

Returns the current condition

The conditions are in a stack, they are added as conditional branches are encountered so the last one is the current condition. When processing a conditional branch, we first follow the ‘then’ branch and only after we process the ‘else’ branch. At that point, the ConditionalContext has the ‘else_branch’

fn is_call_to_no_predicate_function( &self, no_predicates: &HashMap<Id<Function>, bool>, instruction: &Id<Instruction>, ) -> bool

Use the provided map to say if the instruction is a call to a no_predicates function

fn prepare_args(&mut self, args: Vec<ValueId>)

Prepare the arguments for the next block to consume.

Panics if we already have something prepared.

fn consume_args(&mut self) -> Vec<ValueId>

Consume the arguments prepared by the previous block.

Panics if there was nothing prepared.

pub(crate) fn inline_block( &mut self, block: BasicBlockId, no_predicates: &HashMap<Id<Function>, bool>, )

Inline all instructions from the given block into the target block, and track slice capacities. This is done by processing every instruction in the block and using the flattening context to push them in the target block.

no_predicates indicates which functions have no predicates and for which we disable the handling of side effects.

pub(crate) fn handle_terminator( &mut self, block: BasicBlockId, work_list: &IndexSet<BasicBlockId>, ) -> Vec<BasicBlockId>

Returns the list of blocks that need to be processed after the given block, and prepare any arguments for the next-to-be-inlined block to consume.

For a normal block, it would be its successor.

For blocks related to a conditional statement, we ensure to process the ‘then_branch’, then the ‘else_branch’ (if it exists), and finally the exit block.

The update of the context is done by the functions if_start, then_stop and else_stop which perform the business logic when entering a conditional statement, finishing the ‘then_branch’ and the ‘else_branch’, respectively.

We know if a block is related to the conditional statement if is referenced by the work_list. Indeed, the start blocks of the ‘then_branch’ and ‘else_branch’ are added to the work_list when starting to process a conditional statement.

fn if_start( &mut self, condition: &ValueId, then_destination: &BasicBlockId, else_destination: &BasicBlockId, if_entry: &BasicBlockId, call_stack: CallStackId, ) -> Vec<BasicBlockId>

Process a conditional statement by creating a ConditionalContext with information about the branch, and storing it in the dedicated stack. Local allocations are moved to the ‘then_branch’ of the ConditionalContext. Returns the blocks corresponding to the ‘then_branch’, ‘else_branch’, and exit block of the conditional statement, so that they will be processed in this order.

fn then_stop(&mut self, block: &BasicBlockId)

Switch context to the ‘else_branch’:

Negates the condition for the ‘else_branch’ and set it in the ConditionalContext

Move the local allocations to the ‘else_branch’

Reset the predicated values to their old mapping in the inserter

Issues the ‘enable_side_effect’ instruction

fn not_instruction( &mut self, condition: ValueId, call_stack: CallStackId, ) -> ValueId

Negates a boolean value by inserting a Not instruction

fn else_stop(&mut self, block: &BasicBlockId)

Switch context the ‘exit’ block of a conditional statement:

Retrieves the local allocations from the Conditional Context

Reset the predicated values to their old mapping in the inserter

Issues the ‘enable_side_effect’ instruction

Joins the arguments from both branches

fn inline_branch_end( &mut self, destination: BasicBlockId, cond_context: ConditionalContext, )

Inline the ending block of a branch, the point where all blocks from a jmpif instruction join back together. In particular this function must handle merging block arguments from all of the join point’s predecessors, and it must handle any differing side effects from each branch.

The merge of arguments is done by inserting an ‘IfElse’ instructions which returns the argument from the ‘then_branch’ or the ‘else_branch’ depending the the condition.

The arguments are prepared for the destination to consume in the next immediate inlining.

fn predicate_value(&mut self, value: ValueId, predicated_value: ValueId)

Map the value to its predicated value in the current conditional context, and store the previous mapping to the ‘predicated_values’ map if not already stored.

fn reset_predicated_values( &mut self, conditional_context: &mut ConditionalContext, )

Restore the previous mapping of predicated values after a branch is finished.

fn insert_instruction( &mut self, instruction: Instruction, call_stack: CallStackId, ) -> ValueId

Insert a new instruction into the target block. Unlike push_instruction, this function will not map any ValueIds. within the given instruction, nor will it modify self.values in any way.

fn insert_instruction_with_typevars( &mut self, instruction: Instruction, ctrl_typevars: Option<Vec<Type>>, call_stack: CallStackId, ) -> InsertInstructionResult<'_>

Inserts a new instruction into the target block, using the given control type variables to specify result types if needed. Unlike push_instruction, this function will not map any ValueIds. within the given instruction, nor will it modify self.values in any way.

fn insert_current_side_effects_enabled(&mut self)

Checks the branch condition on the top of the stack and uses it to build and insert an EnableSideEffectsIf instruction into the target block.

If the stack is empty, a “true” u1 constant is taken to be the active condition. This is necessary for re-enabling side-effects when re-emerging to a branch depth of 0.

fn push_instruction(&mut self, id: Id<Instruction>)

Push the given instruction to the end of the target block of the current function.

Note that each ValueId of the instruction will be mapped via self.inserter.resolve. As a result, the instruction that will be pushed will actually be a new instruction with a different InstructionId from the original. The results of the given instruction will also be mapped to the results of the new instruction.

fn handle_instruction_side_effects( &mut self, instruction: Instruction, call_stack: CallStackId, ) -> Instruction

If we are currently in a branch, we need to modify instructions that have side effects (e.g. constraints, stores, range checks) to ensure that the side effect is only applied if their branch is taken.

For instance we multiply constrain instructions by the branch’s condition (see optimization #1 in the module comment).

fn handle_call_side_effects( &mut self, condition: ValueId, func: ValueId, arguments: Vec<ValueId>, call_stack: CallStackId, ) -> Vec<ValueId>

Weave the condition into call arguments, returning the modified arguments.

fn handle_intrinsic_side_effects( &mut self, condition: ValueId, intrinsic: Intrinsic, arguments: Vec<ValueId>, call_stack: CallStackId, ) -> Vec<ValueId>

Weave the condition into intrinsic call arguments, returning the modified arguments.

fn handle_blackbox_side_effects( &mut self, condition: ValueId, blackbox: BlackBoxFunc, arguments: Vec<ValueId>, call_stack: CallStackId, ) -> Vec<ValueId>

Weave the condition into blackbox call arguments, returning the modified arguments.

fn cast_condition_to_value_type( &mut self, condition: ValueId, value: ValueId, call_stack: CallStackId, ) -> ValueId

‘Cast’ the ‘condition’ to ‘value’ type

This is needed because we need to multiply the condition with several values in order to ‘nullify’ side-effects when the ‘condition’ is false (in ‘handle_instruction_side_effects’ function).

Since the condition is a boolean, it can be safely casted to any other type.

fn mul_by_condition( &mut self, value: ValueId, condition: ValueId, call_stack: CallStackId, ) -> ValueId

Insert a multiplication between ‘condition’ and ‘value’