noirc_evaluator::ssa::opt::inlining::inline_info
Function compute_bottom_up_order

pub(crate) fn compute_bottom_up_order(
    ssa: &Ssa,
    call_graph: &CallGraph,
) -> Vec<(Id<Function>, (usize, usize))>

Compute something like a topological order of the functions, starting with the ones that do not call any other functions, going towards the entry points. When cycles are detected, take the one which are called by the most to break the ties.

This can be used to simplify the most often called functions first.

Returns the functions paired with their own as well as transitive weight, which accumulates the weight of all the functions they call, as well as own.