noirc_evaluator::ssa::opt::loop_invariant
Struct LoopInvariantContext

struct LoopInvariantContext<'f> {
    inserter: FunctionInserter<'f>,
    outer_induction_variables: FxHashMap<ValueId, (IntegerConstant, IntegerConstant)>,
    all_induction_variables: FxHashMap<ValueId, (IntegerConstant, IntegerConstant)>,
    cfg: ControlFlowGraph,
    post_dom_frontiers: PostDominanceFrontiers,
    true_value: ValueId,
    false_value: ValueId,
}

Context with the scope of an entire function.

Fields
inserter: FunctionInserter<'f>outer_induction_variables: FxHashMap<ValueId, (IntegerConstant, IntegerConstant)>

Maps outer loop induction variable -> fixed lower and upper loop bound This will be used by inner loops to determine whether they have safe operations reliant upon an outer loop’s maximum induction variable

all_induction_variables: FxHashMap<ValueId, (IntegerConstant, IntegerConstant)>

All induction variables collected up front.

cfg: ControlFlowGraphpost_dom_frontiers: PostDominanceFrontiers

Maps a block to its post-dominance frontiers

true_value: ValueIdfalse_value: ValueId
Implementations

impl LoopInvariantContext<'_>

pub(super) fn can_evaluate_binary_op( &self, loop_context: &LoopContext, binary: &Binary, ) -> bool

Checks whether a binary operation can be evaluated using the bounds of a given loop induction variables.

If it cannot be evaluated, it means that we either have a dynamic loop bound or that the operation can potentially overflow during a given loop iteration.

pub(super) fn simplify_from_loop_bounds( &mut self, loop_context: &LoopContext, block_context: &BlockContext, instruction_id: Id<Instruction>, ) -> bool

Some instructions can take advantage of that our induction variable has a fixed minimum/maximum, For instance operations can be transformed from a checked operation to an unchecked operation.

Checked operations require more bytecode and thus we aim to minimize their usage wherever possible.

For example, if one side of an add/mul operation is a constant and the other is an induction variable with a known upper bound, we know whether that binary operation will ever overflow. If we determine that an overflow is not possible we can convert the checked operation to unchecked.

The function returns false if the instruction must be added to the block, which involves further simplification.

fn simplify_not_equal_constraint( &mut self, loop_context: &LoopContext, lhs: &ValueId, rhs: &ValueId, err: &Option<ConstrainError>, call_stack: CallStackId, ) -> SimplifyResult

Replace ‘assert(invariant != induction)’ with assert((invariant < min(induction) || (invariant > max(induction))) For this simplification to be valid, we need to ensure that the induction variable takes all the values from min(induction) up to max(induction) This means that the assert must be executed at each loop iteration, and that the loop processes all the iteration space. This is ensured via control dependence and the check for break patterns, before calling this function.

fn match_induction_and_invariant( &mut self, loop_context: &LoopContext, lhs: &ValueId, rhs: &ValueId, ) -> Option<(bool, ValueId, ValueId)>

If the inputs are an induction and loop invariant variables, it returns the maximum and minimum values of the induction variable, based on the loop bounds, and a boolean indicating if the induction variable is on the lhs or rhs (true for lhs)

fn simplify_induction_variable( &mut self, loop_context: &LoopContext, block_context: &BlockContext, instruction_id: Id<Instruction>, ) -> SimplifyResult

Simplify certain instructions using the lower/upper bounds of induction variables

fn match_induction_and_constant( &self, loop_context: &LoopContext, lhs: &ValueId, rhs: &ValueId, only_outer_induction: bool, ) -> Option<(bool, IntegerConstant, IntegerConstant, IntegerConstant)>

If the inputs are an induction variable and a constant, it returns the constant value, the maximum and minimum values of the induction variable, based on the loop bounds, and a boolean indicating if the induction variable is on the lhs or rhs (true for lhs) if only_outer_induction, we only consider outer induction variables, else we also consider the induction variables from the current loop.

fn simplify_induction_variable_in_binary( &mut self, loop_context: &LoopContext, binary: &Binary, is_header: bool, ) -> SimplifyResult

Given a constant c and an induction variable i:

Replace comparisons i < c by true if max(i) < c, and false if min(i) >= c

Replace comparisons c < i by true if min(i) > c, and false if max(i) <= c

Replace equalities i == c by false if min(i) > c or max(i) < c

Replace checked operations with unchecked version if the induction variable bounds prove that the operation will not overflow

header indicates if we are in the loop header where loop bounds do not apply yet

impl<'f> LoopInvariantContext<'f>

fn new(function: &'f mut Function, loops: &[Loop]) -> Self

fn hoist_loop_invariants( &mut self, loop_: &Loop, all_loops: &[Loop], pre_header: BasicBlockId, )

Perform loop invariant code motion and induction variable analysis. See loop_invariant module for more information.

fn is_control_dependent_post_header( &self, loop_: &Loop, all_loops: &[Loop], block: BasicBlockId, all_predecessors: &BTreeSet<BasicBlockId>, nested_loop_control_dependent_blocks: &mut HashSet<BasicBlockId>, ) -> bool

Checks whether a block is control dependent on any blocks after the given loop’s header (excluding itself and the header).

fn find_control_dependent_nested_loop<'a>( &self, loop_: &Loop, block: BasicBlockId, all_loops: &'a [Loop], all_predecessors: &BTreeSet<BasicBlockId>, ) -> Option<&'a Loop>

Determines if the block is in a nested loop that is control dependent on a block in the outer loop. Returns the first such loop found, if any.

If this is the case, we block hoisting as control is not guaranteed. If the block is not control dependent on the inner loop itself, it will be marked appropriately when the inner loop is processed later.

Control dependence on a nested loop is determined by checking whether the nested loop’s header is control dependent on any blocks between itself and the outer loop’s header. It is expected that all_predecessors contains at least all of these blocks.

fn does_block_execute( &self, loop_context: &LoopContext, all_loops: &[Loop], block: BasicBlockId, ) -> bool

Determine whether a block in the loop body is guaranteed to execute.

We know a loop body will execute if we have constant loop bounds where the upper bound is greater than the lower bound.

The loop will never be executed if we have equal loop bounds or we are unsure if the loop will ever be executed (dynamic loop bounds). If certain instructions were to be hoisted out of a loop that never executed it could potentially cause the program to fail when it is not meant to fail.

A block might be in a nested loop that isn’t guaranteed to execute even if the current loop does.

fn init_block_context( &self, loop_context: &mut LoopContext, loop_: &Loop, all_loops: &[Loop], block: BasicBlockId, ) -> BlockContext

Create a BlockContext.

fn can_hoist_invariant( &mut self, loop_context: &LoopContext, block_context: &BlockContext, instruction_id: Id<Instruction>, ) -> (bool, bool)

Decide if an in instruction can be hoisted into the pre-header of the loop.

Returns 2 flags:

Whether the instruction can be hoisted

If it can be hoisted, does it require an IncrementRc instruction.

fn can_be_hoisted_from_loop_bounds( &self, loop_context: &LoopContext, instruction: &Instruction, ) -> bool

Certain instructions can take advantage of that our induction variable has a fixed minimum/maximum.

For example, an array access can usually only be safely deduplicated when we have a constant index that is below the length of the array.

Checking an array get where the index is the loop’s induction variable on its own would determine that the instruction is not safe for hoisting.

However, if we know that the induction variable’s upper bound will always be in bounds of the array we can safely hoist the array access.

fn map_dependent_instructions(&mut self)

Loop invariant hoisting only operates over loop instructions. The FunctionInserter is used for mapping old values to new values after re-inserting loop invariant instructions. However, there may be instructions which are not within loops that are still reliant upon the instruction results altered during the pass. This method re-inserts all instructions so that all instructions have correct new value IDs based upon the FunctionInserter internal map. Leaving out this mapping could lead to instructions with values that do not exist.