noirc_evaluator::ssa::opt
Module remove_unreachable_instructions

This module defines an SSA pass to remove instructions that are unreachable. For example, if an instruction in a block is constrain u1 0 == u1 1, any subsequent instructions in that block will never be executed. This pass then removes those subsequent instructions and replaces the block’s terminator with a special unreachable value.

This pass might also replace existing instructions with constrain checks, for example binary operations that are guaranteed to overflow.

Handling of constrain

Given an SSA like this:

constrain u1 0 == u1 1
v1 = load v0 -> Field
return v1

Because the constrain is guaranteed to fail, every instruction after it is removed and the terminator is replaced with unreachable:

constrain u1 0 == u1 1
unreachable

Similarly, constrain u1 0 != u1 0 will have the same treatment.

Handling of binary operations

If a binary operation is guaranteed to overflow or fail:

v4 = add u8 254, u8 127 // guaranteed to overflow
v5 = add v4, v6
return v5

the operation is left in the SSA but a constrain failure is added after it, and every subsequent instruction is removed:

v4 = add u8 254, u8 127
constrain u1 0 == u1 1, "attempt to add with overflow"
unreachable

Division by zero is an operation that is guaranteed to fail, but it does not overflow.

Because binary operations can depend on a side-effects variable (enable_side_effects), the constrain will check the value of this variable whenever a non-constant side effect is active. Every subsequent instruction is then replaced with a zeroed-value. So for example this SSA:

enable_side_effects v3
v4 = add u8 254, u8 127 // guaranteed to overflow
v5 = add v4, v6
return v5

is replaced with this one:

enable_side_effects v3
v4 = add u8 254, u8 127 // guaranteed to overflow
constrain u1 0 == v3, "attempt to add with overflow"
return u8 0

Handling of array_get and array_set

If an array operation in ACIR is guaranteed to produce an index-out-of-bounds:

enable_side_effects v0
v1 = make_array [] -> [&mut Field; 0]
v2 = array_get v1, index u32 0 -> &mut Field
v3 = load v2
v4 = add v3, Field 2

the array operation is replaced with a constrain failure and its results with default values:

enable_side_effects v0
v1 = make_array [] -> [&mut Field; 0]
constrain v0 == u1 0, "Index out of bounds"
v2 = allocate -> &mut Field
store Field 0 at v2
v3 <- load v2 -> Field
v4 = add v3, Field 2

For the store and the load to be resolved, this pass has to be followed up with a mem2reg pass before any subsequent DIE pass would remove the store, leaving the load with a reference that never gets stored at.

Handling of slice operations

If a slice operation like slice_push_back or slice_pop_front in ACIR is guaranteed to fail, which can only happen if the slice is empty, the operation is removed and replaced with a constrain failure, then the returned slice is replaced with an empty slice. So this SSA:

v0 = make_array [] -> [u32]
v2, v3, v4 = call slice_pop_front(u32 0, v0) -> (u32, u32, [u32])
v5 = add v2, u32 1
return v4, v5

is replaced with this SSA:

v0 = make_array [] -> [u32]
constrain u1 0 == u1 1, "Index out of bounds"
v1 = make_array [] -> [u32]
return v1, u32 0

Preconditions:

the inlining and flatten_cfg must not run after this pass as they can’t handle the unreachable terminator.

Enums

Reachability 🔒

Functions

binary_operation_always_fails 🔒

If a binary operation is guaranteed to fail, returns the error message. Otherwise returns None.

binary_operator_is_unchecked 🔒

insert_constraint 🔒

Insert a constrain 0 == <predicate>, "<msg>" instruction.

remove_and_replace_with_defaults 🔒

Remove the current instruction and replace it with default values.

should_replace_instruction_with_defaults 🔒

Check if an instruction should be replaced with default values if we are in the UnreachableUnderPredicate mode.

zeroed_value 🔒