noirc_evaluator::ssa
Module ssa_gen

Modules

context 🔒

program 🔒

value 🔒

Structs

Ssa

Contains the entire SSA representation of the program.

Constants

SHOW_INVALID_SSA_ENV_KEY 🔒

SSA_WORD_SIZE 🔒

Functions

generate_ssa

Generates SSA for the given monomorphized program.

is_oracle_func 🔒

Return whether the expression refers to a foreign function.

is_pure_builtin_func 🔒

Return whether the expression refers to a pure builtin or low level function.

validate_ssa

validate_ssa_or_err 🔒

Run the panicky validation, and try to turn it into a RuntimeError if it fails.